% Generated by roxygen2 (4.0.0): do not edit by hand
\name{aggregate_shape_data}
\alias{aggregate_shape_data}
\title{Aggregete shape data}
\usage{
aggregate_shape_data(shp.from, shp.to, variables = NULL)
}
\arguments{
\item{shp.from}{the (detailed) shape object to be aggregated.}

\item{shp.to}{the (global) shape object that defines the aggregates}

\item{variables}{variable names of \code{shp.to}. If missing, all numeric variables are taken.}
}
\description{
Aggregates numeric data from a detailed shape object to a global shape object. It uses an intersection matrix, which stores the intersection ratios of each detailed polygon and each global polygon.
}

