% Generated by roxygen2 (4.0.1): do not edit by hand
\name{animation_geo}
\alias{animation_geo}
\title{Create animations}
\usage{
animation_geo(expr, width = 1000, height = 1000, delay = 40,
  filename = "animation.gif")
}
\arguments{
\item{expr}{R expression to create series of geo plots. This can be achieved by small multiples while setting the nrow and ncol parameters of \code{\link{geo_facets}} to 1.}

\item{width}{width of the animation file (in pixels)}

\item{height}{height of the animation file (in pixels)}

\item{delay}{delay time between images}

\item{filename}{filename of the video (should be a .gif or .mpg file)}
}
\description{
This function creates a gif or mpeg animation from a geo plot. The free tool ImageMagick is required.
}
\keyword{animation}

