% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bb.R
\name{bb}
\alias{bb}
\title{Bounding box generator}
\usage{
bb(x = NA, ext = NULL, xlim = NULL, ylim = NULL, relative = FALSE,
  projection = NULL, q = NULL)
}
\arguments{
\item{x}{Either a shape (from class \code{\link[sp:Spatial]{Spatial}} or \code{\link[raster:Raster-class]{Raster}}) or a bounding box (either 2 by 2 matrix or an \code{\link[raster:Extent]{Extent}} object)}

\item{ext}{Extension factor of the bounding box. If 1, the bounding box is unchanged. Values smaller than 1 reduces the bounding box, and values larger than 1 enlarges the bounding box. If specified, it overrules \code{xlim}, \code{ylim}, and \code{relative}.}

\item{xlim}{limits of the x-axis. These are either absolute or relative (depending on the argument \code{relative}).}

\item{ylim}{limits of the y-axis. See \code{xlim}.}

\item{relative}{boolean that determines whether relative values are used for \code{xlim} and \code{ylim} or absolute.}

\item{projection}{projection string (see \code{\link{set_projection}}) to transform the bounding box to. Only applicable if \code{x} is a shape.}

\item{q}{Open Street Map search query. The bounding is automatically generated by searching for \code{q} in Open Street Map Nominatim. See \url{http://wiki.openstreetmap.org/wiki/Nominatim}. If specified, it overrules all other arguments except \code{ext}.}
}
\description{
Retrieves, modifies, or generates a bounding box. Existing bounding boxes can be retrieved (like \code{\link[sp:bbox]{bbox}}). They can be easily modified in two ways: either by using extension factor (\code{ext}), or by changing the x and y limits. New bounding boxes can be created by setting x and y limits in absolute values. Alternatively, bouding boxes can be automatically generated by a search query.
}

