% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated_functions}
\alias{deprecated_functions}
\alias{animation_tmap}
\alias{save_tmap}
\alias{style_catalogue}
\alias{style_catalog}
\alias{last_map}
\alias{tm_style_white}
\alias{tm_style_gray}
\alias{tm_style_natural}
\alias{tm_style_grey}
\alias{tm_style_cobalt}
\alias{tm_style_col_blind}
\alias{tm_style_albatross}
\alias{tm_style_beaver}
\alias{tm_style_bw}
\alias{tm_style_classic}
\alias{tm_format_World}
\alias{tm_format_World_wide}
\alias{tm_format_Europe}
\alias{style_catalog}
\alias{tm_format_Europe2}
\alias{tm_format_Europe_wide}
\alias{tm_format_NLD}
\alias{tm_format_NLD_wide}
\title{Deprecated tmap functions}
\usage{
animation_tmap(...)

save_tmap(...)

style_catalogue(...)

style_catalog(...)

last_map()

tm_style_white(...)

tm_style_gray(...)

tm_style_natural(...)

tm_style_grey(...)

tm_style_cobalt(...)

tm_style_col_blind(...)

tm_style_albatross(...)

tm_style_beaver(...)

tm_style_bw(...)

tm_style_classic(...)

tm_format_World(...)

tm_format_World_wide(...)

tm_format_Europe(...)

tm_format_Europe2(...)

tm_format_Europe_wide(...)

tm_format_NLD(...)

tm_format_NLD_wide(...)
}
\arguments{
\item{...}{passed on to replacing functions}
}
\description{
For consistency, tmap function names will be prefixed with a tm_ or tmap_. Therefore, function names such as \code{animation_tmap} have been renamed to \code{tmap_animation}.
}
\details{
\itemize{
 \item \code{animation_tmap}: replaced by \code{\link{tmap_animation}}
 \item \code{save_tmap}: replaced by \code{\link{tmap_save}}
 \item \code{style_catalogue}: replaced by \code{\link{tmap_style_catalogue}}
 \item \code{style_catalog}: replaced by \code{\link{tmap_style_catalog}}
 \item \code{last_map}: replaced by \code{\link{tmap_last}}
 \item \code{tm_style_white}: replaced by \code{\link{tm_style}("white")}
 \item \code{tm_style_gray}: replaced by \code{\link{tm_style}("gray")}
 \item \code{tm_style_grey}: replaced by \code{\link{tm_style}("grey")}
 \item \code{tm_style_natural}: replaced by \code{\link{tm_style}("natural")}
 \item \code{tm_style_cobalt}: replaced by \code{\link{tm_style}("cobalt")}
 \item \code{tm_style_col_blind}: replaced by \code{\link{tm_style}("col_blind")}
 \item \code{tm_style_albatross}: replaced by \code{\link{tm_style}("albatross")}
 \item \code{tm_style_beaver}: replaced by \code{\link{tm_style}("beaver")}
 \item \code{tm_style_bw}: replaced by \code{\link{tm_style}("bw")}
 \item \code{tm_style_classic}: replaced by \code{\link{tm_style}("classic")}
 \item \code{tm_format_World}: replaced by \code{\link{tm_format}("World")}
 \item \code{tm_format_World_wide}: replaced by \code{\link{tm_format}("World_wide")}
 \item \code{tm_format_NLD}: replaced by \code{\link{tm_format}("NLD")}
 \item \code{tm_format_NLD_wide}: replaced by \code{\link{tm_format}("NLD_wide")}
 \item \code{tm_format_Europe}: not used anymore, since the dataset Europe is no longer maintained
 \item \code{tm_format_Europe2}: not used anymore, since the dataset Europe is no longer maintained
 \item \code{tm_format_Europe_wide}: not used anymore, since the dataset Europe is no longer maintained
}
}
\keyword{internal}
