% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/double_line.R
\name{double_line}
\alias{double_line}
\title{Create a double line}
\usage{
double_line(shp, width)
}
\arguments{
\item{shp}{The shape object that contains the lines (\code{\link[sp:SpatialLinesDataFrame]{SpatialLinesDataFrame}})}

\item{width}{Width between the double lines}
}
\value{
SpatialLinesDataFrame
}
\description{
Create a double line from a single line. (Experimental, see note)
}
\note{
This function is still in experimental phase, which means that it may not be stable and it may be changed significantly in future versions. Moreover, it is unsure if it will stay in tmap; instead, it may be put in a different package, along with functions of similar tasks.
}

