\docType{data}
\name{Europe}
\alias{Europe}
\title{Europe map}
\description{
Europe map. Lambert azimuthal equal-area projection is used
by default for this map. Several countries are
transcontinental and are partly located in Asia. From these
countries, only Russia and Turkey have been included in
this map as part of Europe since they are widely considered
as European countries. Other transcontinental countries
Azerbaijan, Georgia, and Kazakhstan, are also included in
the map, but only passively. From the other surrounding
countries, only Greenland is removed from the map, since it
interferes with the map title.
}

