% Generated by roxygen2 (4.0.0): do not edit by hand
\name{geo-element}
\alias{geo-element}
\title{Geo element}
\description{
Building block to draw thematic maps.
}
\details{
The only fundamental, and hence required element is
\itemize{
\item \code{\link{geo_shape}} that specifies the shape object, and also controls the projection and bounding box}

The plain drawing elements are
\itemize{
\item \code{\link{geo_borders}} to draw polygon borders
\item \code{\link{geo_fill}} to color the polygons
\item \code{\link{geo_bubbles}} to draw bubbles}

The statistical drawing elements are
\itemize{
\item \code{\link{geo_choropleth}} to color the polygons based on a data variable
\item \code{\link{geo_bubblemap}} to draw bubbles based on data variables for size and color}
The layers can be stacked by simply adding them with the + symbol. The combination of the elements described above form one layer. Multiple layers can be stacked. Each layer should start with \code{\link{geo_shape}}.

The layout elements are
\itemize{
\item \code{\link{geo_theme}} to change the appearance of the map, for instance titles and legend positions. Predefined themes for the example shape files are \code{\link{geo_theme_World}}, \code{\link{geo_theme_Europe}}, and \code{\link{geo_theme_NLD}}.
\item \code{\link{geo_grid}} that specifies how small multiples are created, i.e. how many rows and colums, and whether the statistical data variables have free scales or not.}
}

