\docType{package}
\name{geo-package}
\alias{geo-package}
\title{Statistical geographic maps}
\description{
\tabular{ll}{ Package: \tab geo\cr Type: \tab Package\cr
Version: \tab 0.3\cr Date: \tab 2014-04-17\cr License: \tab
GPL-3\cr LazyLoad: \tab yes\cr }
}
\details{
Tools for creating thematic cartograpic maps of the
Netherlands: \itemize{ \item read shape files directly or
from a network repository: \code{\link{get_shape}}; \item
append data to shape files: \code{\link{append_data}};
\item create statistical maps with layers: e.g.
\code{\link{geo_choropleth}}; \item convert from RD
(rijksdriehoekstelsel) to wgs84 coordinates:
\code{\link{rd2wgs84}}; \item (not working yet) export to
kml format (for Google Earth): \code{shp2kml}; \item (not
working yet) interactive cartography: \code{igeoNL}. }
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{GIS}
\keyword{cartography}

