\name{geo_borders}
\alias{geo_borders}
\title{Draw polygon borders}
\usage{
geo_borders(col = "black", lwd = 1, lty = "solid")
}
\arguments{
  \item{col}{line color}

  \item{lwd}{line width (see
  \code{\link[graphics:par]{par}})}

  \item{lty}{line type (see
  \code{\link[graphics:par]{par}})}
}
\value{
\code{\link{geo-object}}
}
\description{
This layer defines the borders of the polygons. Color, line
width and line type can be set.
}

