% Generated by roxygen2 (4.0.0): do not edit by hand
\name{geo_lines}
\alias{geo_lines}
\title{Draw spatial lines}
\usage{
geo_lines(col = "red", lwd = 1, lty = "solid", palette = NULL,
  by = FALSE)
}
\arguments{
\item{col}{color of the lines. Either a color value or a data variable name.}

\item{lwd}{line width}

\item{lty}{line type}

\item{palette}{color palette, used if \code{col} is a data variable}

\item{by}{logical. If \code{TRUE} and \code{col} is a data variable, draw small multiples, one for each level}
}
\value{
\code{\link{geo-element}}
}
\description{
This layer draw spatial lines.
}

