\name{geo_text}
\alias{geo_text}
\title{Add text labels}
\usage{
geo_text(text, cex = 1)
}
\arguments{
  \item{text}{name of the variable in the shape object that
  contains the text labels}

  \item{cex}{relative size of the text labels}
}
\value{
\code{\link{geo-object}}
}
\description{
This layer adds text labels
}

