\docType{package}
\name{geoNL-package}
\alias{geoNL}
\alias{geoNL-package}
\title{Thematic cartography for the Netherlands}
\description{
  \tabular{ll}{ Package: \tab geoNL\cr Type: \tab
  Package\cr Version: \tab 0.2\cr Date: \tab 2013-02-28\cr
  License: \tab GPL-3\cr LazyLoad: \tab yes\cr }
}
\details{
  Tools for creating thematic cartograpic maps of the
  Netherlands: \itemize{ \item read shape files directly or
  from a network repository: \code{\link{getShape}}; \item
  append data to shape files: \code{\link{appendData}};
  \item create choropleths: \code{\link{choropleth}}; \item
  create bubble maps: \code{\link{bubbleMap}}; \item
  convert from RD (rijksdriehoekstelsel) to wgs84
  coordinates: \code{\link{rd2wgs84}}; \item export to kml
  format (for Google Earth): \code{\link{shp2kml}}; \item
  interactive cartography: \code{\link{igeoNL}}. }
}
\author{
  Martijn Tennekes \email{mtennekes@gmail.com} and Edwin de
  Jonge
}
\keyword{cartography}
\keyword{GIS}
\keyword{NL}

