% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_asp_ratio.R
\name{get_asp_ratio}
\alias{get_asp_ratio}
\title{Get aspect ratio}
\usage{
get_asp_ratio(shp)
}
\arguments{
\item{shp}{shape object, either \code{\link[sp:Spatial]{Spatial}} or a \code{\link[raster:Raster-class]{Raster}}.}
}
\value{
aspect ratio
}
\description{
Get the aspect ratio of a shape object, i.e., the map width devided by the map height.
}

