% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_misc.R
\name{get_fact_levels_na}
\alias{get_fact_levels_na}
\title{Internal tmap function that gets factor levels with NA's}
\usage{
get_fact_levels_na(x, o)
}
\arguments{
\item{x}{vector}

\item{o}{options}
}
\description{
Internal tmap function that gets factor levels with NA's
}
\keyword{internal}
