% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_IDs.R
\name{get_IDs}
\alias{get_IDs}
\title{Get ID's of the shape items}
\usage{
get_IDs(shp)
}
\arguments{
\item{shp}{shape object}
}
\value{
vector of ID's
}
\description{
Get ID's of the shape items. For polygons and lines, the ID attribute is used. For points, the coordinates are used.
}

