% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_polygon_directions.R
\name{get_polygon_directions}
\alias{get_polygon_directions}
\alias{set_polygon_directions}
\title{Get and set polygon direction}
\usage{
get_polygon_directions(shp, y0_bottom = TRUE)

set_polygon_directions(shp, y0_bottom = TRUE, clockwise = TRUE)
}
\arguments{
\item{shp}{shape object, a SpatialPolygons(DataFrame)}

\item{y0_bottom}{logical: is the \code{y=0} coordinate at the bottom of the screen?}

\item{clockwise}{logical; should the direction of the polygons be set to clockwise?}
}
\value{
List of logical vectors, each corresponding to a Polygons object. The values of each vector correspond to the Polygon objects. \code{TRUE} means clockwise.
}
\description{
Get and set the direction of polygons (clockwise or counterclockwise). (Experimental, see ntoe)
}
\note{
This function is still in experimental phase, which means that it may not be stable and it may be changed significantly in future versions. Moreover, it is unsure if it will stay in tmap; instead, it may be put in a different package, along with functions of similar tasks.
}

