% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapScale_misc.R
\name{get_scale_defaults}
\alias{get_scale_defaults}
\title{Internal tmap function get scale values}
\usage{
get_scale_defaults(scale, o, aes, layer, cls, ct = NULL)
}
\arguments{
\item{scale}{scale}

\item{o}{o}

\item{aes}{aes}

\item{layer}{layer}

\item{cls}{cls}

\item{ct}{ct}
}
\description{
Internal tmap function get scale values
}
\keyword{internal}
