% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iso_dasymetric.R
\name{iso_dasymetric}
\alias{iso_dasymetric}
\title{Create contour lines (isolines) and the corresponding polygons}
\usage{
iso_dasymetric(shp, var = NULL, nrow = NA, ncol = NA, N = 250000,
  nlevels = 5, style = ifelse(is.null(breaks), "pretty", "fixed"),
  breaks = NULL, bandwidth = NA, cover = NA, weight = 1, ...)
}
\arguments{
\item{shp}{shape object of class \code{\link[sp:Spatial]{Spatial}} or \code{\link[raster:Raster-class]{Raster}}. Spatial points, polygons, and grids are supported. Spatial lines are not.}

\item{var}{variable name. Not needed for \code{\link[sp:SpatialPoints]{SpatialPoints}}. If missing, the first variable name is taken.}

\item{nrow}{number of rows in the raster that is used to smooth the shape object. Only applicable if shp is not a \code{\link[sp:SpatialGridDataFrame]{SpatialGrid(DataFrame)}} or \code{\link[raster:Raster-class]{Raster}}}

\item{ncol}{number of rows in the raster that is used to smooth the shape object. Only applicable if shp is not a \code{\link[sp:SpatialGridDataFrame]{SpatialGrid(DataFrame)}} or \code{\link[raster:Raster-class]{Raster}}}

\item{N}{preferred number of points in the raster that is used to smooth the shape object. Only applicable if shp is not a \code{\link[sp:SpatialGridDataFrame]{SpatialGrid(DataFrame)}} or \code{\link[raster:Raster-class]{Raster}}}

\item{nlevels}{preferred number of levels}

\item{style}{method to cut the color scale: e.g. "fixed", "equal", "pretty", "quantile", or "kmeans". See the details in \code{\link[classInt:classIntervals]{classIntervals}}.}

\item{breaks}{in case \code{style=="fixed"}, breaks should be specified}

\item{bandwidth}{single numeric value or vector of two numeric values that specifiy the bandwidth of the kernal density estimator. By default, it is determined by this formula: (3 * ncol / bounding_box_width, 3 * nrow / bounding_box_height).}

\item{cover}{\code{\link[sp:SpatialPolygons]{SpatialPolygons}} shape that determines the covered area in which the contour lines are placed. By default (\code{NA}), it is based on \code{shp}: if \code{shp} are spatial points, the bounding box is taken, if \code{shp} are spatial polygons, the union of those polygons are taken, and if \code{shp} is a spatial grid, then a 2D kernal density estimator is applied to all non-missing values are taken. The latter uses the function \code{\link{raster_cover}}. Set \code{cover=NULL} to disable the cover, i.e. the contour lines are placing inside the bounding box of \code{shp}.}

\item{weight}{single number that specifies the weight of a single point. Only applicable if \code{shp} is a \code{\link[sp:SpatialPoints]{SpatialPoints}} object.}

\item{...}{argument passed on to other functions (such as \code{\link{raster_cover}})}
}
\value{
list of two items: \code{iso}, a \code{\link[sp:SpatialLinesDataFrame]{SpatialLinesDataFrame}} containing the contour lines, and \code{dasy}, a \code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygonsDataFrame}} containing the polygons between the contour lines.
}
\description{
Create contour lines (isolines) from a shape object. To make the iso lines smooth, a 2D kernal density estimator is applied on the shape object. These lines are used to draw an isopleth. Also, the polygons between the countour lines are returned. They can be used to create a dasymetric map.
}

