% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{land}
\alias{land}
\title{Spatial data of global land cover}
\source{
\url{http://www.iscgm.org/gm/glcnmo.html}
}
\usage{
data(land)
}
\description{
Spatial data of global land cover, of class \code{\link[sp:SpatialGridDataFrame]{SpatialGridDataFrame}}. The data includes a population times series from 1950 to (forecasted) 2030. All metro areas with over 1 million inhabitants in 2010 are included.
}
\details{
\strong{Important:} publication of these maps is only allowed when cited to Tateishi et al. (2014), and when "Geospatial Information Authority of Japan, Chiba University and collaborating organizations." is shown. See \url{http://www.iscgm.org/gm/glcnmo.html#use}.
}
\references{
Ryutaro Tateishi, Nguyen Thanh Hoan, Toshiyuki Kobayashi, Bayan Alsaaideh, Gegen Tana & Dong Xuan Phong Journal of Geography and Geology; Vol.6, No.3; 2014
}

