% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape_misc.R
\name{make_by_vars}
\alias{make_by_vars}
\title{Internal tmap function to create by variables (used for faceting)}
\usage{
make_by_vars(dt, tmf, smeta)
}
\arguments{
\item{dt}{data.table}

\item{tmf}{tmf object}

\item{smeta}{smeta object}
}
\description{
Internal tmap function to create by variables (used for faceting)
}
\keyword{internal}
