% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map2scale.R
\name{map2divscaleID}
\alias{map2divscaleID}
\title{Map breaks to index numbers of a diverging colour scale}
\usage{
map2divscaleID(breaks, n = 101, contrast = 1)
}
\arguments{
\item{breaks}{vector of breaks}

\item{n}{number of classes, i.e. the length of a diverging colour palette. This should preferable be an odd number, since it contains a neutral middle color.}

\item{contrast}{value between 0 and 1 that determines how much of the \code{(1, n)} range is used. Value \code{contrast=1} means that the most extreme break value, i.e. \code{max(abs(breaks))} is maped to either 1 or n (depending on whether it is a minimum or maximum). There is no contrast at all for \code{contrast=0}, i.e. all index numbers will correspond to the middle class (which has index number \code{((n-1)/2)+1}.}
}
\value{
vector of index numbers
}
\description{
Determines index numbers of a potential diverging colour scale given a vector of breaks.
}

