% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro}
\alias{metro}
\title{Spatial data of metropolitan areas}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 436 rows and 13 columns.
}
\source{
\url{https://population.un.org/wup/}
}
\usage{
metro
}
\description{
\code{metro} includes a population time series from 1950 to (forecasted) 2030. All metro areas with over 1 million inhabitants in 2010 are included.
}
\references{
United Nations, Department of Economic and Social Affairs, Population Division (2014). World Urbanization Prospects: The 2014 Revision, CD-ROM Edition.
}
\keyword{datasets}
