% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLD_prov}
\alias{NLD_prov}
\alias{NLD_muni}
\title{Netherlands datasets}
\source{
\url{https://www.cbs.nl/} for \code{NLD_prov} and \code{NLD_muni}.
}
\usage{
NLD_prov

NLD_muni
}
\description{
Datasets of the Netherlands (province and municipality level),
class \code{\link[sf:sf]{sf}}
}
\details{
The default projections for these objects are
Rijksdriehoekstelsel (Netherlands).

\strong{Important:} publication of these maps is only allowed when citing
Statistics Netherlands (CBS) and Kadaster Nederland as source.
}
\references{
Statistics Netherlands (2014), The Hague/Heerlen, Netherlands, \url{https://www.cbs.nl/}.

Kadaster, the Netherlands' Cadastre, Land Registry, and Mapping Agency (2014), Apeldoorn, Netherlands, \url{https://www.kadaster.nl/}.
}
\keyword{datasets}
