% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layer_donuts.R
\name{opt_tm_donuts}
\alias{opt_tm_donuts}
\title{Options for tm_donuts}
\usage{
opt_tm_donuts()
}
\arguments{
\item{type}{cartogram type, one of: "cont" for contiguous cartogram, "ncont" for non-contiguous cartogram and "dorling" for Dorling cartograms}

\item{itermax, }{maximum number of iterations (see \code{\link[cartogram:cartogram_cont]{cartogram::cartogram_cont()}})}

\item{...}{arguments passed on to \code{\link[cartogram:cartogram_cont]{cartogram::cartogram_cont()}}}
}
\description{
Options for tm_donuts
}
