% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_element.R
\name{+.tm_element_list}
\alias{+.tm_element_list}
\title{Stacking of tmap elements}
\usage{
\method{+}{tm_element_list}(e1, e2)
}
\arguments{
\item{e1}{first \code{\link{tmap-element}}}

\item{e2}{second \code{\link{tmap-element}}}
}
\description{
The plus operator allows you to stack \code{\link{tmap-element}s}, and groups of \code{\link{tmap-element}s}.
}
\references{
Tennekes, M., 2018, {tmap}: Thematic Maps in {R}, Journal of Statistical Software, 84(6), 1-39, \href{https://doi.org/10.18637/jss.v084.i06}{DOI}
}
\seealso{
\code{\link{tmap-element}} and \href{../doc/tmap-getstarted.html}{\code{vignette("tmap-getstarted")}}
}
