% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngGrob.R
\name{pngGrob}
\alias{pngGrob}
\title{Render a png image}
\usage{
pngGrob(file)
}
\arguments{
\item{file}{character value containing the ile path or url}
}
\value{
\code{\link[grid:rasterGrob]{rasterGrob}}
}
\description{
Render a png image (from file path or url) to a grob, which can be used for proportional symbol maps
}
\seealso{
\code{\link{tm_symbols}}
}

