% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_element.R
\name{print.tm_element}
\alias{print.tm_element}
\alias{print.tm_shape}
\title{Print tm_element}
\usage{
\method{print}{tm_element}(x, ...)

\method{print}{tm_shape}(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{passed on}
}
\description{
Print tm_element
}
\keyword{internal}
