% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_aux.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{providers}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{leaflet}{\code{\link[leaflet]{providers}}}
}}

