% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Spatial data of rivers}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1616 rows and 5 columns.
}
\source{
\url{https://www.naturalearthdata.com}
}
\usage{
rivers
}
\description{
Spatial data of rivers
}
\keyword{datasets}
