% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sbind}
\alias{sbind}
\title{Combine shape objects}
\usage{
sbind(...)
}
\arguments{
\item{...}{shape objects. The objects should be of one of the following classes:
\itemize{
 \item{"1)"}\code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygons(DataFrame)}}
 \item{"2)"}\code{\link[sp:SpatialPointsDataFrame]{SpatialPoints(DataFrame)}}
 \item{"3)"}\code{\link[sp:SpatialLinesDataFrame]{SpatialLines(DataFrame)}}
}}
}
\value{
shape object
}
\description{
Combine shape objects into one shape object. It works analogous to \code{\link{rbind}}.
}

