% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_projection.R
\name{set_projection}
\alias{get_projection}
\alias{set_projection}
\title{Set and get the map projection}
\usage{
set_projection(shp, projection = NULL, current.projection = NULL,
  overwrite.current.projection = FALSE)

get_projection(shp)
}
\arguments{
\item{shp}{shape object of class \code{\link[sp:Spatial]{Spatial}} or
\code{\link[raster:Raster-class]{Raster}}}

\item{projection}{character that determines the new projection. Either a \code{PROJ.4} character string or a shortcut. See \code{\link{get_proj4}} for a list of shortcut values. This argument is
only used to transform the \code{shp}. Use \code{current.projection} to
specify the current projection of \code{shp}.}

\item{current.projection}{the current projection of \code{shp}. Only use this
if the current projection is missing or wrong.}

\item{overwrite.current.projection}{logical that determines whether the
current projection is overwritten if it already has a projection that is
different.}
}
\value{
\code{set_projection} returns a (transformed) shape object with
  updated projection information. \code{get_projection} returns the
  \code{PROJ.4} character string of \code{shp}.
}
\description{
The function \code{set_projection} sets the projection of a shape file. It is
a convenient wrapper of \code{\link[sp:spTransform]{spTransform}} and
\code{\link[raster:projectRaster]{projectRaster}} with shortcuts for commonly
used projections. The projection can also be set directly in the plot call
with \code{\link{tm_shape}}. This function is also used to set the current
projection information if this is missing. The function \code{get_projection}
is used to get the projection information.
}

