% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{World}
\alias{Europe}
\alias{NLD_muni}
\alias{NLD_prov}
\alias{World}
\title{World, Europe and Netherlands map}
\source{
http://www.naturalearthdata.com for \code{World} and \code{Europe}

Statistics Netherlands (http://www.cbs.nl) for \code{NLD_prov} and \code{NLD_muni}.
}
\usage{
data(World)

data(Europe)

data(NLD_prov)

data(NLD_muni)
}
\description{
Maps of the world, Europe and the Netherlands (province and municipality level).
}
\details{
The default projections for these maps are Eckhart IV (World), Lambert azimuthal (Europe), and Rijksdriehoekstelsel (Netherlands). See below. To change the projection, use \code{\link{set_projection}}.

\code{World} World map. The default projection for this world map is Eckhart IV since area sizes are preserved, which is a very important property for statistical purposes.

\code{Europe} Europe map. Lambert azimuthal equal-area projection is used by default for this map. Several countries are transcontinental and are partly located in Asia. From these countries, only Russia and Turkey have been included in this map as part of Europe since they are widely considered as European countries. Other transcontinental countries Azerbaijan, Georgia, and Kazakhstan, are also included in the map, but only passively. From the other surrounding countries, only Greenland is removed from the map, since it interferes with the map title.

\code{NLD_prov} and \code{NLD_muni}, maps of the Netherlands at province and municipality level of 2013. The used projection is the Rijksdriehoekstelsel projection. \strong{Important:} publication of these maps is only allowed when cited to Statistics Netherlands (CBS) and Kadaster Nederland as source.
}

