% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{rivers}
\alias{cities}
\alias{rivers}
\title{Spatial data of cities and rivers}
\source{
http://www.naturalearthdata.com/
}
\usage{
data(rivers)

data(cities)
}
\description{
Spatial data of main world cities and rivers.
}

