% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smooth_raster.R
\name{smooth_raster}
\alias{smooth_raster}
\title{Create a smooth spatial grid}
\usage{
smooth_raster(shp, var = NULL, nrow = NA, ncol = NA, N = 250000,
  bandwidth = NA, weight = 1, bbx = NULL, output = "RasterLayer")
}
\arguments{
\item{shp}{shape object of class \code{\link[sp:Spatial]{Spatial}} or \code{\link[raster:Raster-class]{Raster}}. Spatial points, polygons, and grids are supported. Spatial lines are not.}

\item{var}{variable name. Not needed for \code{\link[sp:SpatialPoints]{SpatialPoints}}. If missing, the first variable name is taken.}

\item{nrow}{number of rows in the raster that is used to smooth the shape object. Only applicable if shp is not a \code{\link[sp:SpatialGridDataFrame]{SpatialGrid(DataFrame)}} or \code{\link[raster:Raster-class]{Raster}}}

\item{ncol}{number of rows in the raster that is used to smooth the shape object. Only applicable if shp is not a \code{\link[sp:SpatialGridDataFrame]{SpatialGrid(DataFrame)}} or \code{\link[raster:Raster-class]{Raster}}}

\item{N}{preferred number of points in the raster that is used to smooth the shape object. Only applicable if shp is not a \code{\link[sp:SpatialGridDataFrame]{SpatialGrid(DataFrame)}} or \code{\link[raster:Raster-class]{Raster}}}

\item{bandwidth}{single numeric value or vector of two numeric values that specifiy the bandwidth of the kernal density estimator. By default, it is determined by this formula: (3 * ncol / bounding_box_width, 3 * nrow / bounding_box_height).}

\item{weight}{single number that specifies the weight of a single point. Only applicable if \code{shp} is a \code{\link[sp:SpatialPoints]{SpatialPoints}} object.}

\item{bbx}{bounding box}

\item{output}{character value defining the output. Options: \code{"RasterLayer"}, \code{"SpatialGridDataFrame"}, and \code{"list"}. The latter is the same output as \code{\link[KernSmooth:bkde2D]{bkde2D}}. If a character vector of two or three of these values is specified, a list is returned.}
}
\value{
see \code{output} argument.
}
\description{
Create a smooth spatial grid from a shape object, which can be spatial points, spatial polygons or another spatial grid. For this purpose, a 2D kernal density estimator is applied. This function is used by \code{\link{iso_dasymetric}}.
}

