% Generated by roxygen2 (4.0.1): do not edit by hand
\name{split_lines}
\alias{split_lines}
\title{Split lines by polygons}
\usage{
split_lines(shp.lines, shp.poly, variables.lines, variables.poly)
}
\arguments{
\item{shp.lines}{The shape object that contains the lines}

\item{shp.poly}{The shape object that contains the polygons}

\item{variables.lines}{Names of the variables of \code{shp.lines} that are appended to the split lines shape object.}

\item{variables.poly}{Names of the variables of \code{shp.poly} that are appended to the split lines shape object.}
}
\value{
shape object with splitted lines
}
\description{
Split a lines shape object by a polygon shape object. Data of the corresponding polygons is appended to the line segments
}

