% Generated by roxygen2 (4.0.1): do not edit by hand
\name{split_lines_equal}
\alias{split_lines_equal}
\title{Split lines in segments of equal length.}
\usage{
split_lines_equal(shp, dist = 1000, include.last = FALSE)
}
\arguments{
\item{shp}{The shape object that contains the lines}

\item{dist}{Distance per segment}

\item{include.last}{Include last point, even though the distance is less than dist from the previous point?}
}
\value{
SpatialLinesDataFrame
}
\description{
Split lines in segments of equal length.
}

