% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/split_lines_poly.R
\name{split_lines_poly}
\alias{split_lines_poly}
\title{Split lines by polygons}
\usage{
split_lines_poly(shp.lines, shp.poly, variables.lines, variables.poly)
}
\arguments{
\item{shp.lines}{The shape object that contains the lines}

\item{shp.poly}{The shape object that contains the polygons}

\item{variables.lines}{Names of the variables of \code{shp.lines} that are appended to the split lines shape object.}

\item{variables.poly}{Names of the variables of \code{shp.poly} that are appended to the split lines shape object.}
}
\value{
shape object with splitted lines
}
\description{
Split a lines shape object by a polygon shape object. Data of the corresponding polygons is appended to the line segments (Experimental, see note)
}
\note{
This function is still in experimental phase, which means that it may not be stable and it may be changed significantly in future versions. Moreover, it is unsure if it will stay in tmap; instead, it may be put in a different package, along with functions of similar tasks.
}

