% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_catalogue.R
\name{style_catalogue}
\alias{style_catalogue}
\alias{style_catalog}
\title{Create a style catalogue}
\usage{
style_catalogue(path = "./tmap_style_previews", styles = NA,
  include.global.styles = TRUE)

style_catalog(path = "./tmap_style_previews", styles = NA,
  include.global.styles = TRUE)
}
\arguments{
\item{path}{path where the png images are stored}

\item{styles}{vector of \code{tm_style_XXX} function names. By default, it contains all styles that are included in tmap, and, if \code{include.global.styles=TRUE}, also the user defined \code{tm_style_XXX} function names.}

\item{include.global.styles}{See above}
}
\description{
Create a style catalogue for each predefined tmap style. The result is a set of png images, one for each style.
}
