% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ps.R
\name{theme_ps}
\alias{theme_ps}
\title{ggplot2 theme for proportional symbols}
\usage{
theme_ps(
  base_size = 12,
  base_family = "",
  plot.axes = FALSE,
  plot.legend = FALSE
)
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base family}

\item{plot.axes}{should the axes be shown?}

\item{plot.legend}{should the legend(s) be shown?}
}
\description{
ggplot2 theme for proportional symbols. By default, this theme only shows the plotting area, so without titles, axes, and legend
}
