% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_cartogram.R
\name{tm_cartogram}
\alias{tm_cartogram}
\title{Map layer: cartogram}
\usage{
tm_cartogram(
  size = 1,
  size.scale = tm_scale(),
  size.legend = tm_legend_hide(),
  size.free = NA,
  plot.order = tm_plot_order("size", reverse = FALSE),
  trans.args = list(type = "cont", itermax = 15),
  ...
)
}
\arguments{
\item{size, size.scale, size.legend, size.free}{Transformation variable that determines the size of the polygons.}

\item{plot.order}{Specification in which order the spatial features are drawn. See \code{\link[=tm_plot_order]{tm_plot_order()}} for details.}

\item{trans.args}{lists that are passed on to internal transformation function}

\item{...}{passd on to \code{\link[=tm_polygons]{tm_polygons()}}}
}
\description{
Map layer that draws a cartogram
}
