% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layout.R
\name{tm_check_fix}
\alias{tm_check_fix}
\title{Check and fix invalid geometries}
\usage{
tm_check_fix()
}
\value{
A \code{tmap} element.
}
\description{
Checks whether the geometries of the shapes specified via \code{\link[=tm_shape]{tm_shape()}} are
valid. Invalid geometries can cause errors or maps with artifacts.
}
\details{
This tmap element checks the geometries with \code{\link[sf:st_is_valid]{sf::st_is_valid()}} and, if
needed, applies \code{\link[sf:st_make_valid]{sf::st_make_valid()}}. If that is unsuccessful, it toggles the
s2 backend of the \strong{sf} package (on or off, depending on whether it was
already enabled) and tries again. If the geometries are still invalid, the map
is drawn with a warning and the invalid geometries are removed.
}
