% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_component}
\alias{tm_component}
\title{Map component: general}
\usage{
tm_component(
  x,
  height,
  width,
  dpi,
  margins,
  between_margin,
  stack,
  position,
  frame,
  frame.lwd,
  frame.r,
  group.frame,
  resize_as_group,
  z
)
}
\arguments{
\item{x}{object, e.g. a \code{grid::grob}.}

\item{height}{height of the component in number of text line heights.}

\item{width}{width of the component in number of text line heights.}

\item{dpi}{dpi of the component (in view mode)}

\item{margins}{margins}

\item{between_margin}{Margin between}

\item{stack}{stack with other map components, either \code{"vertical"} or \code{"horizontal"}.}

\item{position}{An object created with \code{tm_pos_in()} or \code{tm_pos_out()}. Or, as a shortcut, a vector of two values, specifying the x and y coordinates. The first is \code{"left"}, \code{"center"} or \code{"right"} (or upper case, meaning tighter to the map frame), the second \code{"top"}, \code{"center"} or \code{"bottom"}. Numeric values are also supported, where 0, 0 means left bottom and 1, 1 right top. See also \href{https://r-tmap.github.io/tmap/articles/adv_positions}{vignette about positioning}.}

\item{frame}{frame}

\item{frame.lwd}{frame line width}

\item{frame.r}{Radius of the rounded frame corners. 0 means no rounding.}

\item{group.frame}{group.frame}

\item{resize_as_group}{resize_as_group}

\item{z}{z}
}
\description{
Map component that adds an graphical object
}
\examples{
if (requireNamespace("ggplot2")) {
	library(ggplot2)
	p = ggplot(World, aes(x = gender, y = press, colour = continent)) +
		geom_point() +
		theme_bw()

	g = ggplotGrob(p)

	tm_shape(World) +
		tm_polygons(
			fill = "gender",
			fill.scale = tm_scale(values = "-cols4all.pu_gn_div")) +
		tm_component(g, height = 15, width = 20, position = tm_pos_in("left", "bottom"))
}
}
