% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_scale_.R
\name{tm_const}
\alias{tm_const}
\title{Constant map variable value}
\usage{
tm_const()
}
\description{
Specify a constant value for a map variable (e.g. a fixed fill color or symbol size), bypassing data-driven scaling.
}
