% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layout.R
\name{tm_place_legends_right}
\alias{tm_place_legends_right}
\alias{tm_place_legends_left}
\alias{tm_place_legends_bottom}
\alias{tm_place_legends_top}
\alias{tm_place_legends_inside}
\alias{tm_extra_innner_margin}
\alias{tm_extra_inner_margin}
\title{tmap layout: helper functions}
\usage{
tm_place_legends_right(width = NA)

tm_place_legends_left(width = NA)

tm_place_legends_bottom(height = NA)

tm_place_legends_top(height = NA)

tm_place_legends_inside(pos.h = NULL, pos.v = NULL)

tm_extra_inner_margin(left = 0, right = 0, top = 0, bottom = 0)
}
\arguments{
\item{width}{width}

\item{height}{height}

\item{pos.h, pos.v}{position (horizontal and vertical)}

\item{left, right, top, bottom}{extra margins}
}
\description{
tmap layout: helper functions
}
