% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_group.R
\name{tm_group}
\alias{tm_group}
\title{Layer group control}
\usage{
tm_group(name, control = NA, zoom_levels = NA)
}
\arguments{
\item{name}{group name that corresponds with the group name specified in the layer functions (e.g.  \code{\link[=tm_polygons]{tm_polygons()}})}

\item{control}{The group control determines how
layer groups can be switched on and off. Options: \code{"radio"} for radio
buttons (meaning only one group can be shown), \code{"check"} for check boxes
(so multiple groups can be shown), and \code{"none"} for no control
(the group cannot be (de)selected).}

\item{zoom_levels}{The zoom levels at which the group is displays at. When specified \code{control} will be set to \code{"none"}.}
}
\description{
Controls the layer groups in interactive maps (view mode): the layer control box (radio buttons or check boxes) and at which zoom levels the layers are displayed at.
}
\seealso{
\href{https://r-tmap.github.io/tmap/articles/adv_groups}{vignette about layer groups}
}
