% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_inset}
\alias{tm_inset}
\title{Map component: inset map}
\usage{
tm_inset(
  bbox,
  height,
  width,
  margins,
  between_margin,
  stack,
  position,
  frame,
  frame.lwd,
  frame.r,
  group.frame,
  resize_as_group,
  z
)
}
\arguments{
\item{bbox}{a bounding box}

\item{height}{height of the component in number of text line heights.}

\item{width}{width of the component in number of text line heights.}

\item{margins}{margins}

\item{between_margin}{Margin between}

\item{stack}{stack with other map components, either \code{"vertical"} or \code{"horizontal"}.}

\item{position}{An object created with \code{tm_pos_in()} or \code{tm_pos_out()}. Or, as a shortcut, a vector of two values, specifying the x and y coordinates. The first is \code{"left"}, \code{"center"} or \code{"right"} (or upper case, meaning tighter to the map frame), the second \code{"top"}, \code{"center"} or \code{"bottom"}. Numeric values are also supported, where 0, 0 means left bottom and 1, 1 right top. See also \href{https://r-tmap.github.io/tmap/articles/adv_positions}{vignette about positioning}.}

\item{frame}{frame}

\item{frame.lwd}{frame line width}

\item{frame.r}{Radius of the rounded frame corners. 0 means no rounding.}

\item{group.frame}{group.frame}

\item{resize_as_group}{resize_as_group}

\item{z}{z}
}
\description{
Map component that adds an inset map
}
\examples{
bb = tmaptools::bb(NLD_prov[NLD_prov$name == "Utrecht",], ext = 1.05)

tm_shape(NLD_dist) +
	tm_polygons(
		fill = "dwelling_value",
		fill.scale = tm_scale_continuous_pseudo_log(values = "-cols4all.pu_gn_div"),
		col = NULL) +
tm_shape(NLD_muni) +
	tm_borders(col = "black", lwd = 0.5) +
tm_shape(NLD_prov) +
	tm_borders(col = "black", lwd = 1.5) +
tm_inset(bbox = bb, height = 15, width = 15)
}
