% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_mouse_coordinates}
\alias{tm_mouse_coordinates}
\title{Map component: mouse coordinates}
\usage{
tm_mouse_coordinates(stack, position, z)
}
\arguments{
\item{stack}{stack}

\item{position}{position}

\item{z}{z}
}
\description{
Map component that adds mouse coordinates
}
