% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapScale_.R
\name{tm_mv_dim}
\alias{tm_mv_dim}
\title{define multivariate variable based on stars dimension}
\usage{
tm_mv_dim(x, values)
}
\arguments{
\item{x}{dimension name}

\item{values}{values to be used}
}
\description{
define multivariate variable based on stars dimension
}
