% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_osm.R
\name{tm_osm_raster}
\alias{tm_osm_raster}
\title{Utility function to plot a raster OSM image
Plots a raster image obtained through \code{read_osm}.}
\usage{
tm_osm_raster(var = "PIXEL__COLOR")
}
\arguments{
\item{var}{name of the variable used to plot the raster image. Shouldn't need to be altered.}
}
\description{
Utility function to plot a raster OSM image
Plots a raster image obtained through \code{read_osm}.
}

