% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layout.R
\name{tm_plot}
\alias{tm_plot}
\title{Plot mode options}
\usage{
tm_plot(use.gradient)
}
\arguments{
\item{use.gradient}{Use gradient fill using \code{\link[grid:linearGradient]{linearGradient}}}
}
\description{
Plot mode options. This option is specific to the plot mode.
}
