% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_pos.R
\name{tm_pos}
\alias{tm_pos}
\alias{tm_pos_out}
\alias{tm_pos_in}
\alias{tm_pos_auto_out}
\alias{tm_pos_auto_in}
\title{Set the position of map components}
\usage{
tm_pos(cell.h, cell.v, pos.h, pos.v, just.h, just.v)

tm_pos_out(cell.h, cell.v, pos.h, pos.v, just.h, just.v)

tm_pos_in(pos.h, pos.v, just.h, just.v)

tm_pos_out(cell.h, cell.v, pos.h, pos.v, just.h, just.v)

tm_pos_auto_out(cell.h, cell.v, pos.h, pos.v, just.h, just.v)

tm_pos_auto_in(just.h, just.v)
}
\arguments{
\item{cell.h, cell.v}{The plotting area is overlaid with a 3x3 grid, of which the middle grid cell is the map area. Components can be drawn into any cell. `cell.h` specifies the horizontal position (column) can take values `"left"`, `"center"`, and `"right"`, and `cell.v` specifies the vertical position (row) and can take values `"top"`, `"center"`, and `"bottom"`.}

\item{pos.h, pos.v}{The position of the component within the cell. The options for `pos.h` are `"left"`, `"center"`, and `"right"`, and for `cell.v` these are `"top"`, `"center"`, and `"bottom"`.}

\item{just.h, just.v}{The justification of the component in case multiple components are stacked. When they are stacked horizontally, `just.v` determines how compnents that are smaller in height than the available height (determined by the outer.margins if specified and otherwise by the heigest component) are justified: `"top"`, `"center"`, or `"bottom"`. Similarly, `just.h` determines how components are justified horizontally when they are stacked vertically: `"left"`, `"center"`, or `"right"`.}
}
\description{
Set the position of map components, such as legends, title, compass, scale bar, etc. `tm_pos` is the
}
