% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_title}
\alias{tm_title}
\alias{tm_title_in}
\alias{tm_title_out}
\title{Map component: title}
\usage{
tm_title(
  text,
  size,
  color,
  padding,
  fontface,
  fontfamily,
  alpha,
  stack,
  just,
  frame,
  frame.lwd,
  frame.r,
  bg.color,
  bg.alpha,
  position,
  width,
  height,
  group.frame,
  resize_as_group,
  z
)

tm_title_in(text, ..., position = tm_pos_in("left", "top"))

tm_title_out(text, ..., position = tm_pos_out("center", "top"))
}
\arguments{
\item{text}{text}

\item{size}{font size}

\item{color}{font color}

\item{padding}{padding}

\item{fontface}{font face, bold, italic}

\item{fontfamily}{font family}

\item{alpha}{alpha transparency of the text}

\item{stack}{stack}

\item{just}{just}

\item{frame}{frame}

\item{frame.lwd}{frame line width}

\item{frame.r}{Radius of the rounded frame corners. 0 means no rounding.}

\item{bg.color}{Background color}

\item{bg.alpha}{Background transparency}

\item{position}{Vector of two values, specifying the x and y coordinates. The first is either \code{"left"} or \code{"right"}, the second either \code{"top"} or \code{"bottom"}.}

\item{width, height}{width and height of the text box.}

\item{group.frame}{group.frame}

\item{resize_as_group}{resize_as_group}

\item{z}{z}

\item{...}{passed on to \code{tm_title()}}
}
\description{
Map component that adds a title
}
