% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_vars.R
\name{tm_vars}
\alias{tm_vars}
\title{Specify variables}
\usage{
tm_vars(
  x = NA,
  dimvalues = NULL,
  n = NA,
  multivariate = FALSE,
  animate = FALSE
)
}
\arguments{
\item{x}{variable names, variable indices, or a dimension name}

\item{dimvalues}{dimension values}

\item{n}{if specified the first \code{n} variables are taken (or the first \code{n} dimension values)}

\item{multivariate}{in case multiple variables are specified, should they serve as facets (FALSE) or as a multivariate map variable?}

\item{animate}{should the variable(s) be animated? (experimental)}
}
\description{
tmap function to specify all variables in the shape object
}
\seealso{
\href{https://r-tmap.github.io/tmap/articles/basics_scales}{Scales}, \href{https://r-tmap.github.io/tmap/articles/basics_vv}{Map variables}, \href{https://r-tmap.github.io/tmap/articles/adv_multivariate}{Multivariate map variables}
}
