% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_xylab.R
\name{tm_xlab}
\alias{tm_xlab}
\alias{tm_ylab}
\title{Map: x and y labels}
\usage{
tm_xlab(text, size, color, rotation, space, fontface, fontfamily, side)

tm_ylab(text, size, color, rotation, space, fontface, fontfamily, side)
}
\arguments{
\item{text}{text of the title}

\item{size}{font size of the title}

\item{color}{color}

\item{rotation}{rotation in degrees}

\item{space}{space between label and map in number of line heights}

\item{fontface}{font face}

\item{fontfamily}{font family}

\item{side}{side: `"top"` or `"bottom"` for `tm_xlab` and `"left"` or `"right"` for `tm_ylab`}
}
\description{
The x and y labels for maps
}
