% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_format.R
\name{tmap_format}
\alias{tmap_format}
\alias{tmap_format_add}
\alias{tm_format}
\title{Deprecated: format}
\usage{
tmap_format(format)

tmap_format_add(..., name)

tm_format(format, ...)
}
\arguments{
\item{format}{Name of the format}

\item{...}{not used}

\item{name}{Name of the format}
}
\description{
In tmap < 4.0 it was possible to set shape-specific options, such as margins and legend position. However, this has become superfluous because in tmap > 4.0 legends are by default placed outside the map area. If needed, a shape-specific set of options can be stored as a style with tmap_options_save.
}
\keyword{internal}
