% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmap2svg.R
\name{tmap2svg}
\alias{tmap2svg}
\title{Interactive SVG}
\usage{
tmap2svg(tm, file = NULL)
}
\arguments{
\item{tm}{tmap object. A tmap object is created with \code{\link{qtm}} or by stacking \code{\link{tmap-element}}s.}

\item{file}{file name. If specified, the SVG image is saved to this file.}
}
\description{
Convert the tmap output to an interactive SVG, that can be plot in RStudio. In development. Also svgPanZoom is not yet released on CRAN. See \url{devtools::install_github("timelyportfolio/svgPanZoom")}.
}
\examples{
data(World)
data(metro)
metro$growth <- (metro$pop2020 - metro$pop2010) / (metro$pop2010 * 10) * 100

tm <- tm_shape(World) +
	tm_polygons("income_grp", palette="-Blues", contrast=.7, id="name") +
	tm_shape(metro) +
	tm_bubbles("pop2010", col = "growth", 
			   border.col = "black", border.alpha = .5, 
			   style="fixed", breaks=c(-Inf, seq(0, 6, by=2), Inf),
			   palette="-RdYlBu", contrast=1, 
			   title.size="Metro population", 
			   title.col="Growth rate (\%)", id="name") + 
	tm_layout_World()

tmap2svg(tm)
}

