% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_mode.R
\name{tmap_devel_mode}
\alias{tmap_devel_mode}
\title{Set the development mode}
\usage{
tmap_devel_mode(devel.mode)
}
\arguments{
\item{devel.mode}{logical value that determines the development mode.
If omitted then the development mode is toggled.}
}
\description{
When the so-called "development mode" is enabled, helpful messages and timings
are printed in the console
}
