% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_grob.R
\name{tmap_grob}
\alias{tmap_grob}
\title{Export to grob object}
\usage{
tmap_grob(tm)
}
\arguments{
\item{tm}{tmap object}
}
\value{
A grob object when one page is generated, or a list of grob objects when multiple pages are generated.
}
\description{
Export a tmap plot object to a grob object (from the grid package).
}
