% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_element.R, R/tmapGpar.R, R/tmapScale_.R,
%   R/tmapTrans.R
\name{tm_element}
\alias{tm_element}
\alias{tm_element_list}
\alias{tmapGpar}
\alias{tmapTpar}
\alias{tmapUsrCls}
\alias{tmapScale}
\alias{tmapTransCentroid}
\alias{tmapTransRaster}
\alias{tmapTransPolygons}
\alias{tmapTransLines}
\alias{tmapTransCartogram}
\title{Internal methods for tmap extensions}
\usage{
tm_element(..., subclass = NULL)

tm_element_list(...)

tmapGpar(
  fill = NULL,
  col = NULL,
  shape = NULL,
  size = NULL,
  fill_alpha = NULL,
  col_alpha = NULL,
  lty = NULL,
  lwd = NULL,
  linejoin = NULL,
  lineend = NULL,
  ...
)

tmapTpar(...)

tmapUsrCls(x)

tmapScale(aes, value, scale, legend, chart, free)

tmapTransCentroid(
  shpTM,
  xmod = NULL,
  ymod = NULL,
  ord__,
  plot.order,
  args,
  scale
)

tmapTransRaster(shpTM, ord__, plot.order, args)

tmapTransPolygons(shpTM, ord__, plot.order, args, scale)

tmapTransLines(shpTM, ord__, plot.order, args, scale)

tmapTransCartogram(shpTM, size, ord__, plot.order, args, scale)
}
\arguments{
\item{...}{args}

\item{subclass}{subclass}

\item{fill, col, shape, size, fill_alpha, col_alpha, lty, lwd, linejoin, lineend}{visual variables}

\item{x}{x}

\item{aes}{aes}

\item{value}{value}

\item{scale}{scale}

\item{legend}{legend}

\item{chart}{chart}

\item{free}{free}

\item{shpTM}{shpTM}

\item{xmod, ymod}{xmod and ymod}

\item{ord__}{ord}

\item{plot.order}{plot.order}

\item{args}{args}
}
\description{
Internal methods for tmap extensions
}
\keyword{internal}
