% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_element.R
\name{tmap_last}
\alias{tmap_last}
\title{Retrieve the last map to be modified or created}
\usage{
tmap_last()
}
\value{
call
}
\description{
Retrieve the last map to be modified or created. Works in the same way
as \code{ggplot2::last_plot()}, although there is a difference:
\code{tmap_last()} returns the last call instead of the stacked \code{\link{tmap-element}}s.
}
\seealso{
\code{\link[=tmap_save]{tmap_save()}}
}
