% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_overview.R
\name{tmap_overview}
\alias{tmap_overview}
\title{Overview of tmap layers}
\usage{
tmap_overview()
}
\value{
A list of three layer types are returned: data layers, aux layers, and components.
}
\description{
Overview of tmap layers, organized by layer type.
}
