% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_pals.R
\docType{data}
\name{tmap_pals}
\alias{tmap_pals}
\title{List of supported color palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
tmap_pals
}
\description{
List of supported color palettes
}
\keyword{datasets}
