% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_mode.R
\name{tmap_style}
\alias{tmap_style}
\title{Set the default tmap style}
\usage{
tmap_style(style)
}
\arguments{
\item{style}{name of the style. The function \code{tm_style_<style>} should exist and be a wrapper of \code{\link{tm_layout}}. The default style when loading the package is "white", which corresponds to the function \code{\link{tm_style_white}}.}
}
\value{
the style before changing
}
\description{
Set the default tmap style, which is contained in the global option \code{tmap.style}.
}
\seealso{
\code{\link{tm_layout}} for predefined styles, and \code{\link{style_catalogue}} for a style catelogue.
}

