% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_style_catalogue.R
\name{tmap_style_catalogue}
\alias{tmap_style_catalogue}
\alias{tmap_style_catalog}
\title{Create a style catalogue}
\usage{
tmap_style_catalogue(path = "./tmap_style_previews", styles = NA)

tmap_style_catalog(path = "./tmap_style_previews", styles = NA)
}
\arguments{
\item{path}{path where the png images are stored}

\item{styles}{vector of styles function names (see \code{\link{tmap_style}}) for which a preview is generated. By default, a preview is generated for all loaded styles.}
}
\description{
Create a style catalogue for each predefined tmap style. The result is a set of png images, one for each style.
}
