% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_options.R
\name{tmapAddLayerOptions}
\alias{tmapAddLayerOptions}
\title{Internal tmap function to add a default value for the layer functions}
\usage{
tmapAddLayerOptions(option, id, value)
}
\arguments{
\item{option, }{one of: `"value.const"`, `"value.na"`, `"value.blank"`, `"values.var"`, `'values.range'`, `"value.neutral"`, `"scales.var"`}

\item{id}{name of the visual variable with layer, in the format `"x.y"`, where x is the visual variable and y is the layer. It is also possible to set x only; then it applies to all layer functions.}

\item{value}{value}
}
\description{
Internal tmap function to add a default value for the layer functions
}
\keyword{internal}
