% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGetShapeMeta.R
\name{tmapGetShapeMeta1}
\alias{tmapGetShapeMeta1}
\title{Internal method that extracts meta data from shape objects}
\usage{
tmapGetShapeMeta1(...)
}
\arguments{
\item{...}{passed on. First should be the shape, second the list of options}
}
\description{
Internal method that extracts meta data from shape objects
}
\keyword{internal}
