% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGetShapeMeta.R
\name{tmapGetShapeMeta2}
\alias{tmapGetShapeMeta2}
\title{Internal method that extracts more meta data from shape objects}
\usage{
tmapGetShapeMeta2(...)
}
\arguments{
\item{...}{passed on. First should be the shape, second shape meta (from tmapGetShapeMeta1), the first the list of options}
}
\description{
Internal method that extracts meta data from shape objects
}
\keyword{internal}
