% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGridAux.R, R/tmapGridComp_leg_landscape.R,
%   R/tmapLeafletAux.R, R/tmapLeafletComp_leg_portrait.R
\name{tmapGridAuxPrepare}
\alias{tmapGridAuxPrepare}
\alias{tmapGridAuxPlot}
\alias{tmapGridAuxPrepare.tm_aux_basemap}
\alias{tmapGridAuxPrepare.tm_aux_tiles}
\alias{tmapGridAuxPrepare.tm_aux_graticules}
\alias{tmapGridAuxPrepare.tm_aux_grid}
\alias{tmapGridAuxPlot.tm_aux_basemap}
\alias{tmapGridAuxPlot.tm_aux_tiles}
\alias{tmapGridAuxPlot.tm_aux_graticules}
\alias{tmapGridAuxPlot.tm_aux_grid}
\alias{tmapGridCompPrepare}
\alias{tmapGridCompPrepare.default}
\alias{tmapLeafletAuxPrepare}
\alias{tmapLeafletAuxPlot}
\alias{tmapLeafletAuxPrepare.tm_aux_basemap}
\alias{tmapLeafletAuxPrepare.tm_aux_tiles}
\alias{tmapLeafletAuxPlot.tm_aux_basemap}
\alias{tmapLeafletAuxPlot.tm_aux_tiles}
\alias{tmapLeafletAuxPrepare.tm_aux_graticules}
\alias{tmapLeafletAuxPrepare.tm_aux_grid}
\alias{tmapLeafletAuxPlot.tm_aux_graticules}
\alias{tmapLeafletAuxPlot.tm_aux_grid}
\alias{tmapLeafletCompPrepare}
\alias{tmapLeafletCompPrepare.default}
\title{Internal methods for map components in plot and view mode}
\usage{
tmapGridAuxPrepare(a, bs, id, o)

tmapGridAuxPlot(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPrepare}{tm_aux_basemap}(a, bs, id, o)

\method{tmapGridAuxPrepare}{tm_aux_tiles}(a, bs, id, o)

\method{tmapGridAuxPrepare}{tm_aux_graticules}(a, bs, id, o)

\method{tmapGridAuxPrepare}{tm_aux_grid}(a, bs, id, o)

\method{tmapGridAuxPlot}{tm_aux_basemap}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPlot}{tm_aux_tiles}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPlot}{tm_aux_graticules}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPlot}{tm_aux_grid}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapGridCompPrepare(comp, o)

\method{tmapGridCompPrepare}{default}(comp, o)

tmapLeafletAuxPrepare(a, bs, id, o)

tmapLeafletAuxPlot(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapLeafletAuxPrepare}{tm_aux_basemap}(a, bs, id, o)

\method{tmapLeafletAuxPrepare}{tm_aux_tiles}(a, bs, id, o)

\method{tmapLeafletAuxPlot}{tm_aux_basemap}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapLeafletAuxPlot}{tm_aux_tiles}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapLeafletAuxPrepare}{tm_aux_graticules}(a, bs, id, o)

\method{tmapLeafletAuxPrepare}{tm_aux_grid}(a, bs, id, o)

\method{tmapLeafletAuxPlot}{tm_aux_graticules}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapLeafletAuxPlot}{tm_aux_grid}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapLeafletCompPrepare(comp, o)

\method{tmapLeafletCompPrepare}{default}(comp, o)
}
\arguments{
\item{o}{options}

\item{comp}{component}
}
\description{
Internal methods for map components
}
\keyword{internal}
