% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_options.R
\name{tmapMode}
\alias{tmapMode}
\title{Internal method for submitting a new mode (deprecated)}
\usage{
tmapMode(id, name, ..., styleOptions = NULL)
}
\arguments{
\item{id}{id of the mode: please use lowercase and one-word. This will be used with \code{\link[=tmap_mode]{tmap_mode()}}.}

\item{name}{name of the mode: please use title case. This will be used to recognize internal functions, e.g. \code{tmapLeafletInit}}

\item{...}{mode specific options}
}
\description{
Internal method for submitting a new mode. Deprecated. Use \code{\link[=tmapSubmitOptions]{tmapSubmitOptions()}} instead.
}
\keyword{internal}
