% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_functions.R
\name{tmapOutput}
\alias{tmapOutput}
\alias{renderTmap}
\title{Wrapper functions for using tmap in shiny}
\usage{
tmapOutput(outputId, width = "100\%", height = 400)

renderTmap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width, height}{the width and height of the map}

\item{expr}{A tmap object. A tmap object is created with \code{\link{qtm}} or by stacking \code{\link{tmap-element}}s.}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable}
}
\description{
Use \code{tmapOutput} to create a UI element, and \code{renderTmap} to render the tmap widget
}
\details{
Two features from tmap are not (yet) supported in Shiny: small multiples (facets) and colored backgrounds (argument bg.color of \code{\link{tm_layout}}). Workarounds for small multiples: create multiple independent maps or specify as.layers = TRUE in \code{\link{tm_facets}}.
}
