% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape.R
\name{tmapReproject}
\alias{tmapReproject}
\title{Internal method that reprojects a shape object}
\usage{
tmapReproject(shp, tmapID, bbox, ..., crs)
}
\arguments{
\item{shp}{shape}

\item{tmapID}{tmap id values}

\item{bbox}{bounding box}

\item{...}{other args}

\item{crs}{crs}
}
\description{
Internal method that reprojects a shape object
}
\keyword{internal}
