% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape.R
\name{tmapShape}
\alias{tmapShape}
\title{Internal method that processed shape objects}
\usage{
tmapShape(...)
}
\arguments{
\item{...}{passed on. Arguments should be: `shp`, `is.main`, `crs`, `bbox`, `unit`, `filter`, `shp_name`, `o`}
}
\description{
Internal method that processed shape objects
}
\keyword{internal}
