% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapSplitShp.R
\name{tmapSplitShp}
\alias{tmapSplitShp}
\title{Internal method that split shape objects}
\usage{
tmapSplitShp(shp, split_stars_dim, smeta)
}
\arguments{
\item{shp}{shape}

\item{split_stars_dim}{name of the dimension to split (\code{""} to skip)}
}
\description{
Internal method that split shape objects. So far, only used to split stars object
(from dimension to attributes)
}
\keyword{internal}
