% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_options.R
\name{tmapSubmitOptions}
\alias{tmapSubmitOptions}
\title{Internal method for submitting options}
\usage{
tmapSubmitOptions(options = NULL, styleOptions = NULL)
}
\arguments{
\item{options}{list of options}

\item{styleOptions}{list of lists of options, each with a style name}
}
\description{
Internal method for submitting options
}
\keyword{internal}
