% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapSubsetShp.R
\name{tmapSubsetShp}
\alias{tmapSubsetShp}
\title{Internal method that subsets data from shape objects}
\usage{
tmapSubsetShp(...)
}
\arguments{
\item{...}{passed on. First should be the shape, second a vector of variable names}
}
\description{
Internal method that subsets data from shape objects
}
\keyword{internal}
