% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v4-not-implemented.R
\name{v4-not-yet}
\alias{v4-not-yet}
\alias{tmap_tip}
\title{Features not yet implemented in tmap4}
\usage{
tmap_tip(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Features not yet implemented in tmap4
}
