% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{World_rivers}
\alias{World_rivers}
\title{Spatial data of rivers}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1632 rows and 5 columns.
}
\source{
\url{https://www.naturalearthdata.com}
}
\usage{
World_rivers
}
\description{
Spatial data of rivers
}
\note{
In tmap <= 3, this dataset was called \code{rivers}.
}
\keyword{datasets}
