% Generated by roxygen2 (4.0.1): do not edit by hand
\name{append_data}
\alias{append_data}
\title{Append data}
\usage{
append_data(shp, data, key.data = NULL, key.shp = NULL,
  ignore.duplicates = FALSE, fixed.order = is.null(key.data) &&
  is.null(key.shp))
}
\arguments{
\item{shp}{shape object}

\item{data}{data.frame}

\item{key.data}{variable name of \code{data} to be matched with \code{key.shp}. If not specified, and \code{fixed.order} is \code{FALSE}, the row names of \code{data} are taken.}

\item{key.shp}{variable name of \code{shp} map data to be matched with \code{key.data}. If not specified, and \code{fixed.order} is \code{FALSE}, the polygon ID's are taken.}

\item{ignore.duplicates}{should duplicated keys in \code{data} be ignored? (\code{FALSE} by default)}

\item{fixed.order}{should the data be append in the same order as the shapes in \code{shp}?}
}
\value{
shape object with appended data
}
\description{
Append data.frame to shape object
}
\examples{
data(NLD_muni)
data(NLD_prov)
data(NLD_ageGroups)

NLD_muni <- append_data(NLD_muni, NLD_ageGroups, key.data="muni", key.shp="name")

geo_shape(NLD_muni) +
	geo_fill(c("pop_0_20", "pop_20_65", "pop_65p"), convert2density=TRUE, total.area.km2=41543) +
	geo_shape(NLD_prov) +
	geo_borders() +
	geo_facets(free.scales=TRUE) +
	geo_theme_NLD(c("Population 0 to 20", "Population 20 to 65", "Population 65 and older"),inner.margins=c(.05, .2, .7, .05), legend.width=.8, legend.max.height=.5, legend.position=c("left", "top"))
}

