% Generated by roxygen2 (4.0.1): do not edit by hand
\name{approx_areas}
\alias{approx_areas}
\title{Approximate area sizes of the shapes.}
\usage{
approx_areas(shp, total.area.km2 = NA, units = NULL)
}
\arguments{
\item{shp}{shape object, i.e. a SpatialPolygons(DataFrame)}

\item{total.area.km2}{total area size of \code{shp} in number of squared kilometers. If \code{NA}, and \code{unit="km2"}, then the polygon coordinates are assumed to be in meters.}

\item{units}{one of
\describe{
    \item{\code{"km2"}:}{Squared kilometers. For this method, \code{total.area.km2} is required.}
    \item{\code{"abs"}:}{Absolute numbers based on polygon coordinates. Only usefull if the projection satisfies the equal-area property. Note: these are just the \code{area} slots of the polygons, where the \code{area} slots of the holes are subtracted.}
    \item{\code{"prop"}:}{Proportional numbers. In other words, the total of the area sizes equals one.}
    \item{\code{"norm"}:}{Normalized numbers. All area sizes are normalized to the largest area, of which the area size equals one.}}
    The default method is \code{"abs"}, unless \code{total.area.km2} is specified (in that case, it is \code{"km2"}).}
}
\value{
numeric vector of area sizes
}
\description{
This function approximates the area sizes of the polygons either in 1) squared kilometers, 2) absolute numbers based on the polygon coordinates, 3) proportional numbers, 4) normalized numbers.
}
\details{
To approximate the sizes in squared kilometer, \code{total.area.km2} is required. Note that this method is an approximation, since it depends on the used projection and the level of detail of the SpatialPolygons object. Projections with equal-area property are highly recommended.
}

