% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bb.R
\name{bb}
\alias{bb}
\title{Bounding box generator}
\usage{
bb(x = NA, ext = NULL, xlim = NULL, ylim = NULL, relative = FALSE,
  current.projection = NULL, projection = NULL, q = NULL)
}
\arguments{
\item{x}{Either a shape (from class \code{\link[sp:Spatial]{Spatial}} or \code{\link[raster:Raster-class]{Raster}}) or a bounding box (either 2 by 2 matrix or an \code{\link[raster:Extent]{Extent}} object). If \code{x} is not specified, a bounding box can be created with \code{xlim} and \code{ylim} or with \code{q}.}

\item{ext}{Extension factor of the bounding box. If 1, the bounding box is unchanged. Values smaller than 1 reduces the bounding box, and values larger than 1 enlarges the bounding box.}

\item{xlim}{limits of the x-axis. These are either absolute or relative (depending on the argument \code{relative}).}

\item{ylim}{limits of the y-axis. See \code{xlim}.}

\item{relative}{boolean that determines whether relative values are used for \code{xlim} and \code{ylim} or absolute.}

\item{current.projection}{projection string (see \code{\link{set_projection}}) of the that corresponds to the}

\item{projection}{projection string (see \code{\link{set_projection}}) to transform the bounding box to.}

\item{q}{Open Street Map search query. The bounding is automatically generated by searching for \code{q} in Open Street Map Nominatim. See \url{http://wiki.openstreetmap.org/wiki/Nominatim}.}
}
\description{
Retrieves, modifies, or generates a bounding box. Existing bounding boxes can be retrieved (like \code{\link[sp:bbox]{bbox}}). They can be easily modified in two ways: either by using extension factor (\code{ext}), or by changing the x and y limits. New bounding boxes can be created by setting x and y limits in absolute values. Alternatively, bouding boxes can be automatically generated by a search query.
}
\examples{
## load Netherlands municipality shape
data(NLD_muni)

## get bounding box (similar to sp's function bbox)
bb(NLD_muni)

## extent it by factor 1.10
bb(NLD_muni, ext=1.10)

## convert to longlat
bb(NLD_muni, projection="longlat")

## find with Open Street Map Nominatim:
bb(q="Netherlands")

## load World shape and plot the Netherlands
data(World)
tm_shape(World, bbox=bb(q="Netherlands", projection = "eck4")) +
	tm_polygons() +
	tm_layout("Kingdom of the Netherlands (also in the Caribbean)")
}

