% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert_shape_data}
\alias{convert_shape_data}
\title{Convert shape data}
\usage{
convert_shape_data(shp.from, shp.to, variables.from = NULL,
  variables.to = NULL)
}
\arguments{
\item{shp.from}{the shape object to be converted. It sould contain data.}

\item{shp.to}{the shape object to be converted into.}

\item{variables.from}{names of the variable of \code{shp.from} to be converted. If missing, all numeric variables are taken.}

\item{variables.to}{variable names to be used. Should be the same number of variable names as \code{variables.from}}
}
\value{
shape object \code{shp.to} with converted data from \code{shp.from}
}
\description{
Convert data from one polygon shape to another. It uses an intersection matrix, which stores the intersection ratios of the two shape objects per polygon
}
\examples{
# \\\dontrun{
data(NLD_muni)
data(NLD_prov)
data(NLD_ageGroups)

## append age group data to municipalities and plot it
NLD_muni <- append_data(NLD_muni, NLD_ageGroups, key.data="muni", key.shp="name")
geo_shape(NLD_muni) +
	geo_fill(c("pop_0_20", "pop_20_65", "pop_65p"), convert2density=TRUE, total.area.km2=41543) +
	geo_shape(NLD_prov) +
	geo_borders() +
	geo_facets(free.scales=TRUE) +
	geo_theme_NLD(c("Population 0 to 20", "Population 20 to 65", "Population 65 and older"),inner.margins=c(.05, .2, .7, .05), legend.width=.8, legend.max.height=.5, legend.position=c("left", "top"))


## aggregate age group data to provinces
NLD_prov <- convert_shape_data(NLD_muni, NLD_prov, c("pop_0_20", "pop_20_65", "pop_65p"))
geo_shape(NLD_prov) +
	geo_fill(c("pop_0_20", "pop_20_65", "pop_65p"), convert2density=TRUE, total.area.km2=41543) +
	geo_borders() +
	geo_facets(free.scales=TRUE) +
	geo_theme_NLD(c("Population 0 to 20", "Population 20 to 65", "Population 65 and older"),inner.margins=c(.05, .2, .7, .05), legend.width=.8, legend.max.height=.5, legend.position=c("left", "top"))

## convert age group data to municipalities again
NLD_muni <- convert_shape_data(NLD_prov, NLD_muni, c("pop_0_20", "pop_20_65", "pop_65p"), 
						  variables.to=c("pop_0_20_new", "pop_20_65_new", "pop_65p_new"))
geo_shape(NLD_muni) +
	geo_fill(c("pop_0_20_new", "pop_20_65_new", "pop_65p_new"), convert2density=TRUE, total.area.km2=41543) +
	geo_borders() +
	geo_facets(free.scales=TRUE) +
	geo_theme_NLD(c("Population 0 to 20", "Population 20 to 65", "Population 65 and older"),inner.margins=c(.05, .2, .7, .05), legend.width=.8, legend.max.height=.5, legend.position=c("left", "top"))
# }
}

