% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{animation_tmap}
\alias{animation_tmap}
\alias{save_tmap}
\alias{style_catalogue}
\alias{style_catalog}
\alias{last_map}
\title{Deprecated tmap functions}
\usage{
animation_tmap(...)

save_tmap(...)

style_catalogue()

style_catalog()

last_map()
}
\arguments{
\item{...}{arguments passed on to new functions}
}
\description{
For consistency, tmap function names will be prefixed with a tm_ or tmap_. Therefore, function names such as \code{animation_tmap} have been renamed to \code{tmap_animation}.
}
