% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/double_line.R
\name{double_line}
\alias{double_line}
\title{Create a double line}
\usage{
double_line(shp, width)
}
\arguments{
\item{shp}{The shape object that contains the lines (\code{\link[sp:SpatialLinesDataFrame]{SpatialLinesDataFrame}})}

\item{width}{Width between the double lines}
}
\value{
SpatialLinesDataFrame
}
\description{
Create a double line from a single line. (Experimental)
}

