% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_functions.R
\name{offset_line}
\alias{double_line}
\alias{offset_line}
\title{Create a double line or offset line}
\usage{
offset_line(shp, offset)

double_line(shp, width, sides = "both")
}
\arguments{
\item{shp}{SpatialLines(DataFrame)}

\item{offset}{offset from the original lines}

\item{width}{width between the left and righthand side}

\item{sides}{character value that specifies which sides are selected: \code{"both"}, \code{"left"}, or \code{"right"}. The default is \code{"both"}. For the other two options, see also the shortcut function \code{offset_line}.}
}
\description{
Create a double line or offset line. The double line can be useful for visualizing two-way tracks or emulating objects such as railway tracks. The offset line can be useful to prevent overlapping of spatial lines.
}

