% Generated by roxygen2 (4.0.1): do not edit by hand
\name{double_line}
\alias{double_line}
\title{Create a double line from a single line.}
\usage{
double_line(shp, width)
}
\arguments{
\item{shp}{The shape object that contains the lines}

\item{width}{Width between the double lines}
}
\value{
SpatialLinesDataFrame
}
\description{
Create a double line from a single line.
}

