% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{geo-package}
\alias{geo-package}
\title{Tools to create thematic maps}
\description{
\tabular{ll}{
Package: \tab geo\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2014-04-25\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\details{
This packages offers a complete workaround to create thematic maps. Written on top of some great packages on spatial data such as \link[sp:sp]{sp} and \href{http://trac.osgeo.org/geos/}{rgeos}, it offers easy to use functions to
\itemize{
\item load shape files with \code{\link{get_shape}};
\item append data with \code{\link{append_data}};
\item change map projections with \code{\link{set_projection}};
\item crop maps with \code{\link{crop_shape}};
\item and approximate area sizes with \code{\link{approx_areas}}}.
The plotting itself is written in a ggplot2 style:
\itemize{
\item Stack drawing elements intuitively with great flexibility. The name of each element starts with \code{geo_}. The required element is \code{geo_shape}, the plain drawing elements are \code{\link{geo_borders}}, \code{\link{geo_fill}}, and \code{\link{geo_bubbles}}, the statistical drawing elements are \code{\link{geo_choropleth}} and \code{\link{geo_bubblemap}}. These elements together form one layer. Multiple layers can be stacked. The theme element is \code{\link{geo_theme}} and the element to control small multiples is \code{\link{geo_grid}}.
\item Create quick maps (like qplot) with \code{\link{geo}}. All the elements above can be called within one function call, although only one layer is supported.
}
A crucial differences with ggplot2 is that the elements are functional building blocks rather than layers from the grammar of graphics.
This package includes ready to use shape files from the World, Europe and the Netherlands (both provinces as municipalities), which are used in the examples.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{GIS}
\keyword{cartography}

