% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{geo-package}
\alias{geo-package}
\title{Tools to create thematic maps}
\description{
\tabular{ll}{
Package: \tab geo\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2014-05-22\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\details{
This packages offers a complete workaround to create thematic maps. Maps are plotted in \pkg{ggplot2} style:
\code{data(World)}
\code{geo_shape(World) +}
\code{geo_fill("pop_est_dens", style="kmeans", palette="YlOrRd") +}
\code{geo_borders() +}
\code{geo_text("iso_a3", cex="AREA", cex.lowerbound=.4, bg.alpha=0) +}
\code{geo_theme_World(title="Population density per km2")}

The name of each element starts with \code{geo_}. The required element is \code{\link{geo_shape}}, which load the shape object (in this case \code{\link{World}}).

The possible drawing elements are \code{\link{geo_borders}}, \code{\link{geo_fill}}, \code{\link{geo_bubbles}}, \code{\link{geo_lines}}, and \code{\link{geo_text}}. One or more of these elements stacked form one layer. Multiple layers can be stacked.

The other elements are \code{\link{geo_theme}} to control the layout, \code{\link{geo_grid}} to draw grid lines, and \code{\link{geo_facets}} to control small multiples.

Create quick maps (like qplot) with \code{\link{geo}}. All the elements above can be called within one function call (although only one layer is supported):
\code{geo(World, fill="pop_est_dens", theme="World", style="kmeans", title="Population per km")}

Note: a crucial difference with ggplot2 is that the elements are functional building blocks rather than layers from the grammar of graphics.

Moreover, this package offers easy to use functions to
\itemize{
\item load shape files with \code{\link{get_shape}};
\item append data with \code{\link{append_data}};
\item change map projections with \code{\link{set_projection}};
\item crop maps with \code{\link{crop_shape}};
\item and approximate area sizes with \code{\link{approx_areas}}}.

This package includes ready to use shape files from the World, Europe and the Netherlands (both provinces as municipalities), which are used in the examples.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{GIS}
\keyword{cartography}

