% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geo}
\alias{geo}
\title{Quick map plot}
\usage{
geo(shp, fill = "grey90", bubble.size = NULL, bubble.col = NULL,
  borders = "grey40", text = NULL, theme = NULL, ...)
}
\arguments{
\item{shp}{shape object. For \code{\link{geo_fill}} and \code{\link{geo_bubbles}}, a \code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygonsDataFrame}} or a \code{\link[sp:SpatialPointsDataFrame]{SpatialPointsDataFrame}} is requied. \code{\link[sp:SpatialPoints]{SpatialPoints}} and \code{\link[sp:SpatialPointsDataFrame]{SpatialPointsDataFrame}} are only used for \code{\link{geo_bubbles}}.}

\item{fill}{either a color to fill the polygons, or name of the data variable in \code{shp} to draw a choropleth.}

\item{bubble.size}{name of the data variable in \code{shp} for the bubblemap that specifies the sizes of the bubbles. If neither \code{bubble.size} nor \code{bubble.col} is specified, no bubblemap is drawn.}

\item{bubble.col}{name of the data variable in \code{shp} for the bubblemap that specifies the colors of the bubbles. If neither \code{bubble.size} nor \code{bubble.col} is specified, no bubblemap is drawn.}

\item{borders}{color of the polygon borders. Use \code{NA} to omit the borders.}

\item{text}{name of a variable in \code{shp} that contain text labels}

\item{theme}{one of "World", "Europe", or "NLD"}

\item{...}{parameters passed on to the \code{geo_*} functions.}
}
\value{
\code{\link{geo-element}}
}
\description{
This function is a convenient wrapper for drawing quick cartographic maps.
}
\examples{
data(Europe)

# just the map
geo(Europe)

# choropleth
geo(Europe, choro.fill="gdp_cap_est", theme="Europe", style="kmeans", title="Estimated GDP per capita")


# bubblemap
geo(Europe, bubble.size="pop_est", bubble.col="part", scale=2, theme="Europe")


# World maps
data(World)
geo(World, choro.fill="pop_est_dens", theme="World", style="kmeans", title="Population per km")
}

