% Generated by roxygen2 (4.0.0): do not edit by hand
\name{geo_bubbles}
\alias{geo_bubbles}
\title{Draw bubblemap}
\usage{
geo_bubbles(size = 1, col = "blueviolet", border = NA, scale = 1,
  n = 5, style = "pretty", breaks = NULL, palette = NULL,
  labels = NULL, auto.palette.mapping = TRUE, contrast = 1,
  colorNA = "#FF1414", xmod = 0, ymod = 0, by = FALSE)
}
\arguments{
\item{size}{\code{shp} data variable that determines the bubble sizes. Multiple variable names create small multiples}

\item{col}{color(s) of the bubble. Either a color (vector), or categorical variable name(s). Multiple variable names create small multiples}

\item{border}{color of the bubble borders. If \code{NA} (default), no bubble borders are drawn.}

\item{scale}{bubble size multiplier number.}

\item{n}{preferred number of color scale classes. Only applicable when \code{col} is a numeric variable name.}

\item{style}{method to cut the color scale: "fixed", "equal", "pretty", "quantile", "kmeans". Only applicable when \code{col} is a numeric variable name.}

\item{breaks}{in case \code{style=="fixed"}, breaks should be specified}

\item{palette}{color palette (see \code{RColorBrewer::display.brewer.all}) for the bubbles. Only when \code{col} is set to a variable.}

\item{labels}{labels of the classes}

\item{auto.palette.mapping}{When diverging colour palettes are used (i.e. "RdBu") this method automatically maps colors to values such that the middle colors (mostly white or yellow) are assigned to values of 0, and the two sides of the color palette are assigned to negative respectively positive values.}

\item{contrast}{number between 0 and 1 (default) that determines the contrast of the palette. Only applicable when \code{auto.palette.mapping=TRUE} and \code{col} is a numeric variable name.}

\item{colorNA}{colour for missing values}

\item{xmod}{horizontal position modification of the bubbles, relatively where 0 means no modification, and 1 means the total width of the frame. Either a single number for all polygons, or a numeric variable in the shape data specifying a number for each polygon. Together with \code{ymod}, it determines position modification of the bubbles. In most coordinate systems (projections), the origin is located at the bottom left, so negative \code{xmod} move the bubbles to the left, and negative \code{ymod} values to the bottom.}

\item{ymod}{vertical position modification. See xmod.}

\item{by}{logical. If \code{TRUE} and \code{col} is a data variable, then small multiples are generated, one for each level (NOT WORKING YET)}
}
\value{
\code{\link{geo-element}}
}
\description{
This layer speficies a bubblemap. Both colors and sizes of the bubbles can be mapped to data variables.
}
\examples{
# World example
data(World)
geo_shape(World) +
	geo_fill() +
	geo_bubbles("pop_est", scale=2) +
	geo_theme_World("World population", legend.width=.4)

# Europe example
data(Europe)

geo_shape(Europe) + geo_borders() + geo_bubbles(size=.75, col="red") + geo_theme("Countries")
geo_shape(Europe) + geo_borders() + geo_bubbles(size=.75, col=ifelse(Europe$name=="Isle of Man", "red", "blue")) + geo_theme("Find Isle of Man...")

geo_shape(Europe) +
	geo_borders() +
	geo_fill() +
	geo_bubbles("gdp_md_est", palette="Set2", col="part", scale=2) + 
	geo_theme_Europe("GDP per country", legend.bubble.col.title="Part of Europe")


# Netherlands example
data(NLD_muni)
data(NLD_prov)

geo_shape(NLD_prov) +
	geo_borders() +
	geo_fill("name") +
	geo_shape(NLD_muni) +
	geo_bubbles(size="pop", col="steelblue",style="kmeans") +
	geo_theme_NLD(title="Population", legend.digits=0, legend.config="bubble.size", legend.width=.4, bg.color="white")

geo_shape(NLD_muni) +
	geo_borders() +
	geo_shape(NLD_prov) +
	geo_bubbles(col="name", size=2)
}

