% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geo_fill}
\alias{geo_fill}
\title{Draw choropleth}
\usage{
geo_fill(col = "grey90", palette = NULL, convert2density = FALSE,
  area = NULL, n = 5, style = "pretty", breaks = NULL, labels = NULL,
  auto.palette.mapping = TRUE, contrast = 1, max.categories = 12,
  colorNA = "grey65", thres.poly = 1e-05)
}
\arguments{
\item{col}{either a single color value or a name of the data variable that is contained in \code{shp}. In the latter case, a choropleth is drawn.}

\item{palette}{palette name. See \code{RColorBrewer::display.brewer.all()} for options. Use a \code{"-"} as prefix to reverse the palette. By default, \code{"RdYlGn"} is taken for numeric variables and \code{"Dark2"} for categorical variables.}

\item{n}{preferred number of classes (in case \code{col} is a numeric variable)}

\item{convert2density}{boolean that determines whether \code{col} is converted to a density variable. Should be \code{TRUE} when \code{col} consists of absolute numbers. The area size is either approximated from the shape object, or given by the argument \code{area}.}

\item{area}{Name of the data variable that contains the area sizes in squared kilometer.}

\item{style}{method to cut the color scale (in case \code{col} is a numeric variable): "fixed", "equal", "pretty", "quantile", "kmeans"}

\item{breaks}{in case \code{style=="fixed"}, breaks should be specified}

\item{labels}{labels of the classes}

\item{auto.palette.mapping}{When diverging colour palettes are used (i.e. "RdBu") this method automatically maps colors to values such that the middle colors (mostly white or yellow) are assigned to values of 0, and the two sides of the color palette are assigned to negative respectively positive values.}

\item{contrast}{number between 0 and 1 (default) that determines the contrast of the palette. Only applicable when \code{auto.palette.mapping=TRUE}}

\item{max.categories}{in case \code{col} is the name of a categorical variable, this value determines how many categories (levels) it can have maximally. If the number of levels is higher than \code{max.categories}, then levels are combined.}

\item{colorNA}{color used for missing values}

\item{thres.poly}{number that specifies the threshold at which polygons are taken into account. The number itself corresponds to the proportion of the area sizes of the polygons to the total polygon size.}
}
\value{
\code{\link{geo-element}}
}
\description{
This layer speficies a choropleth. A color palette is mapped to a data variable. By default, a divering color palette is used for numeric variables and a qualitative palette for categorical variables.
}
\examples{
# World examples
data(World)

geo_shape(World) + geo_fill("green3") + geo_theme_World(title="A green World")

World$highlighted <- ifelse(World$iso_a3 \%in\% c("GRL", "AUS"), "gold", "gray75")
geo_shape(World, projection="merc") + geo_fill("highlighted") + geo_borders() + geo_theme("Mercator projection. Although used in Google Maps, it is discouraged for\\nstatistical purposes. In reality, Australia is 3 times larger than Greenland!", inner.margins=c(0,0,.1,0), title.cex=.6)

geo_shape(World) +
	geo_fill("pop_est_dens", style="kmeans", palette="YlOrRd", auto.palette.mapping=FALSE) +
	geo_borders() +
	geo_text("iso_a3", cex="AREA", cex.lowerbound=.4, bg.alpha=0) +
	geo_theme_World(title="Population density per km2", title.bg.color=TRUE)

geo_shape(World) +
	geo_fill("income_grp", palette="-Blues") +
	geo_borders() +
	geo_text("iso_a3", cex="AREA", scale=1.5, bg.alpha=0) +
	geo_theme_World("Income classification")

# Europe example
data(Europe)
geo_shape(Europe) +
	geo_fill("gdp_cap_est", style="kmeans") +
	geo_borders() +
	geo_text("iso_a3", cex="AREA", scale=2, bg.alpha=0) +
	geo_theme_Europe("GDP per capita")

# Netherlands examples
data(NLD_muni)
data(NLD_prov)

geo_shape(NLD_prov) + 
	geo_fill("name") + 
	geo_shape(NLD_muni) + 
	geo_borders() + 
	geo_shape(NLD_prov) + 
	geo_borders(lwd=2) +
	geo_text("name") +
	geo_theme_NLD("Provinces and municipalities", legend.show=FALSE)

geo_shape(NLD_muni) +
	geo_fill(col="population", convert2density=TRUE, style="kmeans") +
	geo_borders() +
	geo_shape(NLD_prov) +
	geo_borders(, lwd=2) +
	geo_theme_NLD(title="Population (per km2)", legend.digits=0)


}

