\name{geo_grid}
\alias{geo_grid}
\title{Coordinate grid lines}
\usage{
geo_grid(n.x = 8, n.y = 8, col = "grey50", labels.cex = 0.75,
  labels.col = "grey20", on.top = TRUE)
}
\arguments{
  \item{n.x}{Prefered number of grid lines for the x axis.}

  \item{n.y}{Prefered number of grid lines for the y axis.}

  \item{col}{Color for the grid lines.}

  \item{labels.cex}{font size of the tick labels}

  \item{labels.col}{font color fo the tick labels}

  \item{on.top}{Boolean that determines whether the grid
  lines are drawn op top of the map (\code{TRUE}) or under
  the map (\code{FALSE}).}
}
\description{
This layer draws coordinate grid lines.
}

