% Generated by roxygen2 (4.0.0): do not edit by hand
\name{geo_text}
\alias{geo_text}
\title{Add text labels}
\usage{
geo_text(text, cex = 1, fontcolor = NA, fontface = "plain",
  fontfamily = "sans", bg.color = "#888888", bg.alpha = 100,
  cex.lowerbound = 0.2, print.tiny = FALSE, scale = 1, xmod = 0,
  ymod = 0)
}
\arguments{
\item{text}{name of the variable in the shape object that contains the text labels}

\item{cex}{relative size of the text labels. Eiter one number, a name of a numeric variable in the shape data that is used to scale the sizes proportionally, or one of the following special cases
\describe{
        \item{\code{"AREA"}:}{text size is proportional to the squared root of the area size of the polygons.}
        \item{\code{"AREAx"}:}{text size is proportional to the x-th root of the area size of the polygons.}}}

\item{fontcolor}{relative size of the text labels}

\item{fontface}{font face of the text labels}

\item{fontfamily}{font family of the text labels}

\item{bg.color}{background color of the text labels}

\item{bg.alpha}{number between 0 and 255 that specifies the transparancy of the text background (0 is totally transparent, 255 is solid background)}

\item{cex.lowerbound}{lowerbound for \code{cex}. Needed to ignore the tiny labels in case \code{cex} is a variable.}

\item{print.tiny}{boolean that determines if tiny labels (which size is smaller than \code{cex.lowerbound}) are print at size \code{cex.lowerbound}}

\item{scale}{scalar needed in case cex is based}

\item{xmod}{horizontal position modification of the text, relatively where 0 means no modification, and 1 means the total width of the frame. Either a single number for all polygons, or a numeric variable in the shape data specifying a number for each polygon. Together with \code{ymod}, it determines position modification of the text labels. In most coordinate systems (projections), the origin is located at the bottom left, so negative \code{xmod} move the text to the left, and negative \code{ymod} values to the bottom.}

\item{ymod}{vertical position modification. See xmod.}
}
\value{
\code{\link{geo-element}}
}
\description{
This layer adds text labels
}
\examples{
# Europe example
data(Europe)
geo_shape(Europe) +
	geo_fill("gdp_cap_est", style="kmeans") +
	geo_borders() +
	geo_text("iso_a3", cex="AREA3", scale=2, bg.alpha=0) +
	geo_theme_Europe("GDP per capita")
}

