\name{geo_theme}
\alias{geo_theme}
\title{Theme elements of map plots}
\usage{
geo_theme(title = NULL, title.cex = 1.5, bg.color = "yellow",
  draw.frame = FALSE, crop = draw.frame, show.legend.text = NULL,
  type.legend.plot = NULL, legend.position = c("left", "top"),
  legend.plot.size = NA, legend.in.frame = TRUE, legend.cex = 0.8,
  legend.digits = 2, title.position = c("left", "top"), margins = NA,
  frame.lwd = 1, legend.only = FALSE)
}
\arguments{
  \item{title}{Title of the map(s)}

  \item{title.cex}{}

  \item{bg.color}{}

  \item{draw.frame}{}

  \item{crop}{boolean that determines whether the shape
  objects are cropped at the bounding box (see
  \code{\link[sp:bbox]{bbox}})}

  \item{show.legend.text}{}

  \item{type.legend.plot}{}

  \item{legend.position}{}

  \item{legend.plot.size}{}

  \item{legend.cex}{}

  \item{legend.digits}{}

  \item{title.position}{}

  \item{margins}{}

  \item{frame.lwd}{}

  \item{legend.only}{}
}
\description{
This layer specifies thematic layout options for the maps.
}

