% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_asp_ratio.R
\name{get_asp_ratio}
\alias{get_asp_ratio}
\title{Get aspect ratio}
\usage{
get_asp_ratio(shp)
}
\arguments{
\item{shp}{shape object, which is one of
\enumerate{
 \item{\code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygons(DataFrame)}}}
 \item{\code{\link[sp:SpatialPointsDataFrame]{SpatialPoints(DataFrame)}}}
 \item{\code{\link[sp:SpatialLinesDataFrame]{SpatialLines(DataFrame)}}}
 \item{\code{\link[sp:SpatialGridDataFrame]{SpatialGrid(DataFrame)}}}
 \item{\code{\link[sp:SpatialPixelsDataFrame]{SpatialPixels(DataFrame)}}}
}}
}
\value{
aspect ratio
}
\description{
Get the aspect ratio of a shape object, i.e., the map width devided by the map height.
}

