% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IDs.R
\name{get_IDs}
\alias{get_IDs}
\title{Get ID's of the shape items}
\usage{
get_IDs(shp)
}
\arguments{
\item{shp}{shape object, which is one of
\enumerate{
 \item{\code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygons(DataFrame)}}}
 \item{\code{\link[sp:SpatialPointsDataFrame]{SpatialPoints(DataFrame)}}}
 \item{\code{\link[sp:SpatialLinesDataFrame]{SpatialLines(DataFrame)}}}
}}
}
\value{
vector of ID's
}
\description{
Get ID's of the shape items. For polygons and lines, the ID attribute is used. For points, the coordinates are used.
}

