% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_polygon_ranges.R
\name{get_polygon_ranges}
\alias{get_polygon_ranges}
\title{Get ranges of the polygons}
\usage{
get_polygon_ranges(shp, key = NULL)
}
\arguments{
\item{shp}{shape object}

\item{key}{variable name in the shape data that identifies the polygons. If not specified, the default ID is used (see \code{\link{get_IDs}}).}
}
\value{
List containing a data.frame of ranges per polygon, and a range vector of the total shape object.
}
\description{
Get the ranges of the polyons in the current projection. Also the total range of the shape object is returned.
}

