% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{metro}
\alias{metro}
\title{Spatial data of metropolitan areas}
\source{
\url{https://esa.un.org/unpd/wup/}
}
\usage{
data(metro)
}
\description{
Spatial data of metropolitan areas, of class \code{\link[sf:sf]{sf}}. The data includes a population times series from 1950 to (forecasted) 2030. All metro areas with over 1 million inhabitants in 2010 are included.
}
\references{
United Nations, Department of Economic and Social Affairs, Population Division (2014). World Urbanization Prospects: The 2014 Revision, CD-ROM Edition.
}
