% Generated by roxygen2 (4.0.1): do not edit by hand
\name{+.geo}
\alias{+.geo}
\title{Stacking of geo elements}
\usage{
\method{+}{geo}(e1, e2)
}
\arguments{
\item{e1}{first \code{\link{geo-element}}}

\item{e2}{second \code{\link{geo-element}}}
}
\description{
The plus operator allows you to stack \code{\link{geo-element}s}. Always start with \code{\link{geo_shape}} to specify the shape object. If multile layers are used, each layer should start with a \code{\link{geo_shape}} element.
}

