% Generated by roxygen2 (4.0.2): do not edit by hand
\name{+.tmap}
\alias{+.tmap}
\title{Stacking of tmap elements}
\usage{
\method{+}{tmap}(e1, e2)
}
\arguments{
\item{e1}{first \code{\link{tmap-element}}}

\item{e2}{second \code{\link{tmap-element}}}
}
\description{
The plus operator allows you to stack \code{\link{tmap-element}s}.
}
\seealso{
\href{../doc/tmap-nutshell.html}{\code{vignette("tmap-nutshell")}}
}

