% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_misc_elements.R
\name{+.tmap}
\alias{+.tmap}
\title{Stacking of tmap elements}
\usage{
\method{+}{tmap}(e1, e2)
}
\arguments{
\item{e1}{first \code{\link{tmap-element}}}

\item{e2}{second \code{\link{tmap-element}}}
}
\description{
The plus operator allows you to stack \code{\link{tmap-element}s}, and groups of \code{\link{tmap-element}s}.
}
\seealso{
\code{\link{tmap-element}} and \href{../doc/tmap-nutshell.html}{\code{vignette("tmap-nutshell")}}
}
