% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print_tmap.R
\name{print.tmap}
\alias{print.tmap}
\title{Draw tmap plot}
\usage{
\method{print}{tmap}(x, ...)
}
\arguments{
\item{x}{tmap object. A tmap object is created with \code{\link{qtm}} or by stacking \code{\link{tmap-element}}s.}

\item{...}{not used}
}
\description{
Draw tmap plot on current graphics device
}

