% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tmap}
\alias{print.tmap}
\title{Draw thematic map}
\usage{
\method{print}{tmap}(x, return.asp = FALSE, show = TRUE, vp = NULL, ...)
}
\arguments{
\item{x}{tmap object.}

\item{return.asp}{should the aspect ratio be returned?}

\item{show}{show the map}

\item{vp}{viewport (for \code{"plot"} mode)}

\item{...}{not used}
}
\description{
Draw thematic map.
}
