% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tmap.R
\name{print.tmap}
\alias{print.tmap}
\title{Draw thematic map}
\usage{
\method{print}{tmap}(x, vp = NULL, return.asp = FALSE,
  mode = getOption("tmap.mode"), ...)
}
\arguments{
\item{x}{tmap object. A tmap object is created with \code{\link{qtm}} or by stacking \code{\link{tmap-element}}s.}

\item{vp}{\code{\link[grid:viewport]{viewport}} to draw the plot in. This is particularly useful for insets.}

\item{return.asp}{Logical that determines whether the aspect ratio of the map is returned. In that case, \code{\link[grid:grid.newpage]{grid.newpage()}} will be called, but without plotting of the map. This is used by \code{\link{save_tmap}} to determine the aspect ratio of the map.}

\item{...}{not used}

\item{plot}{should \code{x} be plot?}

\item{interactive}{should the plot be prepared for interaction? If \code{TRUE}, latitude longitude ( WGS84) coordinates are used and margins will be set to 0.}
}
\value{
A list of data.frames is silently returned, containing all ID and aesthetic variables per layer group.
}
\description{
Draw thematic map on current graphics device
}

