% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tmap}
\alias{print.tmap}
\title{Draw thematic map}
\usage{
\method{print}{tmap}(x, vp = NULL, ...)
}
\arguments{
\item{x}{tmap object.}

\item{vp}{viewport (for \code{"plot"} mode)}
}
\description{
Draw thematic map.
}
