% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_tmap.R
\name{print.tmap}
\alias{print.tmap}
\title{Draw thematic map}
\usage{
\method{print}{tmap}(x, vp = NULL, ...)
}
\arguments{
\item{x}{tmap object. A tmap object is created with \code{\link{qtm}} or by stacking \code{\link{tmap-element}}s.}

\item{vp}{\code{\link[grid:viewport]{viewport}} to draw the plot in. This is particularly useful for insets.}

\item{...}{not used}
}
\description{
Draw thematic map on current graphics device
}

