% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tmap}
\alias{print.tmap}
\alias{knit_print.tmap}
\title{Draw thematic map}
\usage{
\method{print}{tmap}(
  x,
  return.asp = FALSE,
  show = TRUE,
  vp = NULL,
  knit = FALSE,
  options = NULL,
  in.shiny = FALSE,
  proxy = FALSE,
  ...
)

\method{knit_print}{tmap}(x, ..., options = NULL)
}
\arguments{
\item{x}{tmap object.}

\item{return.asp}{should the aspect ratio be returned?}

\item{show}{show the map}

\item{vp}{viewport (for \code{"plot"} mode)}

\item{knit}{A logical, should knit?}

\item{options}{A vector of options}

\item{in.shiny}{A logical, is the map drawn in \strong{shiny}?}

\item{proxy}{A logical, if \code{in.shiny}, is \code{\link[=tmapProxy]{tmapProxy()}} used?}

\item{...}{passed on internally (for developers: in \code{"view"} mode, the proxy leaflet object is passed to \code{tmapLeafletInit}).}
}
\description{
Draw thematic map
}
