% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Spatial data of rivers}
\source{
\url{http://www.naturalearthdata.com}
}
\usage{
data(rivers)
}
\description{
Spatial data of rivers, of class \code{\link[sp:SpatialLinesDataFrame]{SpatialLinesDataFrame}}
}
