% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tmap.R
\name{save_tmap}
\alias{save_tmap}
\title{Save tmap}
\usage{
save_tmap(tm, filename = shp_name(tm), width = NA, height = NA,
  units = c("in", "cm", "mm", "px"), dpi = 300, outer.margins = 0,
  asp = 0, scale = NA, insets_tm = NULL, insets_vp = NULL, ...)
}
\arguments{
\item{tm}{tmap object}

\item{filename}{filename including extension, and optionally the path. The extensions pdf, eps, svg, wmf (Windows only), png, jpg, bmp, or tiff are supported.}

\item{width}{width. Units are set with the argument \code{units}. If set to \code{NA} and \code{height} is specified, it will be \code{height} * aspect ratio. If both \code{width} and \code{height} are not specified, then the width of the current plotting window will be taken.}

\item{height}{height. Units are set with the argument \code{units}. If set to \code{NA} and \code{width} is specified, it will be \code{width} / aspect ratio. If both \code{width} and \code{height} are not specified, then the height of the current plotting window will be taken.}

\item{units}{units for width and height when either one is explicitly specified (in, cm, or mm)}

\item{dpi}{dots per inch. Only applicable for raster graphics.}

\item{outer.margins}{overrides the outer.margins argument of \code{\link{tm_layout}} (unless set to \code{NA})}

\item{asp}{overrides the asp argument of \code{\link{tm_layout}} (unless set to \code{NA})}

\item{scale}{overrides the scale argument of \code{\link{tm_layout}} (unless set to \code{NA})}

\item{insets_tm}{tmap object of an inset map, or a list of tmap objects of multiple inset maps. The number of tmap objects should be equal to the number of viewports specified with \code{insets_vp}.}

\item{insets_vp}{\code{\link[grid:viewport]{viewport}} of an inset map, or a list of \code{\link[grid:viewport]{viewport}}s of multiple inset maps. The number of viewports should be equal to the number of tmap objects specified with \code{insets_tm}.}

\item{...}{arguments passed on to device functions}
}
\description{
Save tmap to a file, such as png, jpg, or pdf.
}
\examples{
\dontrun{
data(NLD_muni, NLD_prov)
(tm_shape(NLD_muni) +
	 tm_fill(col="population", convert2density=TRUE, 
	   style="kmeans", title="Population (per km2)", legend.hist=FALSE) +
	 tm_borders("black", alpha=.5) + 
tm_shape(NLD_prov) +
	 tm_borders("grey25", lwd=2) +
tm_format_NLD(inner.margins = c(.02, .15, .06, .15)) + 
tm_scale_bar(position = c("left", "bottom")) +
tm_compass(position=c("right", "bottom")) + 
tm_style_classic()) \%>\% save_tmap()
}
}

