% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbind.R
\name{sbind}
\alias{sbind}
\title{Combine shape objects}
\usage{
sbind(...)
}
\arguments{
\item{...}{shape objects. Each shape object is one of
\enumerate{
 \item{\code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygons(DataFrame)}}}
 \item{\code{\link[sp:SpatialPointsDataFrame]{SpatialPoints(DataFrame)}}}
 \item{\code{\link[sp:SpatialLinesDataFrame]{SpatialLines(DataFrame)}}}
}}
}
\value{
shape object
}
\description{
Combine shape objects into one shape object. It works analogous to \code{\link{rbind}}.
}

