\docType{data}
\name{World}
\alias{Europe}
\alias{NLD_muni}
\alias{NLD_prov}
\alias{World}
\title{World, Europe and Netherlands map}
\description{
Maps of the world, Europe and the Netherlands (province and
municipality level).
}
\details{
\code{World} World map. The projection that is chosen for
this world map is Eckhart IV since area sizes are
preserved, which is a very important property for
statistical purposes.

\code{Europe} Europe map. Lambert azimuthal equal-area
projection is used by default for this map. Several
countries are transcontinental and are partly located in
Asia. From these countries, only Russia and Turkey have
been included in this map as part of Europe since they are
widely considered as European countries. Other
transcontinental countries Azerbaijan, Georgia, and
Kazakhstan, are also included in the map, but only
passively. From the other surrounding countries, only
Greenland is removed from the map, since it interferes with
the map title.

\code{NLD_prov} and \code{NLD_muni}, maps of the
Netherlands at province and municipality level of 2013. The
used projection is the Rijksdriehoekstelsel projection.
\strong{Important:} publication of these maps is only
allowed when cited to Statistics Netherlands (CBS) and
Kadaster Nederland as source.
}

