% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{rivers}
\alias{airports}
\alias{cities}
\alias{rivers}
\title{Spatial data of rivers, cities, and airports}
\source{
http://www.naturalearthdata.com/
}
\description{
Spatial data of main world cities, rivers, and airports.
}

