% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeTM.R
\name{shapeTM}
\alias{shapeTM}
\title{Internal tmap function to create a tmap shape}
\usage{
shapeTM(shp, tmapID = NULL, bbox = NULL, ...)
}
\arguments{
\item{shp}{Shape file}

\item{tmapID}{tmap Identifier}

\item{bbox}{bounding box}
}
\description{
Internal tmap function to create a tmap shape
}
\keyword{internal}
