% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/split_lines_equal.R
\name{split_lines_equal}
\alias{split_lines_equal}
\title{Split lines in segments of equal length}
\usage{
split_lines_equal(shp, dist = 1000, include.last = FALSE)
}
\arguments{
\item{shp}{The shape object that contains the lines, i.e. \code{\link[sp:SpatialLinesDataFrame]{SpatialLinesDataFrame}}}

\item{dist}{Distance per segment}

\item{include.last}{Include last point, even though the distance is less than dist from the previous point?}
}
\value{
Shape object, i.e. \code{\link[sp:SpatialLinesDataFrame]{SpatialPolygons(DataFrame)}}.
}
\description{
Split lines in segments of equal length. (Experimental, see note)
}
\note{
This function is still in experimental phase, which means that it may not be stable and it may be changed significantly in future versions. Moreover, it is unsure if it will stay in tmap; instead, it may be put in a different package, along with functions of similar tasks.
}

