% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_shps.R
\name{split.SpatialPolygonsDataFrame}
\alias{split.SpatialLinesDataFrame}
\alias{split.SpatialPointsDataFrame}
\alias{split.SpatialPolygonsDataFrame}
\title{Divide into multiple shape objects}
\usage{
\method{split}{SpatialPolygonsDataFrame}(x, f, drop = FALSE, ...)

\method{split}{SpatialPointsDataFrame}(x, f, drop = FALSE, ...)

\method{split}{SpatialLinesDataFrame}(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{shape object, which is one of
\enumerate{
 \item{\code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygons(DataFrame)}}}
 \item{\code{\link[sp:SpatialPointsDataFrame]{SpatialPoints(DataFrame)}}}
 \item{\code{\link[sp:SpatialLinesDataFrame]{SpatialLines(DataFrame)}}}
}}

\item{f}{factor to split \code{x}}

\item{drop}{unused factor levels are dropped}

\item{...}{other arguments (not used)}
}
\value{
List of shape objects.
}
\description{
Divide a shape object into multiple objects.
}

