% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split.SpatialPolygonsDataFrame}
\alias{split.SpatialLinesDataFrame}
\alias{split.SpatialPointsDataFrame}
\alias{split.SpatialPolygonsDataFrame}
\title{Divide into multiple shape objects}
\usage{
\method{split}{SpatialPolygonsDataFrame}(x, f, drop = FALSE, ...)

\method{split}{SpatialPointsDataFrame}(x, f, drop = FALSE, ...)

\method{split}{SpatialLinesDataFrame}(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{shape object}

\item{f}{factor to split \code{x}}

\item{drop}{unused factor levels are dropped}

\item{...}{other arguments (not used)}
}
\value{
list of shape objects
}
\description{
Divide shape object into multiple objects
}

