% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_add_legend.R
\name{tm_add_legend}
\alias{tm_add_legend}
\title{Map component: manual legend}
\usage{
tm_add_legend(
  ...,
  labels,
  type = "symbols",
  title = "",
  design = NULL,
  orientation = NULL,
  group = NA,
  group.control = "check",
  resize.as.group = FALSE,
  z = NA_integer_
)
}
\arguments{
\item{...}{visual variables and arguments passed on to \code{tm_legend()}.
By default, the argument \code{type} is set to \code{"Symbols"}, which means that the
supported visual variables are: \code{"fill"}, \code{"col"}, \code{"shape"}, \code{"size"},
\code{"fill_alpha"}, \code{"col_alpha"}, \code{"lty"}, \code{"lwd"}, \code{"linejoin"}, and \code{"lineend"}.}

\item{labels}{labels}

\item{type}{the layer type from which the visual variables (see \code{...}) are taken.
Options: \code{"symbols"} (default), \code{"lines"}, \code{"polygons"}, and \code{"text"}.}

\item{title}{text of the title}

\item{design}{legend design}

\item{orientation}{legend orientation}

\item{group}{Name of the group to which this layer belongs. This is only
relevant in view mode, where layer groups can be switched (see \code{group.control})}

\item{group.control}{In view mode, the group control determines how
layer groups can be switched on and off. Options: \code{"radio"} for radio
buttons (meaning only one group can be shown), \code{"check"} for check boxes
(so multiple groups can be shown), and \code{"none"} for no control
(the group cannot be (de)selected).}

\item{resize.as.group}{resize.as.group}

\item{z}{z}
}
\description{
Map component that adds a manual legend
}
