% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_animate.R
\name{tm_animate}
\alias{tm_animate}
\alias{tm_animate_slow}
\title{Specify an animation (experimental)}
\usage{
tm_animate(
  by = "VARS__",
  nframes = 60L,
  fps = 20L,
  play = c("loop", "pingpong", "once"),
  dpr = 2,
  ...
)

tm_animate_slow(
  by = "VARS__",
  nframes = 60L,
  fps = 2L,
  play = c("loop", "pingpong", "once"),
  dpr = 2,
  ...
)
}
\arguments{
\item{by}{group by variable used to create the animation frames. Note: it is called \code{pages} in the core function \code{\link[=tm_facets]{tm_facets()}}.}

\item{nframes}{number of animation frames. This only app}

\item{fps}{frames per second. Default: 30 for \code{tm_facets_animate} and 2 for \code{tm_facets_animate_slow}.}

\item{play}{how should the animation be played? One of \code{"loop"}, \code{"pingpong"}, and \code{"once"}}

\item{dpr}{device pixel ratio. The ratio between the physical pixel density of a device and its logical pixel density.}

\item{...}{passed on to \code{\link[=tm_facets]{tm_facets()}}}
}
\description{
\code{tm_animate}
}
\seealso{
\code{\link[=tm_facets]{tm_facets()}} which is the core function, and \code{\link[=tmap_animation]{tmap_animation()}} used to save the animation
}
