% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_comp_group.R
\name{tm_comp_group}
\alias{tm_comp_group}
\title{Group components}
\usage{
tm_comp_group(
  id,
  position,
  stack,
  frame_combine,
  resize_as_group,
  stack_margin,
  offset,
  frame,
  bg.color,
  bg.alpha
)
}
\arguments{
\item{id}{id of the component group. Refers to the \code{group_id} argument of each component function, such as \code{\link[=tm_legend]{tm_legend()}} and \code{\link[=tm_title]{tm_title()}}.}

\item{stack}{stacking Horizontal or vertical}

\item{frame_combine}{put frame around all components that are drawn on the same location. Whether a frame is drawn is still decided by the \code{frame} argument of the  'main' (first) component.}

\item{resize_as_group}{in case a component if rescaled because of the limited space, rescale the other components proportionally?}

\item{stack_margin}{Margin between components}

\item{offset}{Offset margin between frame and the components block}

\item{frame}{Frane of the components. Is usually set in each component function, but if specified here, it will overwrite them.}

\item{bg.color}{Background color the components block. Is usually set in each component function, but if specified here, it will overwrite them.}

\item{bg.alpha}{Background alpha transparency of the components block. Is usually set in each component function, but if specified here, it will overwrite them.}
}
\value{
A \code{\link{tmap-element}}
}
\description{
Group components
}
