% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_compass}
\alias{tm_compass}
\title{Map component: compass}
\usage{
tm_compass(
  north,
  type,
  text.size,
  size,
  show.labels,
  cardinal.directions,
  text.color,
  color.dark,
  color.light,
  lwd,
  position,
  bg.color,
  bg.alpha,
  stack,
  just,
  frame,
  frame.lwd,
  frame.r,
  margins,
  z
)
}
\arguments{
\item{north}{north}

\item{type}{type}

\item{text.size}{text.size}

\item{size}{size}

\item{show.labels}{show.labels}

\item{cardinal.directions}{cardinal.directions}

\item{text.color}{text.color}

\item{color.dark}{color.dark}

\item{color.light}{color.light}

\item{lwd}{lwd}

\item{position}{position}

\item{bg.color}{bg.color}

\item{bg.alpha}{bg.alpha}

\item{stack}{stack}

\item{just}{just}

\item{frame}{frame}

\item{frame.lwd}{frame.lwd}

\item{frame.r}{frame.r}

\item{margins}{margins}

\item{z}{z}
}
\description{
Map component that adds a compass
}
