% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_compass}
\alias{tm_compass}
\title{Map component: compass}
\usage{
tm_compass(
  north,
  type,
  text.size,
  size,
  show.labels,
  cardinal.directions,
  text.color,
  color.dark,
  color.light,
  lwd,
  position,
  bg.color,
  bg.alpha,
  stack,
  just,
  frame,
  frame.lwd,
  frame.r,
  margins,
  z,
  ...
)
}
\arguments{
\item{north}{north}

\item{type}{Compass type}

\item{text.size}{text.size}

\item{size}{size}

\item{show.labels}{show.labels}

\item{cardinal.directions}{cardinal.directions}

\item{text.color}{text.color}

\item{color.dark}{color.dark}

\item{color.light}{color.light}

\item{lwd}{lwd}

\item{position}{Vector of two values, specifying the x and y coordinates. The first is either \code{"left"} or \code{"right"}, the second either \code{"top"} or \code{"bottom"}.}

\item{bg.color}{Background color}

\item{bg.alpha}{Background transparency}

\item{stack}{stack}

\item{just}{just}

\item{frame}{frame}

\item{frame.lwd}{frame line width}

\item{frame.r}{Radius of the rounded frame corners. 0 means no rounding.}

\item{margins}{margins}

\item{z}{z}

\item{...}{to catch deprecated arguments (alpha)}
}
\description{
Map component that adds a compass
}
