% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_credits}
\alias{tm_credits}
\title{Map component: (credits) text}
\usage{
tm_credits(
  text,
  size,
  color,
  padding,
  fontface,
  fontfamily,
  stack,
  just,
  frame,
  frame.lwd,
  frame.r,
  bg.color,
  bg.alpha,
  position,
  width,
  height,
  group.frame,
  resize.as.group,
  z
)
}
\arguments{
\item{text}{text of the title}

\item{size}{font size of the title}

\item{color}{color}

\item{padding}{padding}

\item{fontface}{font face}

\item{fontfamily}{font family}

\item{stack}{stack}

\item{just}{just}

\item{frame}{frame}

\item{frame.lwd}{frame.lwd}

\item{frame.r}{frame.r}

\item{bg.color}{bg.color}

\item{bg.alpha}{bg.alpha}

\item{position}{position}

\item{width}{width}

\item{height}{height}

\item{group.frame}{group.frame}

\item{resize.as.group}{resize.as.group}

\item{z}{z}
}
\description{
Map component that adds a text, typically used as credits
}
