% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_facets.R
\name{tm_facets}
\alias{tm_facets}
\alias{tm_facets_grid}
\alias{tm_facets_wrap}
\alias{tm_facets_pagewise}
\alias{tm_facets_stack}
\alias{tm_facets_hstack}
\alias{tm_facets_vstack}
\alias{tm_facet_flip}
\alias{tm_facets_flip}
\title{Facets}
\usage{
tm_facets(
  by = NULL,
  rows = NULL,
  columns = NULL,
  pages = NULL,
  as.layers = FALSE,
  nrows = NA,
  ncols = NA,
  byrow = TRUE,
  orientation = NA,
  free.coords = NA,
  drop.units = TRUE,
  drop.empty.facets = TRUE,
  drop.NA.facets = FALSE,
  sync = TRUE,
  showNA = NA,
  textNA = "Mssing",
  scale.factor = 2,
  type = NA,
  along = NULL
)

tm_facets_grid(rows = NULL, columns = NULL, pages = NULL, ...)

tm_facets_wrap(by = "VARS__", nrows = NA, ncols = NA, byrow = TRUE, ...)

tm_facets_pagewise(by = "VARS__", nrows = 1, ncols = 1, byrow = TRUE, ...)

tm_facets_stack(by = "VARS__", orientation = NA, ...)

tm_facets_hstack(by = "VARS__", ...)

tm_facets_vstack(by = "VARS__", ...)

tm_facets_flip()
}
\arguments{
\item{by}{Group by variable (only for a facet wrap or facet stack)}

\item{rows}{Variable that specifies the rows (only for a facet grid)}

\item{columns}{Variable that specifies the columns (only for a facet grid)}

\item{pages}{Variable that specifies the pages (only for a facet grid)}

\item{as.layers}{show facets as layers?}

\item{nrows}{Number of rows}

\item{ncols}{Number of columns}

\item{byrow}{Should facets be wrapped by row?}

\item{orientation}{For facet stack: horizontal or vertical?}

\item{free.coords}{Logical. If the \code{by} argument is specified, should each map has its own coordinate ranges? By default \code{TRUE}, unless facets are shown in as different layers (\code{as.layers = TRUE})}

\item{drop.units}{Logical. If the \code{by} argument is specified, should non-selected spatial units be dropped? If \code{FALSE}, they are plotted where mapped aesthetics are regarded as missing values. Not applicable for raster shapes. By default \code{TRUE}.}

\item{drop.empty.facets}{Logical. If the \code{by} argument is specified, should empty facets be dropped? Empty facets occur when the \code{by}-variable contains unused levels. When \code{TRUE} and two \code{by}-variables are specified, empty rows and columns are dropped.}

\item{drop.NA.facets}{Logical. If the \code{by} argument is specified, and all data values for specific facets are missing, should these facets be dropped? \code{FALSE} by default.}

\item{sync}{Logical. Should the navigation in view mode (zooming and panning) be synchronized? By default \code{TRUE} if the facets have the same bounding box. This is generally the case when rasters are plotted, or when free.coords is \code{FALSE}.}

\item{showNA}{If the \code{by} argument is specified, should missing values of the \code{by}-variable be shown in a facet? If two \code{by}-variables are specified, should missing values be shown in an additional row and column? If \code{NA}, missing values only are shown if they exist. Similar to the \code{useNA} argument of \link[base:table]{table()}, where \code{TRUE}, \code{FALSE}, and \code{NA} correspond to \code{"always"}, \code{"no"}, and \code{"ifany"} respectively.}

\item{textNA}{Text used for facets of missing values.}

\item{scale.factor}{Number that determines how the elements (e.g. font sizes, symbol sizes, line widths) of the small multiples are scaled in relation to the scaling factor of the shapes. The elements are scaled to the \code{scale.factor}th root of the scaling factor of the shapes. So, for \code{scale.factor=1}, they are scaled proportional to the scaling of the shapes. Since elements, especially text, are often too small to read, a higher value is recommended. By default, \code{scale.factor=2}.}

\item{type}{\code{"grid"}, \code{"wrap"} or \code{"stack"}}

\item{along}{deceprated Please use \code{tm_facets_page()}}

\item{...}{paseed on to \code{tm_facets()}}
}
\description{
Specify facets. \code{tm_facets()} is the core function, but recommended is to use \code{tm_facets_wrap()}, \code{tm_facets_stack()} or \code{tm_facets_grid()}. The former two specify facets for one grouping variable (so one faceting dimension). The difference is that wrap may place facets in multiple rows and columns whereas \code{tm_facets_stack()} stacks the facets either horizontally or vertically. \code{tm_facets_grid()} supports up to three faceting dimensions.
}
