% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tm_misc_elements.R
\name{tm_grid}
\alias{tm_grid}
\title{Coordinate grid lines}
\usage{
tm_grid(n.x = NA, n.y = NA, projection = NA, col = "grey50",
  alpha = NA, labels.size = 0.75, labels.col = "grey20",
  labels.inside.frame = TRUE, on.top = TRUE)
}
\arguments{
\item{n.x}{prefered number of grid lines for the x axis.}

\item{n.y}{prefered number of grid lines for the y axis.}

\item{projection}{projection character. If specified, the grid lines are projected accordingly. See \code{\link{set_projection}} for projection details. Many world maps are projected, but still have latitude longitude (\code{"longlat"}) grid lines.}

\item{col}{color for the grid lines.}

\item{alpha}{alpha transparency of the grid lines. Number between 0 and 1. By default, the alpha transparency of \code{col} is taken.}

\item{labels.size}{font size of the tick labels}

\item{labels.col}{font color fo the tick labels}

\item{labels.inside.frame}{Show labels inside the frame?}

\item{on.top}{Boolean that determines whether the grid lines are drawn op top of the map (\code{TRUE}) or under the map (\code{FALSE}).}
}
\description{
Creates a \code{\link{tmap-element}} that draws coordinate grid lines.
}

