% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tm_misc_elements.R
\name{tm_grid}
\alias{tm_grid}
\title{Coordinate grid lines}
\usage{
tm_grid(n.x = 8, n.y = 8, col = "grey50", labels.size = 0.75,
  labels.col = "grey20", labels.inside.frame = TRUE, on.top = TRUE)
}
\arguments{
\item{n.x}{Prefered number of grid lines for the x axis.}

\item{n.y}{Prefered number of grid lines for the y axis.}

\item{col}{Color for the grid lines.}

\item{labels.size}{font size of the tick labels}

\item{labels.col}{font color fo the tick labels}

\item{labels.inside.frame}{Show labels inside the frame?}

\item{on.top}{Boolean that determines whether the grid lines are drawn op top of the map (\code{TRUE}) or under the map (\code{FALSE}).}
}
\description{
Creates a \code{\link{tmap-element}} that draws coordinate grid lines.
}

