% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers.R
\name{tm_iso}
\alias{tm_iso}
\title{Draw iso (contour) lines with labels}
\usage{
tm_iso(col = NA, text = "level", size = 0.5, remove.overlap = TRUE,
  along.lines = TRUE, overwrite.lines = TRUE, group = NA, ...)
}
\arguments{
\item{col}{line color. See \code{\link{tm_lines}}.}

\item{text}{text to display. By default, it is the variable named \code{"level"} of the shape that is created with \code{\link[tmaptools:smooth_map]{smooth_map}}}

\item{size}{text size (see \code{\link{tm_text}})}

\item{remove.overlap}{see \code{\link{tm_text}}}

\item{along.lines}{see \code{\link{tm_text}}}

\item{overwrite.lines}{see \code{\link{tm_text}}}

\item{group}{name of the group to which this layer belongs in view mode. Each group can be selected or deselected in the layer control item. Groups can either be specified as base or overlay groups in \code{\link{tm_view}} (arguments \code{base.groups} and \code{overlay.groups}).}

\item{...}{arguments passed on to \code{\link{tm_lines}} or \code{\link{tm_text}}}
}
\description{
This function is a wrapper of \code{\link{tm_lines}} and \code{\link{tm_text}} aimed to draw isopleths, which can be created with \code{\link[tmaptools:smooth_map]{smooth_map}}.
}
\seealso{
\code{\link[tmaptools:smooth_map]{smooth_map}}
}
