% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layout.R, R/tmap_options.R
\name{tm_layout}
\alias{tm_layout}
\alias{tm_style}
\alias{tm_format}
\title{Layout options}
\usage{
tm_layout(
  scale,
  asp,
  bg.color,
  outer.bg.color,
  frame,
  frame.lwd,
  frame.r,
  frame.double.line,
  outer.margins,
  inner.margins,
  inner.margins.extra,
  meta.margins,
  meta.auto.margins,
  between.margin,
  component.offset,
  component.stack.margin,
  grid.mark.height,
  xylab.height,
  coords.height,
  xlab.show,
  ylab.show,
  xlab.pos,
  ylab.pos,
  grid.show,
  grid.label.pos,
  panel.type,
  panel.wrap.pos,
  panel.xtab.pos,
  unit,
  color.sepia.intensity,
  color.saturation,
  color.vision.deficiency.sim,
  fontface,
  fontfamily,
  legend.show,
  legend.design,
  legend.orientation,
  legend.position,
  legend.width,
  legend.height,
  legend.stack,
  legend.group.frame,
  legend.resize.as.group,
  legend.reverse,
  legend.title.color,
  legend.title.size,
  legend.title.fontface,
  legend.title.fontfamily,
  legend.title.just,
  legend.text.color,
  legend.text.size,
  legend.text.fontface,
  legend.text.fontfamily,
  legend.frame,
  legend.frame.lwd,
  legend.frame.r,
  legend.bg.color,
  legend.bg.alpha,
  compass.type,
  title.size,
  title.color,
  title.fontface,
  title.fontfamily,
  title.bg.color,
  title.bg.alpha,
  title.padding,
  title.frame,
  title.frame.lwd,
  title.frame.r,
  title.stack,
  title.position,
  title.width,
  title.heigth,
  title.group.frame,
  title.resize.as.group,
  panel.show,
  panel.labels,
  panel.label.size,
  panel.label.color,
  panel.label.fontface,
  panel.label.fontfamily,
  panel.label.bg.color,
  panel.label.height,
  panel.label.rot
)

tm_style(style, ...)

tm_format(format, ...)
}
\arguments{
\item{scale}{TODO}

\item{asp}{TODO}

\item{bg.color}{TODO}

\item{outer.bg.color}{TODO}

\item{frame}{TODO}

\item{frame.lwd}{TODO}

\item{frame.r}{TODO}

\item{frame.double.line}{TODO}

\item{outer.margins}{TODO}

\item{inner.margins}{TODO}

\item{inner.margins.extra}{TODO}

\item{meta.margins}{TODO}

\item{meta.auto.margins}{TODO}

\item{between.margin}{TODO}

\item{component.offset}{TODO}

\item{component.stack.margin}{TODO}

\item{grid.mark.height}{TODO}

\item{xylab.height}{TODO}

\item{coords.height}{TODO}

\item{xlab.show}{TODO}

\item{ylab.show}{TODO}

\item{xlab.pos}{TODO}

\item{ylab.pos}{TODO}

\item{grid.show}{TODO}

\item{grid.label.pos}{TODO}

\item{panel.type}{TODO}

\item{panel.wrap.pos}{TODO}

\item{panel.xtab.pos}{TODO}

\item{unit}{TODO}

\item{color.sepia.intensity}{TODO}

\item{color.saturation}{TODO}

\item{color.vision.deficiency.sim}{TODO}

\item{fontface}{TODO}

\item{fontfamily}{TODO}

\item{legend.show}{TODO}

\item{legend.design}{TODO}

\item{legend.orientation}{TODO}

\item{legend.position}{TODO}

\item{legend.width}{TODO}

\item{legend.height}{TODO}

\item{legend.stack}{TODO}

\item{legend.group.frame}{TODO}

\item{legend.resize.as.group}{TODO}

\item{legend.reverse}{TODO}

\item{legend.title.color}{TODO}

\item{legend.title.size}{TODO}

\item{legend.title.fontface}{TODO}

\item{legend.title.fontfamily}{TODO}

\item{legend.title.just}{TODO}

\item{legend.text.color}{TODO}

\item{legend.text.size}{TODO}

\item{legend.text.fontface}{TODO}

\item{legend.text.fontfamily}{TODO}

\item{legend.frame}{TODO}

\item{legend.frame.lwd}{TODO}

\item{legend.frame.r}{TODO}

\item{legend.bg.color}{TODO}

\item{legend.bg.alpha}{TODO}

\item{compass.type}{TODO}

\item{title.size}{TODO}

\item{title.color}{TODO}

\item{title.fontface}{TODO}

\item{title.fontfamily}{TODO}

\item{title.bg.color}{TODO}

\item{title.bg.alpha}{TODO}

\item{title.padding}{TODO}

\item{title.frame}{TODO}

\item{title.frame.lwd}{TODO}

\item{title.frame.r}{TODO}

\item{title.stack}{TODO}

\item{title.position}{TODO}

\item{title.width}{TODO}

\item{title.heigth}{TODO}

\item{title.group.frame}{TODO}

\item{title.resize.as.group}{TODO}

\item{panel.show}{TODO}

\item{panel.labels}{TODO}

\item{panel.label.size}{TODO}

\item{panel.label.color}{TODO}

\item{panel.label.fontface}{TODO}

\item{panel.label.fontfamily}{TODO}

\item{panel.label.bg.color}{TODO}

\item{panel.label.height}{TODO}

\item{panel.label.rot}{TODO}

\item{style}{name of the style}

\item{format}{name of the format}
}
\description{
Set of tmap options that are directly related to the layout.
}
\examples{
data(land, World)

tm_shape(land) +
	tm_raster("elevation", col.scale = tm_scale_intervals(breaks=c(-Inf, 250, 500, 1000, 1500, 2000, 2500, 3000, 4000, Inf),  
														  values = terrain.colors(9), 
														  midpoint = NA),
			  col.legend = tm_legend(title="Elevation", position = tm_pos_in("left", "bottom"))) +
tm_shape(World, is.main = TRUE, crs = "+proj=eck4") +
	tm_borders("grey20") +
	#tm_graticules(labels.size = .5) +
	#tm_text("name", size="AREA") +
	#tm_compass(position = c(.65, .15), color.light = "grey90") +
	#tm_credits("Eckert IV projection", position = c("right", "BOTTOM")) +
	tm_style("classic") +
	tm_layout(bg.color="lightblue",
			  inner.margins=c(0, 0, .02, 0))
}
