% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tm_layout.R
\name{tm_layout}
\alias{tm_colors_albatross}
\alias{tm_colors_beaver}
\alias{tm_colors_cobalt}
\alias{tm_layout}
\alias{tm_layout_Europe}
\alias{tm_layout_Europe_wide}
\alias{tm_layout_NLD}
\alias{tm_layout_NLD_wide}
\alias{tm_layout_World}
\alias{tm_layout_World_wide}
\title{Layout elements of cartographic maps}
\usage{
tm_layout(title = NA, scale = 1, title.size = 1.3, bg.color = "grey90",
  aes.color = c(fill = "grey70", borders = "grey40", bubbles = "blueviolet",
  dots = "black", lines = "red", text = "black", na = "grey60"),
  aes.palette = list(seq = "YlOrBr", div = "RdYlGn", cat = "Set3"),
  attr.color = "black", frame = TRUE, asp = NA,
  outer.margins = rep(0.02, 4), inner.margins = NA, outer.bg.color = NULL,
  earth.boundary = FALSE, earth.boundary.color = attr.color,
  earth.boundary.lwd = 1, earth.datum = "WGS84", space.color = NULL,
  legend.show = TRUE, legend.only = FALSE, legend.position = NULL,
  legend.width = 0.3, legend.height = 0.9, legend.hist.height = 0.3,
  legend.hist.width = legend.width, legend.title.size = 1.1,
  legend.text.size = 0.7, legend.hist.size = 0.7,
  legend.format = list(scientific = FALSE, digits = NA, text.separator = "to",
  text.less.than = "Less than", text.or.more = "or more"),
  legend.frame = FALSE, legend.text.color = attr.color,
  legend.bg.color = NA, legend.bg.alpha = 1, legend.hist.bg.color = NA,
  legend.hist.bg.alpha = 1, title.snap.to.legend = FALSE,
  title.position = c("left", "top"), title.color = attr.color,
  title.bg.color = NA, title.bg.alpha = 1, attr.position = c("right",
  "bottom"), design.mode = FALSE)

tm_colors_cobalt(bg.color = "#002240", aes.color = c(fill = "#0088FF",
  borders = "#002240", bubbles = "#FF9D00", dots = "#FF9D00", lines = "#FFEE80",
  text = "white", na = "grey60"), aes.palette = list(seq = "YlGn", div =
  "RdYlGn", cat = "Set3"), attr.color = "white", ...)

tm_colors_albatross(bg.color = "#00007F", aes.color = c(fill = "#4C4C88",
  borders = "#BFBFFF", bubbles = "#BFBFFF", dots = "#BFBFFF", lines = "#BFBFFF",
  text = "#FFE700", na = "grey60"), aes.palette = list(seq = "YlOrRd", div =
  "RdYlGn", cat = "Set3"), attr.color = "#BFBFFF", ...)

tm_colors_beaver(bg.color = "#FFFFFF", aes.color = c(fill = "#FFE200",
  borders = "#000000", bubbles = "#A30000", dots = "#A30000", lines = "#A30000",
  text = "#000000", na = "#E0E0E0"), aes.palette = list(seq = "YlOrBr", div =
  "RdYlGn", cat = "Dark2"), attr.color = "black", ...)

tm_layout_World(title = NA, inner.margins = c(0, 0.05, 0.025, 0.01),
  legend.position = c("left", "bottom"), scale = 0.8, ...)

tm_layout_World_wide(title = NA, inner.margins = c(0, 0.2, 0.025, 0.01),
  legend.position = c("left", "bottom"), legend.width = 0.4, scale = 0.8,
  ...)

tm_layout_Europe(title = NA, title.position = c("left", "top"),
  legend.position = c("left", "top"), inner.margins = c(0, 0.1, 0, 0), ...)

tm_layout_Europe_wide(title = NA, title.position = c("left", "top"),
  legend.position = c("left", "top"), inner.margins = c(0, 0.25, 0, 0),
  legend.width = 0.4, legend.hist.width = 0.4, ...)

tm_layout_NLD(title = NA, frame = FALSE, inner.margins = c(0.02, 0.2,
  0.06, 0.02), legend.position = c("left", "top"), legend.width = 0.4, ...)

tm_layout_NLD_wide(title = NA, frame = FALSE, inner.margins = c(0.02, 0.3,
  0.06, 0.02), legend.position = c("left", "top"), legend.width = 0.5,
  legend.hist.width = 0.35, ...)
}
\arguments{
\item{title}{Title(s). By default, the name of the statistical variable of which the legend is drawn at the top (see \code{legend.config}) is used as a title.}

\item{scale}{numeric value that serves as the global scale parameter. All font sizes, bubble sizes, border widths, and line widths are controled by this value. Each of these elements can be scaled independantly with the \code{scale}, \code{lwd}, or \code{size} arguments provided by the \code{\link{tmap-element}s}.}

\item{title.size}{Relative size of the title}

\item{bg.color}{Background color. By default it is \code{"white"}. A recommended alternative for choropleths is light grey (e.g., \code{"grey85"}).}

\item{aes.color}{Default color values for the aesthetics layers. Should be a named vector with the names chosen from: \code{fill}, \code{borders}, \code{bubbles}, \code{dots}, \code{lines}, \code{text}, \code{na}.}

\item{aes.palette}{Default color palettes for the aesthetics. It takes a list of three items: \code{seq} for sequential palettes, \code{div} for diverging palettes, and \code{cat} for categorical palettes. By default, Color Brewer palettes (see (see \code{RColorBrewer::display.brewer.all})) are used. It is also possible provide a vector of colors for any of these items.}

\item{attr.color}{Default color value for map attributes}

\item{frame}{Either a boolean that determines whether a frame is drawn, or a color value that specifies the color of the frame.}

\item{asp}{Aspect ratio. The aspect ratio of the map (width/height). If \code{NA}, it is determined by the bounding box (see argument \code{bbox} of \code{\link{tm_shape}}), the \code{outer.margins}, and the \code{inner.margins}. If \code{0}, then the aspect ratio is adjusted to the aspect ratio of the device.}

\item{outer.margins}{Relative margins between device and frame. Vector of four values specifying the bottom, left, top, and right margin. Values are between 0 and 1.}

\item{inner.margins}{Relative margins inside the frame. Vector of four values specifying the bottom, left, top, and right margin. Values are between 0 and 1. By default, 0 for each side if master shape is a raster, otherwise 0.02.}

\item{outer.bg.color}{Background color outside the frame.}

\item{earth.boundary}{Should the earth boundary be shown? Only applicable for projected world maps.}

\item{earth.boundary.color}{Color of the earth boundary.}

\item{earth.boundary.lwd}{Line width of the earth boundary.}

\item{earth.datum}{Geodetic datum to determine the earth boundary. By default \code{"WGS84"}, other frequently used datums are \code{"NAD83"} and \code{"NAD27"}. Any other \code{PROJ.4} character string can be used.}

\item{space.color}{Color of the space, i.e. the region inside the frame, and outsise the earth boundary.}

\item{legend.show}{Logical that determines whether the legend is shown.}

\item{legend.only}{logical. Only draw the legend (without map)? Particularly useful for small multiples with a common legend.}

\item{legend.position}{Position of the legend. Vector of two values, specifing the x and y coordinates. Either this vector contains "left", "center" or "right" for the first value and "top", "center", or "bottom" for the second value, or this vector contains two numeric values between 0 and 1 that specifies the x and y value of the left bottom corner of the legend. By default, it is automatically placed in the corner with most space based on the (first) shape object.}

\item{legend.width}{maximum width of the legend}

\item{legend.height}{maximum height of the legend.}

\item{legend.hist.height}{height of the histogram. This hight is initial. If the total legend is downscaled to \code{legend.height}, the histogram is downscaled as well.}

\item{legend.hist.width}{width of the histogram. By default, it is equal to the \code{legend.width}.}

\item{legend.title.size}{Relative font size for the legend title}

\item{legend.text.size}{Relative font size for the legend text elements}

\item{legend.hist.size}{Relative font size for the choropleth histogram}

\item{legend.format}{list of formatting options for the legend numbers. Only applicable if \code{labels} is undefined. Parameters are:
\describe{
\item{scientific}{Should the labels be formatted scientically? If so, square brackets are used, and the \code{format} of the numbers is \code{"g"}. Otherwise, \code{format="f"}, and \code{text.separator}, \code{text.less.than}, and \code{text.or.more} are used. Also, the numbers are automatically  rounded to millions or billions if applicable.}
\item{format}{By default, \code{"f"}, i.e. the standard notation \code{xxx.xxx}, is used. If \code{scientific=TRUE} then \code{"g"}, which means that numbers are formatted scientically, i.e. \code{n.dddE+nn} if needed to save space.}
\item{digits}{Number of digits after the decimal point if \code{format="f"}, and the number of significant digits otherwise.}
\item{text.separator}{Character string to use to separate numbers in the legend (default: "to").}
\item{text.less.than}{Character string to use to translate "Less than" (which is the default).}
\item{text.or.more}{Character string to use to translate "or more" (which is the default). }
\item{...}{Other arguments passed on to \code{\link[base:formatC]{formatC}}}
}}

\item{legend.frame}{either a logical that determines whether the legend is placed inside a frame, or a color that directly specifies the frame border color. The width of the frame is automatically determined, but is upper-bounded by \code{legend.width}.}

\item{legend.text.color}{color of the legend text}

\item{legend.bg.color}{Background color of the legend. Use \code{TRUE} to match with the overall background color \code{bg.color}.}

\item{legend.bg.alpha}{Transparency number between 0 (totally transparent) and 1 (not transparent). By default, the alpha value of the \code{legend.bg.color} is used (normally 1).}

\item{legend.hist.bg.color}{Background color of the histogram}

\item{legend.hist.bg.alpha}{Transparency number between 0 (totally transparent) and 1 (not transparent). By default, the alpha value of the \code{legend.hist.bg.color} is used (normally 1).}

\item{title.snap.to.legend}{Logical that determines whether the title is part of the legend.}

\item{title.position}{Position of the title. Vector of two values, specifing the x and y coordinates. Either this vector contains "left", "center" or "right" for the first value and "top", "center", or "bottom" for the second value, or this vector contains two numeric values between 0 and 1 that specifies the x and y value of the left bottom corner of the legend. By default the title is placed on top of the legend (determined by \code{legend.position})}

\item{title.color}{color of the title}

\item{title.bg.color}{background color of the title. Use \code{TRUE} to match with the overall background color \code{bg.color}.}

\item{title.bg.alpha}{Transparency number between 0 (totally transparent) and 1 (not transparent). By default, the alpha value of the \code{title.bg.color} is used (normally 1).}

\item{attr.position}{Position of the map attributes, which are \code{\link{tm_credits}}, \code{\link{tm_scale_bar}} and \code{\link{tm_compass}}. Vector of two values, specifing the x and y coordinates. The first value is "left", "center" or "right", and the second value "top", "center", or "bottom". Positions can also be set separately in the map attribute fuctions.}

\item{design.mode}{Logical that enables the design mode. If \code{TRUE}, inner and outer margins, legend position, aspect ratio are explicitely shown. Also, feedback text in the console is given.}

\item{...}{other arguments from \code{tm_layout}}
}
\description{
This element specifies the map layout. The main function \code{tm_layout} can be seen as a general layout theme. The functions \code{tm_layout_World}, \code{tm_layout_Europe}, and \code{tm_layout_NLD} are layout themes specified for the world, Europe, and Netherlands maps (which are contained in this package). For each of these layout themes, there is also an extra wide variant, with more space for the legend. Tip: create a format layout theme for your own map (see example below). The functions starting with \code{tm_color} are predefined color themes for fixed colors. The map can be further styled with \code{\link{tm_style}} (see details).
}
\details{
The difference between \code{tm_layout} and \code{\link{tm_style}} is the following. Specifications regarding position and size, for instance margins and legend size, are controleld wtih \code{tm_layout}. These specifications are typically dependent on the shapes, and, to a lesser extent, on the type of thematic map. Therefore, map dependent wrappers such as \code{tm_layout_World} can be useful. Furthermore, the default fixed colors (for background, title, attributes, and fixed aestethics) are defined with \code{tm_layout}. On the other hand, \code{\link{tm_style}} controls the styling of the map, that is independent of the used shapes, thematic map type, or used colors. There are two main flavours: modern (default) or classic (\code{\link{tm_style_classic}}).
}
\examples{
data(World)

tm_shape(World) + 
    tm_fill("pop_est_dens", style="kmeans", title="Population density") + 
tm_layout_World("World Population", bg.color="lightblue")

# A custom layout wrapper for Africa
tm_layout_Africa <- function(title=NA,
							 inner.margins = c(.02, .25, .02, .02),
							 draw.frame = FALSE,
							 title.position=c("left", "bottom"), 
							 legend.position = c("left", "bottom"),
							 legend.width = 0.5,
							 bg.color = "lightskyblue2", ...) {
	args <- c(as.list(environment()), list(...))
	do.call("tm_layout", args)
}

Africa <- World[World$continent=="Africa", ]

qtm(Africa, fill="pop_est_dens", fill.style="kmeans", fill.title="Population density") + 
    tm_layout_Africa("Africa")


\dontrun{
# classic styled world map
data(land)
	
# convert to Eckert IV projection
land_eck4 <- set_projection(land, "eck4")

tm_shape(land_eck4) +
	tm_raster("elevation", breaks=c(-Inf, 250, 500, 1000, 1500, 2000, 2500, 3000, 4000, Inf),  
		palette = terrain.colors(9), title="Elevation") +
	tm_shape(World) +
	tm_borders("grey20") +
	tm_grid(projection="longlat", labels.size = .5) +
	tm_text("name", size="AREA") +
	tm_compass(position = c(.65, .15), color.light = "grey90") +
	tm_credits("Eckert IV projection", position = c(.85, 0)) +
	tm_layout(inner.margins=c(.04,.03, .02, .01), legend.position = c("left", "bottom"), 
		legend.frame = TRUE, bg.color="lightblue", legend.bg.color="lightblue", 
		earth.boundary = TRUE, space.color="grey90") + 
	tm_style_classic()
}

# Color theme comparison
two_by_two_plot <- function(tm) {
	require(grid)
	grid.newpage()
	pushViewport(viewport(layout = grid.layout(2,2)))
	print(tm + tm_layout("Default theme"),vp = viewport(layout.pos.row = 1, layout.pos.col = 1))
	print(tm + tm_colors_cobalt(title="Cobalt theme"), vp = viewport(layout.pos.row = 1, layout.pos.col = 2))
	print(tm + tm_colors_albatross(title="Albatross theme"), vp = viewport(layout.pos.row = 2, layout.pos.col = 1))
	print(tm + tm_colors_beaver(title="Beaver theme"), vp = viewport(layout.pos.row = 2, layout.pos.col = 2))
	upViewport()
}

# fixed aesthetics
data(Europe, rivers, metro)
two_by_two_plot(tm_shape(Europe) +
	tm_fill() +
	tm_shape(rivers) +
	tm_lines() +
	tm_shape(metro) +
	tm_bubbles(size="pop2010") + 
	tm_text("name", size="pop2010", auto.placement = TRUE))


# for choropleth aesthetics
data(NLD_muni, NLD_prov)
two_by_two_plot(tm_shape(NLD_muni) +
	tm_fill("population", convert2density = TRUE, style="kmeans") +
tm_shape(NLD_prov) +
	tm_borders(lwd=2) +
	tm_text("name") +
tm_layout(scale=.7))

}
\seealso{
\code{\link{tm_style}}, \href{../doc/tmap-nutshell.html}{\code{vignette("tmap-nutshell")}}
}

