% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layout.R, R/tmap_options.R
\name{tm_layout}
\alias{tm_layout}
\alias{tm_style}
\alias{tm_format}
\title{Layout options}
\usage{
tm_layout(
  scale,
  asp,
  bg.color,
  outer.bg.color,
  frame,
  frame.lwd,
  frame.r,
  frame.double.line,
  outer.margins,
  inner.margins,
  inner.margins.extra,
  meta.margins,
  meta.auto.margins,
  between.margin,
  component.offset,
  component.stack.margin,
  grid.mark.height,
  xylab.height,
  coords.height,
  xlab.show,
  ylab.show,
  xlab.pos,
  ylab.pos,
  grid.show,
  grid.label.pos,
  panel.type,
  panel.wrap.pos,
  panel.xtab.pos,
  unit,
  color.sepia.intensity,
  color.saturation,
  color.vision.deficiency.sim,
  fontface,
  fontfamily,
  legend.show,
  legend.design,
  legend.orientation,
  legend.position,
  legend.width,
  legend.height,
  legend.stack,
  legend.group.frame,
  legend.resize.as.group,
  legend.reverse,
  legend.title.color,
  legend.title.size,
  legend.title.fontface,
  legend.title.fontfamily,
  legend.title.just,
  legend.text.color,
  legend.text.size,
  legend.text.fontface,
  legend.text.fontfamily,
  legend.frame,
  legend.frame.lwd,
  legend.frame.r,
  legend.bg.color,
  legend.bg.alpha,
  compass.type,
  title.size,
  title.color,
  title.fontface,
  title.fontfamily,
  title.bg.color,
  title.bg.alpha,
  title.padding,
  title.frame,
  title.frame.lwd,
  title.frame.r,
  title.stack,
  title.position,
  title.width,
  title.heigth,
  title.group.frame,
  title.resize.as.group,
  panel.show,
  panel.labels,
  panel.label.size,
  panel.label.color,
  panel.label.fontface,
  panel.label.fontfamily,
  panel.label.bg.color,
  panel.label.height,
  panel.label.rot
)

tm_style(style, ...)

tm_format(format, ...)
}
\arguments{
\item{scale}{PARAM_DESCRIPTION}

\item{asp}{PARAM_DESCRIPTION}

\item{bg.color}{PARAM_DESCRIPTION}

\item{outer.bg.color}{PARAM_DESCRIPTION}

\item{frame}{PARAM_DESCRIPTION}

\item{frame.lwd}{PARAM_DESCRIPTION}

\item{frame.r}{PARAM_DESCRIPTION}

\item{frame.double.line}{PARAM_DESCRIPTION}

\item{outer.margins}{PARAM_DESCRIPTION}

\item{inner.margins}{PARAM_DESCRIPTION}

\item{inner.margins.extra}{PARAM_DESCRIPTION}

\item{meta.margins}{PARAM_DESCRIPTION}

\item{meta.auto.margins}{PARAM_DESCRIPTION}

\item{between.margin}{PARAM_DESCRIPTION}

\item{component.offset}{PARAM_DESCRIPTION}

\item{component.stack.margin}{PARAM_DESCRIPTION}

\item{grid.mark.height}{PARAM_DESCRIPTION}

\item{xylab.height}{PARAM_DESCRIPTION}

\item{coords.height}{PARAM_DESCRIPTION}

\item{xlab.show}{PARAM_DESCRIPTION}

\item{ylab.show}{PARAM_DESCRIPTION}

\item{xlab.pos}{PARAM_DESCRIPTION}

\item{ylab.pos}{PARAM_DESCRIPTION}

\item{grid.show}{PARAM_DESCRIPTION}

\item{grid.label.pos}{PARAM_DESCRIPTION}

\item{panel.type}{PARAM_DESCRIPTION}

\item{panel.wrap.pos}{PARAM_DESCRIPTION}

\item{panel.xtab.pos}{PARAM_DESCRIPTION}

\item{unit}{PARAM_DESCRIPTION}

\item{color.sepia.intensity}{PARAM_DESCRIPTION}

\item{color.saturation}{PARAM_DESCRIPTION}

\item{color.vision.deficiency.sim}{PARAM_DESCRIPTION}

\item{fontface}{PARAM_DESCRIPTION}

\item{fontfamily}{PARAM_DESCRIPTION}

\item{legend.show}{PARAM_DESCRIPTION}

\item{legend.design}{PARAM_DESCRIPTION}

\item{legend.orientation}{PARAM_DESCRIPTION}

\item{legend.position}{PARAM_DESCRIPTION}

\item{legend.width}{PARAM_DESCRIPTION}

\item{legend.height}{PARAM_DESCRIPTION}

\item{legend.stack}{PARAM_DESCRIPTION}

\item{legend.group.frame}{PARAM_DESCRIPTION}

\item{legend.resize.as.group}{PARAM_DESCRIPTION}

\item{legend.reverse}{PARAM_DESCRIPTION}

\item{legend.title.color}{PARAM_DESCRIPTION}

\item{legend.title.size}{PARAM_DESCRIPTION}

\item{legend.title.fontface}{PARAM_DESCRIPTION}

\item{legend.title.fontfamily}{PARAM_DESCRIPTION}

\item{legend.title.just}{PARAM_DESCRIPTION}

\item{legend.text.color}{PARAM_DESCRIPTION}

\item{legend.text.size}{PARAM_DESCRIPTION}

\item{legend.text.fontface}{PARAM_DESCRIPTION}

\item{legend.text.fontfamily}{PARAM_DESCRIPTION}

\item{legend.frame}{PARAM_DESCRIPTION}

\item{legend.frame.lwd}{PARAM_DESCRIPTION}

\item{legend.frame.r}{PARAM_DESCRIPTION}

\item{legend.bg.color}{PARAM_DESCRIPTION}

\item{legend.bg.alpha}{PARAM_DESCRIPTION}

\item{compass.type}{PARAM_DESCRIPTION}

\item{title.size}{PARAM_DESCRIPTION}

\item{title.color}{PARAM_DESCRIPTION}

\item{title.fontface}{PARAM_DESCRIPTION}

\item{title.fontfamily}{PARAM_DESCRIPTION}

\item{title.bg.color}{PARAM_DESCRIPTION}

\item{title.bg.alpha}{PARAM_DESCRIPTION}

\item{title.padding}{PARAM_DESCRIPTION}

\item{title.frame}{PARAM_DESCRIPTION}

\item{title.frame.lwd}{PARAM_DESCRIPTION}

\item{title.frame.r}{PARAM_DESCRIPTION}

\item{title.stack}{PARAM_DESCRIPTION}

\item{title.position}{PARAM_DESCRIPTION}

\item{title.width}{PARAM_DESCRIPTION}

\item{title.heigth}{PARAM_DESCRIPTION}

\item{title.group.frame}{PARAM_DESCRIPTION}

\item{title.resize.as.group}{PARAM_DESCRIPTION}

\item{panel.show}{PARAM_DESCRIPTION}

\item{panel.labels}{PARAM_DESCRIPTION}

\item{panel.label.size}{PARAM_DESCRIPTION}

\item{panel.label.color}{PARAM_DESCRIPTION}

\item{panel.label.fontface}{PARAM_DESCRIPTION}

\item{panel.label.fontfamily}{PARAM_DESCRIPTION}

\item{panel.label.bg.color}{PARAM_DESCRIPTION}

\item{panel.label.height}{PARAM_DESCRIPTION}

\item{panel.label.rot}{PARAM_DESCRIPTION}

\item{style}{name of the style}

\item{format}{name of the format}
}
\description{
Set of tmap options that are directly related to the layout.
}
