% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_legend.R
\name{tm_legend}
\alias{tm_legend}
\alias{tm_legend_hide}
\alias{tm_legend_combine}
\title{Legend}
\usage{
tm_legend(
  title,
  show,
  orientation,
  design,
  reverse,
  na.show,
  position,
  width,
  height,
  stack,
  z,
  group.frame,
  resize_as_group,
  title.color,
  title.size,
  title.fontface,
  title.fontfamily,
  title.padding,
  text.color,
  text.size,
  text.fontface,
  text.fontfamily,
  format,
  frame,
  frame.lwd,
  frame.r,
  bg.color,
  bg.alpha,
  item.height,
  item.width,
  item.space,
  item.na.height,
  item.na.width,
  item.na.space,
  item.shape,
  ticks,
  ticks.disable.na,
  ticks.col,
  ticks.lwd,
  title.align,
  margins,
  margin.item.text,
  ...
)

tm_legend_hide()

tm_legend_combine(variable)
}
\arguments{
\item{title}{Legend title}

\item{show}{Show legend?}

\item{orientation}{Orientation of the legend: \code{"portrait"} or \code{"landscape"}}

\item{design}{PARAM_DESCRIPTION}

\item{reverse}{Should the legend be reversed?}

\item{na.show}{PARAM_DESCRIPTION}

\item{position}{PARAM_DESCRIPTION}

\item{width}{Width of the legend}

\item{height}{Height of the legend}

\item{stack}{PARAM_DESCRIPTION}

\item{z}{PARAM_DESCRIPTION}

\item{group.frame}{PARAM_DESCRIPTION}

\item{resize_as_group}{PARAM_DESCRIPTION}

\item{title.color}{Color of the legend title}

\item{title.size}{Size of the legend title}

\item{title.fontface}{Font face of the legend title}

\item{title.fontfamily}{Font family of the legend title}

\item{title.padding}{PARAM_DESCRIPTION}

\item{text.color}{Color of the legend text}

\item{text.size}{Size of the legend text}

\item{text.fontface}{Font face of the legend text}

\item{text.fontfamily}{Font family of the legend text}

\item{format}{PARAM_DESCRIPTION}

\item{frame}{PARAM_DESCRIPTION}

\item{frame.lwd}{PARAM_DESCRIPTION}

\item{frame.r}{PARAM_DESCRIPTION}

\item{bg.color}{Background color of the legend}

\item{bg.alpha}{Background transparency of the legend}

\item{item.height}{PARAM_DESCRIPTION}

\item{item.width}{PARAM_DESCRIPTION}

\item{item.space}{PARAM_DESCRIPTION}

\item{item.na.height}{PARAM_DESCRIPTION}

\item{item.na.width}{PARAM_DESCRIPTION}

\item{item.na.space}{PARAM_DESCRIPTION}

\item{item.shape}{PARAM_DESCRIPTION}

\item{ticks}{PARAM_DESCRIPTION}

\item{ticks.disable.na}{PARAM_DESCRIPTION}

\item{ticks.col}{PARAM_DESCRIPTION}

\item{ticks.lwd}{PARAM_DESCRIPTION}

\item{title.align}{PARAM_DESCRIPTION}

\item{margins}{PARAM_DESCRIPTION}

\item{margin.item.text}{PARAM_DESCRIPTION}

\item{...}{passed on (?)}

\item{variable}{visual (or transformation) variable to combine the legend with: e.g. \code{"fill"} or \code{"size"}}
}
\value{
A tm_legend component
}
\description{
Legend specification
}
