% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_legend.R
\name{tm_legend}
\alias{tm_legend}
\alias{tm_legend_hide}
\alias{tm_legend_combine}
\title{Legend}
\usage{
tm_legend(
  title,
  show,
  orientation,
  design,
  reverse,
  na.show,
  position,
  width,
  height,
  stack,
  z,
  group.frame,
  resize_as_group,
  title.color,
  title.size,
  title.fontface,
  title.fontfamily,
  title.padding,
  title.align,
  text.color,
  text.size,
  text.fontface,
  text.fontfamily,
  format,
  frame,
  frame.lwd,
  frame.r,
  bg.color,
  bg.alpha,
  item.height,
  item.width,
  item.space,
  item.na.height,
  item.na.width,
  item.na.space,
  item.shape,
  ticks,
  ticks.disable.na,
  ticks.col,
  ticks.lwd,
  margins,
  margin.item.text,
  ...
)

tm_legend_hide()

tm_legend_combine(variable)
}
\arguments{
\item{title}{Legend title}

\item{show}{Show legend?}

\item{orientation}{Orientation of the legend: \code{"portrait"} or \code{"landscape"}}

\item{design}{Legend design \code{"standard"}.}

\item{reverse}{Should the legend be reversed?}

\item{na.show}{Show \code{NA} values in legend?}

\item{position}{Vector of two values, specifying the x and y coordinates. The first is either \code{"left"} or \code{"right"}, the second either \code{"top"} or \code{"bottom"}.}

\item{width}{Width of the legend}

\item{height}{Height of the legend}

\item{stack}{stack}

\item{z}{z}

\item{group.frame}{group.frame}

\item{resize_as_group}{PARAM_DESCRIPTION}

\item{title.color}{Color of the legend title}

\item{title.size}{Size of the legend title}

\item{title.fontface}{Font face of the legend title}

\item{title.fontfamily}{Font family of the legend title}

\item{title.padding}{Padding of the legend title}

\item{title.align}{Title alignment}

\item{text.color}{Color of the legend text}

\item{text.size}{Size of the legend text}

\item{text.fontface}{Font face of the legend text}

\item{text.fontfamily}{Font family of the legend text}

\item{format}{Use the format argument of the \verb{tm_scale_*()} functions instead.}

\item{frame}{frame}

\item{frame.lwd}{frame line width}

\item{frame.r}{Radius of the rounded frame corners. 0 means no rounding.}

\item{bg.color}{Background color of the legend}

\item{bg.alpha}{Background transparency of the legend}

\item{item.height}{PARAM_DESCRIPTION}

\item{item.width}{PARAM_DESCRIPTION}

\item{item.space}{PARAM_DESCRIPTION}

\item{item.na.height}{PARAM_DESCRIPTION}

\item{item.na.width}{PARAM_DESCRIPTION}

\item{item.na.space}{PARAM_DESCRIPTION}

\item{item.shape}{PARAM_DESCRIPTION}

\item{ticks}{List of vectors of size 2 that determines the horizontal tick mark lines (for portrait legends). The values are the y-values of begin and endpoint of each tick mark.}

\item{ticks.disable.na}{PARAM_DESCRIPTION}

\item{ticks.col}{PARAM_DESCRIPTION}

\item{ticks.lwd}{PARAM_DESCRIPTION}

\item{margins}{PARAM_DESCRIPTION}

\item{margin.item.text}{PARAM_DESCRIPTION}

\item{...}{visual values, e.g. \code{col}, \code{fill}, \code{lwd}, can be specified. If so, they overrule the default visual values, which are determined by the drawn map objects (e.g. polygons)}

\item{variable}{visual (or transformation) variable to combine the legend with: e.g. \code{"fill"} or \code{"size"}}
}
\value{
A tm_legend component
}
\description{
Legend specification
}
