% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_legend.R
\name{tm_legend}
\alias{tm_legend}
\alias{tm_legend_hide}
\alias{tm_legend_combine}
\title{Legend}
\usage{
tm_legend(
  title,
  show,
  orientation,
  design,
  reverse,
  na.show,
  position,
  width,
  height,
  stack,
  z,
  group.frame,
  resize.as.group,
  title.color,
  title.size,
  title.fontface,
  title.fontfamily,
  title.padding,
  text.color,
  text.size,
  text.fontface,
  text.fontfamily,
  format,
  frame,
  frame.lwd,
  frame.r,
  bg.color,
  bg.alpha,
  item.height,
  item.width,
  item.space,
  item.na.height,
  item.na.width,
  item.na.space,
  ticks,
  ticks.disable.na,
  ticks.col,
  ticks.lwd,
  title.just,
  margins,
  margin.item.text
)

tm_legend_hide()

tm_legend_combine(aes)
}
\arguments{
\item{title}{PARAM_DESCRIPTION}

\item{show}{PARAM_DESCRIPTION}

\item{orientation}{PARAM_DESCRIPTION}

\item{design}{PARAM_DESCRIPTION}

\item{reverse}{PARAM_DESCRIPTION}

\item{position}{PARAM_DESCRIPTION}

\item{width}{PARAM_DESCRIPTION}

\item{height}{PARAM_DESCRIPTION}

\item{stack}{PARAM_DESCRIPTION}

\item{z}{PARAM_DESCRIPTION}

\item{group.frame}{PARAM_DESCRIPTION}

\item{resize.as.group}{PARAM_DESCRIPTION}

\item{title.color}{PARAM_DESCRIPTION}

\item{title.size}{PARAM_DESCRIPTION}

\item{title.fontface}{PARAM_DESCRIPTION}

\item{title.fontfamily}{PARAM_DESCRIPTION}

\item{title.padding}{PARAM_DESCRIPTION}

\item{text.color}{PARAM_DESCRIPTION}

\item{text.size}{PARAM_DESCRIPTION}

\item{text.fontface}{PARAM_DESCRIPTION}

\item{text.fontfamily}{PARAM_DESCRIPTION}

\item{format}{PARAM_DESCRIPTION}

\item{frame}{PARAM_DESCRIPTION}

\item{frame.lwd}{PARAM_DESCRIPTION}

\item{frame.r}{PARAM_DESCRIPTION}

\item{bg.color}{PARAM_DESCRIPTION}

\item{bg.alpha}{PARAM_DESCRIPTION}

\item{item.height}{PARAM_DESCRIPTION}

\item{item.width}{PARAM_DESCRIPTION}

\item{item.space}{PARAM_DESCRIPTION}

\item{item.na.height}{PARAM_DESCRIPTION}

\item{item.na.width}{PARAM_DESCRIPTION}

\item{item.na.space}{PARAM_DESCRIPTION}

\item{ticks}{PARAM_DESCRIPTION}

\item{ticks.disable.na}{PARAM_DESCRIPTION}

\item{ticks.col}{PARAM_DESCRIPTION}

\item{ticks.lwd}{PARAM_DESCRIPTION}

\item{title.just}{PARAM_DESCRIPTION}

\item{margins}{PARAM_DESCRIPTION}

\item{margin.item.text}{PARAM_DESCRIPTION}

\item{aes}{aes}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Legend specification
}
