% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_logo}
\alias{tm_logo}
\title{Map component: logo}
\usage{
tm_logo(
  file,
  height,
  margins,
  stack,
  position,
  frame,
  frame.lwd,
  frame.r,
  group.frame,
  resize.as.group,
  z
)
}
\arguments{
\item{file}{either a filename or url of a png image. If multiple files/urls are provided with a character vector, the logos are placed near each other. To specify logos for small multiples use a list of character values/vectors. In order to stack logos vertically, multiple \code{tm_logo} elements can be stacked.}

\item{height}{height of the logo in number of text line heights. The width is scaled based the height and the aspect ratio of the logo. If multiple logos are specified by a vector or list, the heights can be specified accordingly.}

\item{margins}{margins}

\item{position}{position}

\item{z}{z}
}
\description{
Map component that adds a scale bar. As of version 4.0, \code{tm_scalebar()} is
used instead of \code{tm_scale_bar()} (now deprecated), because of the potential
confusion with the \verb{tm_scale_*()} scaling functions (like \code{\link[=tm_scale_continuous]{tm_scale_continuous()}}).
}
