% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_misc_elements.R
\name{tm_minimap}
\alias{tm_minimap}
\title{Minimap}
\usage{
tm_minimap(position = NA, ...)
}
\arguments{
\item{position}{position of the scale bar Vector of two values, specifying the x and y coordinates. The first is either "left" or "right", the second either "top" or "bottom". The default value is controlled by the argument \code{"attr.position"} of \code{\link{tm_layout}}.}

\item{...}{arguments passed on to \code{\link[leaflet:addMiniMap]{addMiniMap}}.}
}
\description{
Creates a minimap in view mode. See \code{\link[leaflet:addMiniMap]{addMiniMap}}.
}
\seealso{
\code{\link[leaflet:addMiniMap]{addMiniMap}}
}
