% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_minimap}
\alias{tm_minimap}
\title{Map component: minimap}
\usage{
tm_minimap(server, toggle, stack, position, z, ...)
}
\arguments{
\item{server}{name of the provider or an URL (see \code{\link{tm_tiles}}). By default, it shows the same map as the basemap, and moreover, it will automatically change when the user switches basemaps. Note the latter does not happen when \code{server} is specified.}

\item{toggle}{should the minimap have a button to minimise it? By default \code{TRUE}.}

\item{stack}{stack}

\item{position}{position}

\item{z}{z}

\item{...}{arguments passed on to \code{\link[leaflet:addMiniMap]{addMiniMap}}.}
}
\description{
Map component that adds a minimap in view mode
}
\seealso{
\code{\link[leaflet:addMiniMap]{addMiniMap}}
}
