% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_misc_elements.R
\name{tm_minimap}
\alias{tm_minimap}
\title{Minimap}
\usage{
tm_minimap(server = NA, position = c("left", "bottom"),
  toggle = TRUE, ...)
}
\arguments{
\item{server}{name of the provider or an URL (see \code{\link{tm_tiles}}). By default, it shows the same map as the basemap, and moreover, it will automatically change when the user switches basemaps. Note the latter does not happen when \code{server} is specified.}

\item{position}{position of the scale bar Vector of two values, specifying the x and y coordinates. The first is either "left" or "right", the second either "top" or "bottom".}

\item{toggle}{should the minimap have a button to minimise it? By default \code{TRUE}.}

\item{...}{arguments passed on to \code{\link[leaflet:addMiniMap]{addMiniMap}}.}
}
\description{
Creates a minimap in view mode. See \code{\link[leaflet:addMiniMap]{addMiniMap}}.
}
\seealso{
\code{\link[leaflet:addMiniMap]{addMiniMap}}
}
