% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_mouse_coordinates}
\alias{tm_mouse_coordinates}
\title{Map component: mouse coordinates}
\usage{
tm_mouse_coordinates(stack, position, z)
}
\arguments{
\item{stack}{stack with other map components, either \code{"vertical"} or \code{"horizontal"}.}

\item{position}{An object created with \code{tm_pos_in()} or \code{tm_pos_out()}. Or, as a shortcut, a vector of two values, specifying the x and y coordinates. The first is \code{"left"}, \code{"center"} or \code{"right"} (or upper case, meaning tighter to the map frame), the second \code{"top"}, \code{"center"} or \code{"bottom"}. Numeric values are also supported, where 0, 0 means left bottom and 1, 1 right top. See also \href{https://r-tmap.github.io/tmap/articles/adv_positions}{vignette about positioning}.}

\item{z}{z}
}
\description{
Map component that adds mouse coordinates
}
\seealso{
\href{https://r-tmap.github.io/tmap/articles/basics_components}{Vignette about components}
}
