% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_misc_elements.R
\name{tm_mouse_coordinates}
\alias{tm_mouse_coordinates}
\title{Mouse coordinates}
\usage{
tm_mouse_coordinates()
}
\description{
Adds mouse coordinates in view mode. See \code{\link[leafem:addMouseCoordinates]{addMouseCoordinates}}.
}
\seealso{
\code{\link[leafem:addMouseCoordinates]{addMouseCoordinates}}
}
