% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_rgb.R
\name{tm_rgb}
\alias{tm_rgb}
\title{Map layer: RGB}
\usage{
tm_rgb(
  col = tm_mv(1:3),
  col.scale = tm_scale_rgb(),
  col.legend = tm_legend(),
  col.chart = tm_chart_none(),
  col.free = NA,
  options = opt_tm_rgb()
)
}
\arguments{
\item{col, col.scale, col.legend, col.chart, col.free}{Visual variable that determines
the col color.}

\item{options}{options passed on to the corresponding \verb{opt_<layer_function>} function}
}
\description{
Map layer: RGB
}
