% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers.R
\name{tm_rgb}
\alias{tm_rgb}
\title{Map layer: RGB}
\usage{
tm_rgb(
  col = tm_mv(1:3),
  col.scale = tm_scale_rgb(),
  col.legend = tm_legend(),
  col.free = NA
)
}
\arguments{
\item{col, col.scale, col.legend, col.free}{Visual variable that determines the col color.}
}
\description{
Map layer: RGB
}
