% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_rgb.R
\name{opt_tm_rgb}
\alias{opt_tm_rgb}
\alias{tm_rgb}
\alias{tm_rgba}
\title{Map layer: rgb images}
\usage{
opt_tm_rgb(interpolate = FALSE)

tm_rgb(
  col = tm_mv(1:3),
  col.scale = tm_scale_rgb(),
  col.legend = tm_legend(),
  col.chart = tm_chart_none(),
  col.free = NA,
  options = opt_tm_rgb()
)

tm_rgba(
  col = tm_mv(1:4),
  col.scale = tm_scale_rgba(),
  col.legend = tm_legend(),
  col.chart = tm_chart_none(),
  col.free = NA,
  options = opt_tm_rgb()
)
}
\arguments{
\item{interpolate}{Should the raster image be interpolated? Currently only applicable in view mode (passed on to \code{\link[grid:grid.raster]{grid}})}

\item{col, col.scale, col.legend, col.chart, col.free}{Visual variable that determines
the col color. \code{col} is a multivariate variable, with 3 (\code{tm_rgb}) or 4 (\code{tm_rgba}) numeric data variables. These can be specified via \code{\link[=tm_mv]{tm_mv()}} or \code{\link[=tm_mv_dim]{tm_mv_dim()}}}

\item{options}{options passed on to the corresponding \verb{opt_<layer_function>} function}
}
\description{
Map layer that an rgb image.. The used (multivariate) visual variable is \code{col}, which should be specified with 3 or 4 variables for \code{tm_rgb} and \code{tm_rgba} respectively. The first three correspond to the red, green, and blue channels. The optional fourth is the alpha transparency channel.
}
