% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_scale_bar}
\alias{tm_scale_bar}
\title{Map component: scale bar}
\usage{
tm_scale_bar(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=tm_scalebar]{tm_scalebar}}
  \describe{
    \item{\code{breaks}}{breaks}
    \item{\code{width}}{width}
    \item{\code{text.size}}{text.size}
    \item{\code{text.color}}{text.color}
    \item{\code{color.dark}}{color.dark}
    \item{\code{color.light}}{color.light}
    \item{\code{lwd}}{lwd}
    \item{\code{position}}{position}
    \item{\code{bg.color}}{bg.color}
    \item{\code{bg.alpha}}{bg.alpha}
    \item{\code{size}}{size}
    \item{\code{stack}}{stack}
    \item{\code{frame}}{frame}
    \item{\code{frame.lwd}}{frame.lwd}
    \item{\code{frame.r}}{frame.r}
    \item{\code{margins}}{margins}
    \item{\code{z}}{z}
  }}
}
\description{
This function was renamed to [tm_scalebar()] in tmap v4.0
}
