% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_scalebar}
\alias{tm_scalebar}
\title{Map component: scale bar}
\usage{
tm_scalebar(
  breaks,
  width,
  text.size,
  text.color,
  color.dark,
  color.light,
  lwd,
  position,
  bg.color,
  bg.alpha,
  size = "deprecated",
  stack,
  frame,
  frame.lwd,
  frame.r,
  margins,
  z
)
}
\arguments{
\item{breaks}{breaks}

\item{width}{width}

\item{text.size}{text size}

\item{text.color}{text.color}

\item{color.dark}{color.dark}

\item{color.light}{color.light}

\item{lwd}{linewidth}

\item{position}{Vector of two values, specifying the x and y coordinates. The first is either \code{"left"} or \code{"right"}, the second either \code{"top"} or \code{"bottom"}.}

\item{bg.color}{Background color}

\item{bg.alpha}{Background transparency}

\item{size}{Deprecated (use \code{text.size} instead)}

\item{stack}{stack}

\item{frame}{frame}

\item{frame.lwd}{frame line width}

\item{frame.r}{Radius of the rounded frame corners. 0 means no rounding.}

\item{margins}{margins}

\item{z}{z}
}
\description{
Map component that adds a scale bar.
}
