% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_components.R
\name{tm_scalebar}
\alias{tm_scalebar}
\title{Map component: scale bar}
\usage{
tm_scalebar(
  breaks,
  width,
  text.size,
  text.color,
  color.dark,
  color.light,
  lwd,
  position,
  bg.color,
  bg.alpha,
  size = "deprecated",
  stack,
  frame,
  frame.lwd,
  frame.r,
  margins,
  z
)
}
\arguments{
\item{breaks}{breaks}

\item{width}{width}

\item{text.size}{text.size}

\item{text.color}{text.color}

\item{color.dark}{color.dark}

\item{color.light}{color.light}

\item{lwd}{lwd}

\item{position}{position}

\item{bg.color}{bg.color}

\item{bg.alpha}{bg.alpha}

\item{size}{size}

\item{stack}{stack}

\item{frame}{frame}

\item{frame.lwd}{frame.lwd}

\item{frame.r}{frame.r}

\item{margins}{margins}

\item{z}{z}
}
\description{
Map component that adds a scale bar. As of version 4.0, `tm_scalebar` is used instead of `tm_scale_bar` (now deprecated), because of the potential confusion with the `tm_scale_x` scaling functions (like \code{\link{tm_scale_continuous}}).
}
