% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_shape.R
\name{tm_shape}
\alias{tm_shape}
\title{Shape (spatial object) specification}
\usage{
tm_shape(
  shp,
  name = NULL,
  is.main = NA,
  crs = NULL,
  bbox = NULL,
  unit = NULL,
  filter = NULL,
  projection = NULL
)
}
\arguments{
\item{shp}{spatial object}

\item{name}{name of the shape}

\item{is.main}{is \code{shp} the main shape, which determines the crs and bounding box of the map?}

\item{crs}{crs to which \code{shp} is reprojected (only used if \code{is.main = TRUE})}

\item{bbox}{bounding box of he map (only used if \code{is.main = TRUE})}

\item{unit}{unit of the coordinates}

\item{filter}{filter features}
}
\description{
Specify a shape, which is a spatial object from one of these spatial object class packages: \code{sf}, \code{stars}, \code{terra}.
}
