% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_shape.R
\name{tm_shape}
\alias{tm_shape}
\title{Shape (spatial object) specification}
\usage{
tm_shape(
  shp,
  bbox = NULL,
  crs = NULL,
  is.main = NA,
  name = NULL,
  unit = NULL,
  filter = NULL,
  ...
)
}
\arguments{
\item{shp}{Spatial object}

\item{bbox}{Bounding box of he map (only used if \code{is.main = TRUE})}

\item{crs}{CRS to which \code{shp} is reprojected (only used if \code{is.main = TRUE})}

\item{is.main}{Is \code{shp} the main shape, which determines the crs and
bounding box of the map?}

\item{name}{Name of the shape}

\item{unit}{Unit of the coordinates}

\item{filter}{Filter features}

\item{...}{to catch deprecated arguments from version < 4.0}
}
\description{
Specify a shape, which is a spatial object from one of these spatial object
class packages: \code{\link[sf:sf]{sf}}, \code{\link[stars:st_as_stars]{stars}}, or \code{terra}.
}
\examples{
tm_shape(World, crs = "+proj=robin", filter = World$continent=="Africa") + 
	tm_polygons()
}
