% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers_text.R
\name{tm_text}
\alias{tm_text}
\title{Map layer: text}
\usage{
tm_text(
  text = tm_const(),
  text.scale = tm_scale(),
  text.legend = tm_legend(),
  text.free = NA,
  size = tm_const(),
  size.scale = tm_scale(),
  size.legend = tm_legend(),
  size.free = NA,
  col = tm_const(),
  col.scale = tm_scale(),
  col.legend = tm_legend(),
  col.free = NA,
  col_alpha = tm_const(),
  col_alpha.scale = tm_scale(),
  col_alpha.legend = tm_legend(),
  col_alpha.free = NA,
  fontface = tm_const(),
  fontface.scale = tm_scale(),
  fontface.legend = tm_legend(),
  fontface.free = NA,
  fontfamily = "",
  bgcol = tm_const(),
  bgcol.scale = tm_scale(),
  bgcol.legend = tm_legend(),
  bgcol.free = NA,
  bgcol_alpha = tm_const(),
  bgcol_alpha.scale = tm_scale(),
  bgcol_alpha.legend = tm_legend(),
  bgcol_alpha.free = NA,
  shadow = FALSE,
  plot.order = tm_plot_order("AREA", reverse = FALSE, na.order = "bottom"),
  trans.args = list(points.only = "ifany"),
  mapping.args = list(clustering = FALSE),
  zindex = NA,
  group = NA,
  group.control = "check",
  ...
)
}
\arguments{
\item{text, text.scale, text.legend, text.free}{Visual variable that determines
the text. See details.}

\item{size, size.scale, size.legend, size.free}{Visual variable that determines
the font size. See details.}

\item{col, col.scale, col.legend, col.free}{Visual variable that determines
the col color. See details.}

\item{col_alpha, col_alpha.scale, col_alpha.legend, col_alpha.free}{Visual variable that determines
the border color alpha transparency. See Details.}

\item{fontface, fontface.scale, fontface.legend, fontface.free}{Visual variable that determines
the font face. See Details.}

\item{fontfamily}{The font family. See \link[grid:gpar]{gpar()} for details.}

\item{bgcol, bgcol.scale, bgcol.legend, bgcol.free}{Visual variable that determines
the background color. See Details.}

\item{bgcol_alpha, bgcol_alpha.scale, bgcol_alpha.legend, bgcol_alpha.free}{Visual variable that determines
the background color transparency. See Details.}

\item{shadow}{Shadow behind the text. Logical or color.}

\item{plot.order}{Specification in which order the spatial features are drawn.
See \code{\link[=tm_plot_order]{tm_plot_order()}} for details.}

\item{trans.args, mapping.args}{lists that are passed on to internal transformation
and mapping functions respectively.}

\item{zindex}{Map layers are drawn on top of each other. The \code{zindex} numbers
(one for each map layer) determines the stacking order.
By default the map layers are drawn in the order they are called.}

\item{group}{Name of the group to which this layer belongs. This is only
relevant in view mode, where layer groups can be switched (see \code{group.control})}

\item{group.control}{In view mode, the group control determines how layer
groups can be switched on and off. Options: \code{"radio"} for radio buttons
(meaning only one group can be shown), \code{"check"} for check boxes (so multiple groups can be shown),
and \code{"none"} for no control (the group cannot be (de)selected).}

\item{...}{to catch deprecated arguments from version < 4.0}
}
\description{
Map layer that draws symbols Supported visual variables are: \code{text}
(the text itself) \code{col} (color), \code{size} (font size), and \code{fontface} (font face).
}
\details{
The visual variable arguments (e.g. \code{col}) can be specified with either a
data variable name (of the object specified in \code{\link[=tm_shape]{tm_shape()}}), or with a visual
value (for \code{col}, a color is expected). Multiple values can be specified:
in that case facets are created. These facets can be combined with other
faceting data variables, specified with \code{\link[=tm_facets]{tm_facets()}}.

The \code{.scale} arguments determine the used scale to map the data values to
visual variable values. These can be specified with one of the available
\code{tm_scale_()} functions. The default scale that is used is specified by the
tmap option \code{scales.var}.

The \code{.legend} arguments determine the used legend, specified with \code{\link[=tm_legend]{tm_legend()}}.
The default legend and its settings are determined by the tmap options \code{legend.}.

The \code{.free} arguments determine whether scales are applied freely across facets,
or shared. A logical value is required. They can also be specified with a
vector of three logical values; these determine whether scales are applied
freely per facet dimension. This is only useful when facets are applied
(see \code{\link[=tm_facets]{tm_facets()}}). There are maximally three facet dimensions: rows, columns,
and pages. This only applies for a facet grid (\code{\link[=tm_facets_grid]{tm_facets_grid()}}).
For instance, \code{col.free = c(TRUE, FALSE, FALSE)} means that for the visual
variable \code{col}, each row of facets will has its own scale, and therefore its
own legend. For facet wraps and stacks (\code{\link[=tm_facets_wrap]{tm_facets_wrap()}} and \code{\link[=tm_facets_stack]{tm_facets_stack()}})
there is only one facet dimension, so the \code{.free} argument requires only one logical value.
}
\examples{
tm_shape(World) +
  tm_text(text = "name", 
    size = .4, 
    bgcol = "economy")

tm_shape(World) +
  tm_text(text = "name", 
    size = .4, 
    bgcol = "economy", 
    bgcol.scale = tm_scale_categorical(values = cols4all::.P$hcl$cat$set2),
    bgcol_alpha = "pop_est",
    bgcol_alpha.scale = tm_scale_intervals(style = "kmeans"))
}
