% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layers.R
\name{tm_basemap}
\alias{tm_basemap}
\alias{tm_tiles}
\title{Draw tile layer}
\usage{
tm_basemap(server = NA, group = NA, alpha = NA)

tm_tiles(server, group = NA, alpha = 1)
}
\arguments{
\item{server}{name of the provider or url}

\item{group}{group}

\item{alpha}{alpha}
}
\description{
Creates a \code{\link{tmap-element}} that draws a tile layer. This feature is only available in view mode. For plot mode, a tile image can be retrieved by \code{\link[tmaptools:read_osm]{read_osm}}.
}
