% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_vars.R
\name{tm_vars}
\alias{tm_vars}
\title{tmap function to specify variables}
\usage{
tm_vars(x = NA, dimvalues = NULL, n = NA, multivariate = FALSE)
}
\arguments{
\item{x}{variable names, variable indices, or a dimension name}

\item{dimvalues}{dimension values}

\item{n}{if specified the first \code{n} variables are taken (or the first \code{n} dimension values)}

\item{multivariate}{in case multiple variables are specified, should they serve as facets (FALSE) or as a multivariate visual variable?}
}
\description{
tmap function to specify all variables in the shape object
}
