% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layout.R
\name{tm_view}
\alias{tm_view}
\title{View mode options}
\usage{
tm_view(
  use.WebGL,
  control.position,
  control.bases,
  control.overlays,
  set.bounds,
  set.view,
  set.zoom.limits,
  leaflet.options
)
}
\arguments{
\item{use.WebGL}{use webGL layers with leafgl}

\item{control.position}{position of the control attribute}

\item{control.bases}{base layers}

\item{control.overlays}{overlay layers}

\item{set.bounds}{logical that determines whether maximum bounds are set, or a bounding box. Not applicable in plot mode. In view mode, this is passed on to \code{\link[leaflet:setMaxBounds]{setMaxBounds}}}

\item{set.view}{numeric vector that determines the view. Either a vector of three: lng, lat, and zoom, or a single value: zoom. See \code{\link[leaflet:setView]{setView}}. Only applicable if \code{bbox} is not specified}

\item{set.zoom.limits}{numeric vector of two that set the minimum and maximum zoom levels (see \code{\link[leaflet:tileOptions]{tileOptions}}).}

\item{leaflet.options}{options passed on to \code{\link[leaflet:leafletOptions]{leafletOptions}}}
}
\description{
View mode options. These options are specific to the view mode.
}
