% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_element.R
\name{tmap-element}
\alias{tmap-element}
\alias{+.tmap}
\title{Stacking of tmap elements}
\usage{
\method{+}{tmap}(e1, e2)
}
\arguments{
\item{e1}{first tmap element}

\item{e2}{second tmap element}
}
\description{
The plus operator allows you to stack tmap elements (functions with a prefix \code{tm_})
}
