% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{tmap-package}
\alias{tmap-package}
\title{Tools to create thematic maps}
\description{
\tabular{ll}{
Package: \tab tmap\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2014-07-28\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\details{
This packages offers a complete work flow to create thematic maps (also called statistical maps). Maps are plotted in \pkg{ggplot2} style.
This package includes ready to use shape files from the World, Europe, and the Netherlands (both provinces as municipalities), which are used in the examples.

To get started, see \href{../doc/tm-nutshell.html}{\code{vignette("tm-nutshell")}}.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{GIS,}
\keyword{bubble}
\keyword{choropleth,}
\keyword{map}
\keyword{maps,}
\keyword{statistical}
\keyword{thematic}

