% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{tmap-package}
\alias{tmap}
\alias{tmap-package}
\title{Thematic Maps}
\description{
\tabular{ll}{
Package: \tab tmap\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2015-04-24\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\details{
Thematic maps are geographical maps in which statistical data are visualized. This package offers a flexible, layer-based, way to create thematic maps, such as choropleths and bubble maps.
To get started, see \href{../doc/tmap-nutshell.html}{\code{vignette("tmap-nutshell")}}.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{GIS,}
\keyword{bubble}
\keyword{choropleth,}
\keyword{map}
\keyword{maps,}
\keyword{statistical}
\keyword{thematic}

