% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{tmap-package}
\alias{tmap-package}
\alias{tmap}
\title{Thematic Map Visualization}
\description{
Thematic maps are geographical maps in which spatial data distributions are visualized. This package offers a flexible, layer-based, and easy to use approach to create thematic maps, such as choropleths and bubble maps. It is based on the grammar of graphics, and resembles the syntax of ggplot2.
}
\details{
This page provides a brief overview of all package functions. See \href{../doc/tmap-getstarted.html}{\code{vignette("tmap-getstarted")}} for a short introduction with examples.
}
\section{Quick plotting method}{

\tabular{ll}{
\code{\link{qtm}}\tab Plot a thematic map \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}
}

\section{Main plotting method}{

Shape specification:
\tabular{ll}{
\code{\link{tm_shape}}\tab Specify a shape object \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}

Aesthetics base layers:
\tabular{ll}{
\code{\link{tm_polygons}}\tab Create a polygon layer (with borders) \cr
\code{\link{tm_symbols}}\tab Create a layer of symbols \cr
\code{\link{tm_lines}}\tab Create a layer of lines \cr
\code{\link{tm_raster}}\tab Create a raster layer \cr
\code{\link{tm_text}}\tab Create a layer of text labels \cr
\code{\link{tm_basemap}}\tab Create a layer of basemap tiles \cr
\code{\link{tm_tiles}}\tab Create a layer of overlay tiles \cr
}

Aesthetics derived layers:
\tabular{ll}{
\code{\link{tm_fill}}\tab Create a polygon layer (without borders) \cr
\code{\link{tm_borders}}\tab Create polygon borders \cr
\code{\link{tm_bubbles}}\tab Create a layer of bubbles \cr
\code{\link{tm_squares}}\tab Create a layer of squares \cr
\code{\link{tm_dots}}\tab Create a layer of dots \cr
\code{\link{tm_markers}}\tab Create a layer of markers \cr
\code{\link{tm_iso}}\tab Create a layer of iso/contour lines \cr
\code{\link{tm_rgb}}\tab Create a raster layer of an image \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}

Faceting (small multiples)
\tabular{ll}{
\code{\link{tm_facets}}\tab Define facets \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}

Attributes:
\tabular{ll}{
\code{\link{tm_grid}}\tab Create grid lines \cr
\code{\link{tm_scale_bar}}\tab Create a scale bar \cr
\code{\link{tm_compass}}\tab Create a map compass \cr
\code{\link{tm_credits}}\tab Create a text for credits \cr
\code{\link{tm_logo}}\tab Create a logo \cr
\code{\link{tm_xlab} and \link{tm_ylab}}\tab Create axis labels \cr
\code{\link{tm_minimap}}\tab Create a minimap (view mode only) \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}

Layout element:
\tabular{ll}{
\code{\link{tm_layout}}\tab Adjust the layout (main function)\cr
\code{\link{tm_legend}}\tab Adjust the legend \cr
\code{\link{tm_view}}\tab Configure the interactive view mode \cr
\code{\link{tm_style}}\tab Apply a predefined style \cr
\code{\link{tm_format}}\tab Apply a predefined format \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}

Change options:
\tabular{ll}{
\code{\link{tmap_mode}}\tab Set the tmap mode: \code{"plot"} or \code{"view"}\cr
\code{\link{ttm}}\tab Toggle between the modes \cr
\code{\link{tmap_options}}\tab Set global tmap options (from \code{\link{tm_layout}}, \code{\link{tm_view}}, and a couple of others) \cr
\code{\link{tmap_style}}\tab Set the default style \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}

Create icons:
\tabular{ll}{
\code{\link{tmap_icons}}\tab Specify icons for markers or proportional symbols \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}
}

\section{Output functions}{
 
\tabular{ll}{
\code{\link{print}}\tab Plot in graphics device or view interactively in web browser or RStudio's viewer pane \cr
\code{\link{tmap_last}}\tab Redraw the last map \cr
\code{\link{tmap_leaflet}}\tab Obtain a leaflet widget object \cr
\code{\link{tmap_animation}}\tab Create an animation \cr
\code{\link{tmap_arrange}}\tab Create small multiples of separate maps \cr
\code{\link{tmap_save}}\tab Save thematic maps (either as image or HTML file) \cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}
}

\section{Spatial datasets}{
 
\tabular{ll}{
\code{\link{World}}\tab World country data (\code{\link[sf:sf]{sf}} object of polygons) \cr
\code{\link{NLD_prov}}\tab Netherlands province data (\code{\link[sf:sf]{sf}} object of  polygons) \cr
\code{\link{NLD_muni}}\tab Netherlands municipal data (\code{\link[sf:sf]{sf}} object of  polygons) \cr
\code{\link{metro}}\tab Metropolitan areas (\code{\link[sf:sf]{sf}} object of points) \cr
\code{\link{rivers}}\tab Rivers (\code{\link[sf:sf]{sf}} object of lines) \cr
\code{\link{land}}\tab Global land cover (\code{\link[stars:st_as_stars]{stars}} object)\cr
--------------------------- \tab --------------------------------------------------------------------------------------------------- \cr
}
}

\references{
Tennekes, M., 2018, {tmap}: Thematic Maps in {R}, Journal of Statistical Software, 84(6), 1-39, \href{https://doi.org/10.18637/jss.v084.i06}{DOI}
}
\seealso{
\href{../doc/tmap-getstarted.html}{\code{vignette("tmap-getstarted")}}
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\concept{GIS}
\concept{bubble map}
\concept{choropleth}
\concept{statistical maps}
\concept{thematic maps}
