% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_mode.R
\name{tmap_design_mode}
\alias{tmap_design_mode}
\title{Toggle design mode}
\usage{
tmap_design_mode()
}
\description{
When the so-called "design mode" is enabled, inner and outer margins, legend position, and aspect ratio are shown explicitly in plot mode. Also, information about aspect ratios is printed in the console. This function toggles the tmap option `design.mode`.
}
\seealso{
\code{\link{tmap_options}}
}
