% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_mode.R
\name{tmap_design_mode}
\alias{tmap_design_mode}
\title{Set the design mode}
\usage{
tmap_design_mode(design.mode)
}
\arguments{
\item{design.mode}{Logical value that determines the design mode.
If omitted then the design mode is toggled.}
}
\description{
When the so-called "design mode" is enabled,
the composition of the plot is shown explicitly in plot mode.
The used color codings is printed in the console as well as information about plot size and aspect ratio.
}
\details{
This function sets the global option \code{tmap.design.mode}.
It can be used as toggle function without arguments.
}
\seealso{
\code{\link[=tmap_options]{tmap_options()}}
}
