% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_add_legend.R, R/tm_element.R,
%   R/tmapChart.R, R/tmapGpar.R, R/tmapGridComp_charts.R, R/tmapGridLines.R,
%   R/tmapGridPolygons.R, R/tmapGridRaster.R, R/tmapGridSymbols.R,
%   R/tmapGridText.R, R/tmapLeaflet_layers.R, R/tmapOnLoad.R, R/tmapScaleAsIs.R,
%   R/tmapScaleBivariate.R, R/tmapScaleCategorical.R, R/tmapScaleIntervals.R,
%   R/tmapScaleRank.R, R/tmapScale_.R, R/tmapScale_defaults.R, R/tmapTrans.R
\docType{data}
\name{toTitleCase}
\alias{toTitleCase}
\alias{tmap_internal}
\alias{tm_element}
\alias{tm_element_list}
\alias{tmapChartBinned}
\alias{tmapChartBinned2d}
\alias{tmapChartBinned2d_numcat}
\alias{tmapChartBinned2d_numnum}
\alias{tmapChartBinned2d_catcat}
\alias{tmapChartRaw}
\alias{tmapChartRaw_nna}
\alias{tmapChartNone}
\alias{tmapChartPass}
\alias{tmapChartBinned_categorical}
\alias{tmapChartBinned_numeric}
\alias{tmapGpar}
\alias{tmapTpar}
\alias{lwd_to_mm}
\alias{tmapGridLines}
\alias{tmapGridPolygons}
\alias{tmapGridRaster}
\alias{tmapGridSymbols}
\alias{tmapGridText}
\alias{tmapLeafletPolygons}
\alias{tmapLeafletLines}
\alias{tmapLeafletSymbols}
\alias{tmapLeafletRaster}
\alias{tmapLeafletText}
\alias{.TMAP}
\alias{.TMAP_LEAFLET}
\alias{.TMAP_GRID}
\alias{tmapScaleAsIs}
\alias{tmapScaleBivariate}
\alias{tmapScaleCategorical}
\alias{tmapScaleIntervals}
\alias{tmapScaleRank}
\alias{tmapUsrCls}
\alias{format_aes_results}
\alias{chart_save}
\alias{data_type}
\alias{data_class}
\alias{tmapScale}
\alias{tmapScaleAuto}
\alias{tmapValuesCheck_col}
\alias{tmapValuesCheck_fill}
\alias{tmapValuesCheck_bgcol}
\alias{tmapValuesCheck_shape}
\alias{tmapValuesCheck_size}
\alias{tmapValuesCheck_area}
\alias{tmapValuesCheck_lwd}
\alias{tmapValuesCheck_lty}
\alias{tmapValuesCheck_xmod}
\alias{tmapValuesCheck_angle}
\alias{tmapValuesCheck_col_alpha}
\alias{tmapValuesCheck_fill_alpha}
\alias{tmapValuesCheck_bgcol_alpha}
\alias{tmapValuesCheck_text}
\alias{tmapValuesCheck_fontface}
\alias{tmapValuesIsDiv_fill}
\alias{tmapValuesIsDiv_col}
\alias{tmapValuesIsDiv_bgcol}
\alias{tmapValuesIsDiv_size}
\alias{tmapValuesIsDiv_area}
\alias{tmapValuesIsDiv_lwd}
\alias{tmapValuesIsDiv_lty}
\alias{tmapValuesIsDiv_col_alpha}
\alias{tmapValuesIsDiv_fill_alpha}
\alias{tmapValuesIsDiv_bgcol_alpha}
\alias{tmapValuesIsDiv_xmod}
\alias{tmapValuesIsDiv_angle}
\alias{tmapValuesIsDiv_shape}
\alias{tmapValuesIsDiv_text}
\alias{tmapValuesIsDiv_fontface}
\alias{tmapValuesRange_fill}
\alias{tmapValuesRange_col}
\alias{tmapValuesRange_bgcol}
\alias{tmapValuesRange_shape}
\alias{tmapValuesRange_lty}
\alias{tmapValuesRange_size}
\alias{tmapValuesRange_area}
\alias{tmapValuesRange_lwd}
\alias{tmapValuesRange_col_alpha}
\alias{tmapValuesRange_fill_alpha}
\alias{tmapValuesRange_bgcol_alpha}
\alias{tmapValuesRange_xmod}
\alias{tmapValuesRange_angle}
\alias{tmapValuesRange_text}
\alias{tmapValuesRange_fontface}
\alias{tmapValuesVV_fill}
\alias{tmapValuesVV_col}
\alias{tmapValuesVV_bgcol}
\alias{tmapValuesVV_shape}
\alias{tmapValuesVV_lty}
\alias{tmapValuesVV_size}
\alias{tmapValuesVV_area}
\alias{tmapValuesVV_lwd}
\alias{tmapValuesVV_col_alpha}
\alias{tmapValuesVV_fill_alpha}
\alias{tmapValuesVV_bgcol_alpha}
\alias{tmapValuesVV_xmod}
\alias{tmapValuesVV_angle}
\alias{tmapValuesVV_text}
\alias{tmapValuesVV_fontface}
\alias{tmapValuesSubmit_col}
\alias{tmapValuesSubmit_fill}
\alias{tmapValuesSubmit_bgcol}
\alias{tmapValuesSubmit_size}
\alias{tmapValuesSubmit_area}
\alias{tmapValuesSubmit_xmod}
\alias{tmapValuesSubmit_ymod}
\alias{tmapValuesSubmit_angle}
\alias{tmapValuesSubmit_lwd}
\alias{tmapValuesSubmit_lty}
\alias{tmapValuesSubmit_shape}
\alias{tmapValuesSubmit_col_alpha}
\alias{tmapValuesSubmit_fill_alpha}
\alias{tmapValuesSubmit_bgcol_alpha}
\alias{tmapValuesSubmit_text}
\alias{tmapValuesSubmit_fontface}
\alias{tmapValuesScale_col}
\alias{tmapValuesScale_fill}
\alias{tmapValuesScale_bgcol}
\alias{tmapValuesScale_size}
\alias{tmapValuesScale_area}
\alias{tmapValuesScale_lwd}
\alias{tmapValuesScale_lty}
\alias{tmapValuesScale_shape}
\alias{tmapValuesScale_col_alpha}
\alias{tmapValuesScale_fill_alpha}
\alias{tmapValuesScale_bgcol_alpha}
\alias{tmapValuesScale_text}
\alias{tmapValuesScale_fontface}
\alias{tmapValuesScale_xmod}
\alias{tmapValuesScale_angle}
\alias{tmapValuesColorize_col}
\alias{tmapValuesColorize_fill}
\alias{tmapValuesColorize_bgcol}
\alias{tmapValuesColorize_size}
\alias{tmapValuesColorize_area}
\alias{tmapValuesColorize_lwd}
\alias{tmapValuesColorize_lty}
\alias{tmapValuesColorize_shape}
\alias{tmapValuesColorize_col_alpha}
\alias{tmapValuesColorize_fill_alpha}
\alias{tmapValuesColorize_bgcol_alpha}
\alias{tmapValuesColorize_text}
\alias{tmapValuesColorize_fontface}
\alias{tmapValuesColorize_xmod}
\alias{tmapValuesColorize_angle}
\alias{tmapSeq}
\alias{transform_values}
\alias{tmapValuesCVV_fill}
\alias{tmapValuesCVV_col}
\alias{tmapValuesCVV_bgcol}
\alias{tmapValuesCVV_size}
\alias{tmapValuesCVV_area}
\alias{tmapValuesCVV_lwd}
\alias{tmapValuesCVV_col_alpha}
\alias{tmapValuesCVV_fill_alpha}
\alias{tmapValuesCVV_bgcol_alpha}
\alias{tmapValuesCVV_xmod}
\alias{tmapValuesCVV_ymod}
\alias{tmapValuesCVV_angle}
\alias{tmapValuesCVV_shape}
\alias{tmapValuesCVV_lty}
\alias{tmapValuesCVV_text}
\alias{tmapValuesCVV_fontface}
\alias{tmapValuesBVV_fill}
\alias{tmapValuesBVV_col}
\alias{tmapValuesBVV_bgcol}
\alias{tmapValuesCheck_num}
\alias{tmapValuesIsDiv_num}
\alias{tmapValuesRange_num}
\alias{tmapValuesVV_num}
\alias{tmapValuesSubmit_num}
\alias{tmapValuesScale_num}
\alias{tmapValuesColorize_num}
\alias{tmapValuesCVV_num}
\alias{tmapValuesCheck_skip}
\alias{tmapValuesIsDiv_skip}
\alias{tmapValuesRange_skip}
\alias{tmapValuesVV_skip}
\alias{tmapValuesSubmit_skip}
\alias{tmapValuesScale_skip}
\alias{tmapValuesColorize_skip}
\alias{tmapValuesCVV_skip}
\alias{tmapTransCentroid}
\alias{tmapTransRaster}
\alias{tmapTransPolygons}
\alias{tmapTransLines}
\title{Internal methods for tmap extensions}
\format{
An object of class \code{environment} of length 9.

An object of class \code{environment} of length 0.

An object of class \code{environment} of length 1.
}
\usage{
toTitleCase(x)

tm_element(..., subclass = NULL)

tm_element_list(...)

tmapChartBinned(chart, bin_colors, breaks_def, na.show, x1)

tmapChartBinned2d(chart, chart1, chart2)

tmapChartBinned2d_numcat(chart, chart1, chart2)

tmapChartBinned2d_numnum(chart, chart1, chart2)

tmapChartBinned2d_catcat(chart, chart1, chart2)

tmapChartRaw(chart, na.show, x1, ...)

tmapChartRaw_nna(chart, na.show, x1, ...)

tmapChartNone(chart, na.show, x1, ...)

tmapChartPass(chart, na.show, x1, ...)

tmapChartBinned_categorical(chart, bin_colors, breaks_def, na.show, x1)

tmapChartBinned_numeric(chart, bin_colors, breaks_def, na.show, x1)

tmapGpar(
  fill = NULL,
  col = NULL,
  shape = NULL,
  size = NULL,
  fill_alpha = NULL,
  col_alpha = NULL,
  lty = NULL,
  lwd = NULL,
  linejoin = NULL,
  lineend = NULL,
  ...
)

tmapTpar(...)

lwd_to_mm(value, unit = "bigpts")

tmapGridLines(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapGridPolygons(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapGridRaster(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapGridSymbols(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapGridText(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapLeafletPolygons(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapLeafletLines(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapLeafletSymbols(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapLeafletRaster(
  shpTM,
  dt,
  gp,
  pdt,
  popup.format,
  hdt,
  idt,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapLeafletText(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

.TMAP

.TMAP_LEAFLET

.TMAP_GRID

tmapScaleAsIs(
  x1,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE
)

tmapScaleBivariate(
  x1,
  x2,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE
)

tmapScaleCategorical(
  x1,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE
)

tmapScaleIntervals(
  x1,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE
)

tmapScaleRank(
  x1,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE
)

tmapUsrCls(x)

format_aes_results(values, ord = NULL, legend, chart)

chart_save(legend)

data_type(x)

data_class(x, check_for_color_class = FALSE, midpoint_enabled = FALSE)

tmapScale(aes, value, scale, legend, chart, free)

tmapScaleAuto(
  x1,
  scale,
  legend,
  chart,
  o,
  aes,
  layer,
  layer_args,
  sortRev,
  bypass_ord,
  submit_legend = TRUE,
  ...
)

tmapValuesCheck_col(x, is_var = TRUE)

tmapValuesCheck_fill(x, is_var = TRUE)

tmapValuesCheck_bgcol(x, is_var = TRUE)

tmapValuesCheck_shape(x, is_var = TRUE)

tmapValuesCheck_size(x, is_var = TRUE)

tmapValuesCheck_area(x, is_var = TRUE)

tmapValuesCheck_lwd(x, is_var = TRUE)

tmapValuesCheck_lty(x, is_var = TRUE)

tmapValuesCheck_xmod(x, is_var = TRUE)

tmapValuesCheck_angle(x, is_var = TRUE)

tmapValuesCheck_col_alpha(x, is_var = TRUE)

tmapValuesCheck_fill_alpha(x, is_var = TRUE)

tmapValuesCheck_bgcol_alpha(x, is_var = TRUE)

tmapValuesCheck_area(x, is_var = TRUE)

tmapValuesCheck_text(x, is_var = TRUE)

tmapValuesCheck_fontface(x, is_var = TRUE)

tmapValuesIsDiv_fill(x)

tmapValuesIsDiv_col(x)

tmapValuesIsDiv_bgcol(x)

tmapValuesIsDiv_size(x)

tmapValuesIsDiv_area(x)

tmapValuesIsDiv_lwd(x)

tmapValuesIsDiv_lty(x)

tmapValuesIsDiv_col_alpha(x)

tmapValuesIsDiv_fill_alpha(x)

tmapValuesIsDiv_bgcol_alpha(x)

tmapValuesIsDiv_area(x)

tmapValuesIsDiv_xmod(x)

tmapValuesIsDiv_angle(x)

tmapValuesIsDiv_shape(x)

tmapValuesIsDiv_text(x)

tmapValuesIsDiv_fontface(x)

tmapValuesRange_fill(x, n, isdiv)

tmapValuesRange_col(x, n, isdiv)

tmapValuesRange_bgcol(x, n, isdiv)

tmapValuesRange_shape(x, n, isdiv)

tmapValuesRange_lty(x, n, isdiv)

tmapValuesRange_size(x, n, isdiv)

tmapValuesRange_area(x, n, isdiv)

tmapValuesRange_lwd(x, n, isdiv)

tmapValuesRange_col_alpha(x, n, isdiv)

tmapValuesRange_fill_alpha(x, n, isdiv)

tmapValuesRange_bgcol_alpha(x, n, isdiv)

tmapValuesRange_area(x, n, isdiv)

tmapValuesRange_xmod(x, n, isdiv)

tmapValuesRange_angle(x, n, isdiv)

tmapValuesRange_text(x, n, isdiv)

tmapValuesRange_fontface(x, n, isdiv)

tmapValuesVV_fill(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o,
  aes = "fill"
)

tmapValuesVV_col(...)

tmapValuesVV_bgcol(...)

tmapValuesVV_shape(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesVV_lty(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesVV_size(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesVV_area(...)

tmapValuesVV_lwd(...)

tmapValuesVV_col_alpha(...)

tmapValuesVV_fill_alpha(...)

tmapValuesVV_bgcol_alpha(...)

tmapValuesVV_area(...)

tmapValuesVV_xmod(...)

tmapValuesVV_angle(...)

tmapValuesVV_text(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesVV_fontface(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesSubmit_col(x, args)

tmapValuesSubmit_fill(x, args)

tmapValuesSubmit_bgcol(x, args)

tmapValuesSubmit_size(x, args)

tmapValuesSubmit_area(x, args)

tmapValuesSubmit_xmod(x, args)

tmapValuesSubmit_ymod(x, args)

tmapValuesSubmit_angle(x, args)

tmapValuesSubmit_lwd(x, args)

tmapValuesSubmit_lty(x, args)

tmapValuesSubmit_shape(x, args)

tmapValuesSubmit_col_alpha(x, args)

tmapValuesSubmit_fill_alpha(x, args)

tmapValuesSubmit_bgcol_alpha(x, args)

tmapValuesSubmit_text(x, args)

tmapValuesSubmit_fontface(x, args)

tmapValuesScale_col(x, scale)

tmapValuesScale_fill(x, scale)

tmapValuesScale_bgcol(x, scale)

tmapValuesScale_size(x, scale)

tmapValuesScale_area(x, scale)

tmapValuesScale_lwd(x, scale)

tmapValuesScale_lty(x, scale)

tmapValuesScale_shape(x, scale)

tmapValuesScale_col_alpha(x, scale)

tmapValuesScale_fill_alpha(x, scale)

tmapValuesScale_bgcol_alpha(x, scale)

tmapValuesScale_text(x, scale)

tmapValuesScale_fontface(x, scale)

tmapValuesScale_xmod(x, scale)

tmapValuesScale_angle(x, scale)

tmapValuesColorize_col(x, pc)

tmapValuesColorize_fill(x, pc)

tmapValuesColorize_bgcol(x, pc)

tmapValuesColorize_size(x, pc)

tmapValuesColorize_area(x, pc)

tmapValuesColorize_lwd(x, pc)

tmapValuesColorize_lty(x, pc)

tmapValuesColorize_shape(x, pc)

tmapValuesColorize_col_alpha(x, pc)

tmapValuesColorize_fill_alpha(x, pc)

tmapValuesColorize_bgcol_alpha(x, pc)

tmapValuesColorize_text(x, pc)

tmapValuesColorize_fontface(x, pc)

tmapValuesColorize_xmod(x, pc)

tmapValuesColorize_angle(x, pc)

tmapSeq(s, n = NULL)

transform_values(x, lim, rng, power, scale, include.neutral = TRUE)

tmapValuesCVV_fill(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_col(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_bgcol(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_size(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_area(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_lwd(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_col_alpha(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_fill_alpha(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_bgcol_alpha(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_area(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_xmod(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_ymod(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_angle(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_shape(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_lty(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_text(x, value.na, n, range, scale, rep, o)

tmapValuesCVV_fontface(x, value.na, n, range, scale, rep, o)

tmapValuesBVV_fill(x, value.na, m, n, scale, rep, o)

tmapValuesBVV_col(x, value.na, m, n, scale, rep, o)

tmapValuesBVV_bgcol(x, value.na, m, n, scale, rep, o)

tmapValuesCheck_num(x, is_var = TRUE)

tmapValuesIsDiv_num(x)

tmapValuesRange_num(x, n, isdiv)

tmapValuesVV_num(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesSubmit_num(x, args)

tmapValuesScale_num(x, scale)

tmapValuesColorize_num(x, pc)

tmapValuesCVV_num(x, value.na, n, range, scale, rep, o)

tmapValuesCheck_skip(x, is_var)

tmapValuesIsDiv_skip(x)

tmapValuesRange_skip(x, n, isdiv)

tmapValuesVV_skip(
  x,
  value.na,
  isdiv,
  n,
  dvalues,
  are_breaks,
  midpoint,
  range,
  scale,
  rep,
  o
)

tmapValuesSubmit_skip(x, args)

tmapValuesScale_skip(x, scale)

tmapValuesColorize_skip(x, pc)

tmapValuesCVV_skip(x, value.na, n, range, scale, rep, o)

tmapTransCentroid(
  shpTM,
  xmod = NULL,
  ymod = NULL,
  ord__,
  plot.order,
  args,
  scale
)

tmapTransRaster(shpTM, ord__, plot.order, args)

tmapTransPolygons(shpTM, ord__, plot.order, args, scale)

tmapTransLines(shpTM, ord__, plot.order, args, scale)
}
\arguments{
\item{...}{args}

\item{subclass}{subclass}

\item{fill, col, shape, size, fill_alpha, col_alpha, lty, lwd, linejoin, lineend}{visual variables}

\item{shpTM}{shpTM}

\item{scale}{scale}

\item{args}{args}

\item{xmod, ymod}{xmod and ymod}

\item{ord__}{ord}

\item{plot.order}{plot.order}
}
\description{
Internal methods for tmap extensions
}
\keyword{datasets}
\keyword{internal}
