% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_leaflet.R
\name{tmap_leaflet}
\alias{tmap_leaflet}
\alias{tmap_grid}
\title{Export tmap to the format of the used graphics mode}
\usage{
tmap_leaflet(x, show = FALSE, ...)

tmap_grid(x, show = FALSE, ...)
}
\arguments{
\item{x}{description}

\item{show}{show the map?}

\item{...}{arguments passed on the \code{\link[=print.tmap]{print.tmap()}}}
}
\value{
\code{tmap_grid} returns a \code{\link[grid:grid.grob]{grid::grob()}} object (\code{"plot"} mode) and \code{tmap_leaflet} a \code{\link[leaflet:leaflet]{leaflet::leaflet()}} object (\code{"view"} mode). In case small multiples are shown, a list is returned.
}
\description{
Export tmap to the format of the used graphics mode. \code{tmap_grid} returns a \code{\link[grid:grid.grob]{grid::grob()}} object (\verb{"plot" mode}) and \code{tmap_leaflet} a \code{\link[leaflet:leaflet]{leaflet::leaflet()}} object (\code{"view"} mode).
}
\examples{
map = tm_shape(World) + tm_polygons()
tmap_leaflet(map, show = TRUE)
}
