% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_options.R
\name{tmap_options}
\alias{tmap_options}
\alias{tmap_options_mode}
\alias{tm_options}
\title{tmap options}
\usage{
tmap_options(...)

tmap_options_mode(mode = NA, default.options = FALSE)

tm_options(...)
}
\arguments{
\item{...}{See details}
}
\description{
tmap options
}
\details{
| option        | description |
| ------------- |:-------------:|
| modes		|  Mode specific options. It is a named list where names correspond to the available modes. Each item is a list of options. |
| crs		|  Map crs (see \code{\link{tm_shape}}). \code{NA} means the crs is specified in \code{\link{tm_shape}}. The crs that is used by the transformation functions is defined in \code{\link{tm_shape}}.|
| facet.max		| Maximum number of facets |
| facet.flip		| Should facets be flipped (in case of facet wrap)? This can also be set via \code{\link{tm_facets_flip}} |
| raster.max.cells		| Maximum number of raster grid cells  |
| show.messages		| Show messages? |
| show.warnings		| Show warnings? |
| output.format		| Output format |
| output.size		| Output size |
| output.dpi		| Output dpi |
| output.dpi.animation		| Output dpi for animations |
| value.const		| Default visual value constants e.g. the default fill color for \code{tm_shape(World) + tm_polygons()}. A list is required with per visual variable a value. |
| value.na		| Default visual values that are used to visualize NA data values. A list is required with per visual variable a value.|
| value.null		| Default visual values that are used to visualize null (out-of-scope) data values. A list is required with per visual variable a value.|
| value.blank		| Default visual values that correspond to blank. For color these are \code{"#00000000"} meaning transparent. A list is required with per visual variable a value. |
| values.var		| Default values when a data variable to mapped to a visual variable, e.g. a color palette. A list is required with per visual variable a value. |
| values.range		| Default range for values. See \code{values.range} of \code{\link{tm_scale_categorical}}. A list is required with per visual variable a value.
| value.neutral		| Default values for when a data variable to mapped to a visual variable, e.g. a color palette. A list is required with per visual variable a value. | |
| scales.var		| Default scales. |
| label.format		| Format for the labels (was legend.format in tmap v3) |
| label.na		| Default label for missing values |
See \code{\link{tm_layout}} for layout specific options
}
