% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_options.R
\name{tmap_options}
\alias{tmap_options}
\alias{tmap_options_diff}
\alias{tmap_options_reset}
\title{Options for tmap}
\usage{
tmap_options(unit, limits, ...)

tmap_options_diff()

tmap_options_reset()
}
\arguments{
\item{unit}{This is the default value for the \code{unit} argument of \code{\link{tm_shape}}. It specifies the unit of measurement, which is used in the scale bar and the calculation of density values. By default (when loading the package), it is \code{"metric"}. Other valid values are \code{"imperial"}, \code{"km"}, \code{"m"}, \code{"mi"}, and \code{"ft"}.}

\item{limits}{This option determines how many facets (small multiples) are allowed for per mode. It should be a vector of two numeric values named \code{facets.view} and \code{facets.plot}. By default (i.e. when loading the package), it is set to \code{c(facets.view = 4, facets.plot = 64)}}

\item{...}{options from \code{\link{tm_layout}} or \code{\link{tm_view}}. Note that the difference with using \code{\link{tm_layout}} or \code{\link{tm_view}} directly, is that options set with \code{tmap_options} remain for the entire session (unless changed with \code{tmap_options} or \code{\link{tmap_style}}).}
}
\description{
Get or set global options for tmap. The behaviour of \code{tmap_options} is similar to \code{\link[base:options]{options}}: all tmap options are retrieved when this function is called without arguments. When arguments are specified, the corresponding options are set, and the old values are silently returned. The function \code{tmap_options_reset} is used to reset all options back to the default values (also the \code{style} is reset to \code{"white"}). Differences with the default values can be shown with \code{tmap_options_diff}.
}
\examples{
# load data
data(World)

# get current options
tmap_options()

# get current style
tmap_style()

# plot map (with default options)
tm_shape(World) + tm_polygons("HPI")

# change style to cobalt
tmap_style("cobalt")

# observe the changed options
tmap_options_diff()

# plot the map again
tm_shape(World) + tm_polygons("HPI")

# change the background color
tmap_options(bg.color = "red")

# observe the changed options
tmap_options_diff()

# note that the current style is modified
tmap_style()

# plot the map again
tm_shape(World) + tm_polygons("HPI")

# the specified arguments of tm_layout and tm_view will override the options temporarily:
tm_shape(World) + tm_polygons("HPI") + tm_layout(bg.color="purple")

# when tm_style_ is called, it will override all options temporarily:
tm_shape(World) + tm_polygons("HPI") + tm_layout(bg.color="purple") + tm_style_classic()

# reset all options
tmap_options_reset()

# check style and options
tmap_style()
tmap_options_diff()
}
\seealso{
\code{\link{tm_layout}}, \code{\link{tm_view}}, and \code{\link{tmap_style}}
}
