% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_layout.R, R/tmap_options.R
\name{tm_check_fix}
\alias{tm_check_fix}
\alias{tmap_options}
\alias{tmap_options_mode}
\alias{tmap_options_diff}
\alias{tmap_options_reset}
\alias{tmap_options_save}
\title{tmap options}
\usage{
tm_check_fix()

tmap_options(
  ...,
  crs,
  facet.max,
  facet.flip,
  free.scales,
  raster.max_cells,
  show.messages,
  show.warnings,
  output.format,
  output.size,
  output.dpi,
  output.dpi.animation,
  value.const,
  value.na,
  value.null,
  value.blank,
  values.var,
  values.range,
  value.neutral,
  values.scale,
  scales.var,
  scale.misc.args,
  continuous.nclass_per_legend_break,
  continuous.nclasses,
  label.format,
  label.na,
  scale,
  asp,
  bg.color,
  outer.bg.color,
  frame,
  frame.lwd,
  frame.r,
  frame.double_line,
  outer.margins,
  inner.margins,
  inner.margins.extra,
  meta.margins,
  meta.auto_margins,
  between_margin,
  panel.margin,
  component.offset,
  component.stack_margin,
  grid.mark.height,
  xylab.height,
  coords.height,
  xlab.show,
  xlab.text,
  xlab.size,
  xlab.color,
  xlab.rotation,
  xlab.space,
  xlab.fontface,
  xlab.fontfamily,
  xlab.side,
  ylab.show,
  ylab.text,
  ylab.size,
  ylab.color,
  ylab.rotation,
  ylab.space,
  ylab.fontface,
  ylab.fontfamily,
  ylab.side,
  panel.type,
  panel.wrap.pos,
  panel.xtab.pos,
  unit,
  color.sepia_intensity,
  color.saturation,
  color_vision_deficiency_sim,
  text.fontface,
  text.fontfamily,
  component.position,
  component.autoscale,
  legend.show,
  legend.design,
  legend.orientation,
  legend.position,
  legend.width,
  legend.height,
  legend.stack,
  legend.group.frame,
  legend.resize_as_group,
  legend.reverse,
  legend.na.show,
  legend.title.color,
  legend.title.size,
  legend.title.fontface,
  legend.title.fontfamily,
  legend.xlab.color,
  legend.xlab.size,
  legend.xlab.fontface,
  legend.xlab.fontfamily,
  legend.ylab.color,
  legend.ylab.size,
  legend.ylab.fontface,
  legend.ylab.fontfamily,
  legend.text.color,
  legend.text.size,
  legend.text.fontface,
  legend.text.fontfamily,
  legend.frame,
  legend.frame.lwd,
  legend.frame.r,
  legend.bg.color,
  legend.bg.alpha,
  legend.only,
  legend.settings.standard.portrait,
  legend.settings.standard.landscape,
  chart.show,
  chart.plot.axis.x,
  chart.plot.axis.y,
  chart.position,
  chart.width,
  chart.height,
  chart.stack,
  chart.group.frame,
  chart.resize_as_group,
  chart.reverse,
  chart.na.show,
  chart.title.color,
  chart.title.size,
  chart.title.fontface,
  chart.title.fontfamily,
  chart.xlab.color,
  chart.xlab.size,
  chart.xlab.fontface,
  chart.xlab.fontfamily,
  chart.ylab.color,
  chart.ylab.size,
  chart.ylab.fontface,
  chart.ylab.fontfamily,
  chart.text.color,
  chart.text.size,
  chart.text.fontface,
  chart.text.fontfamily,
  chart.frame,
  chart.frame.lwd,
  chart.frame.r,
  chart.bg.color,
  chart.bg.alpha,
  chart.object.color,
  title.show,
  title.size,
  title.color,
  title.fontface,
  title.fontfamily,
  title.bg.color,
  title.bg.alpha,
  title.padding,
  title.frame,
  title.frame.lwd,
  title.frame.r,
  title.stack,
  title.position,
  title.width,
  title.group.frame,
  title.resize_as_group,
  credits.show,
  credits.size,
  credits.color,
  credits.fontface,
  credits.fontfamily,
  credits.bg.color,
  credits.bg.alpha,
  credits.padding,
  credits.frame,
  credits.frame.lwd,
  credits.frame.r,
  credits.stack,
  credits.position,
  credits.width,
  credits.heigth,
  credits.group.frame,
  credits.resize_as_group,
  compass.north,
  compass.type,
  compass.text.size,
  compass.size,
  compass.show.labels,
  compass.cardinal.directions,
  compass.text.color,
  compass.color.dark,
  compass.color.light,
  compass.lwd,
  compass.bg.color,
  compass.bg.alpha,
  compass.margins,
  compass.show,
  compass.stack,
  compass.position,
  compass.frame,
  compass.frame.lwd,
  compass.frame.r,
  compass.group.frame,
  compass.resize_as_group,
  logo.height,
  logo.margins,
  logo.between_margin,
  logo.show,
  logo.stack,
  logo.position,
  logo.frame,
  logo.frame.lwd,
  logo.frame.r,
  logo.group.frame,
  logo.resize_as_group,
  scalebar.show,
  scalebar.breaks,
  scalebar.width,
  scalebar.text.size,
  scalebar.text.color,
  scalebar.color.dark,
  scalebar.color.light,
  scalebar.lwd,
  scalebar.bg.color,
  scalebar.bg.alpha,
  scalebar.size,
  scalebar.margins,
  scalebar.stack,
  scalebar.position,
  scalebar.frame,
  scalebar.frame.lwd,
  scalebar.frame.r,
  scalebar.group.frame,
  scalebar.resize_as_group,
  grid.show,
  grid.labels.pos,
  grid.x,
  grid.y,
  grid.n.x,
  grid.n.y,
  grid.crs,
  grid.col,
  grid.lwd,
  grid.alpha,
  grid.labels.show,
  grid.labels.size,
  grid.labels.col,
  grid.labels.rot,
  grid.labels.format,
  grid.labels.cardinal,
  grid.labels.margin.x,
  grid.labels.margin.y,
  grid.labels.space.x,
  grid.labels.space.y,
  grid.labels.inside_frame,
  grid.ticks,
  grid.lines,
  grid.ndiscr,
  mouse_coordinates.stack,
  mouse_coordinates.position,
  mouse_coordinates.show,
  minimap.server,
  minimap.toggle,
  minimap.stack,
  minimap.position,
  minimap.show,
  panel.show,
  panel.labels,
  panel.label.size,
  panel.label.color,
  panel.label.fontface,
  panel.label.fontfamily,
  panel.label.bg.color,
  panel.label.frame,
  panel.label.frame.lwd,
  panel.label.frame.r,
  panel.label.height,
  panel.label.rot,
  bbox,
  set.bounds,
  set.view,
  set.zoom.limits,
  qtm.scalebar,
  qtm.minimap,
  qtm.mouse_coordinates,
  earth.boundary,
  earth.boundary.color,
  earth.boundary.lwd,
  earth.datum,
  space.color,
  check_and_fix,
  basemap.show,
  basemap.server,
  basemap.alpha,
  basemap.zoom,
  tiles.show,
  tiles.server,
  tiles.alpha,
  tiles.zoom,
  attr.color,
  title = NULL,
  main.title = NULL,
  view.legend.position
)

tmap_options_mode(mode = NA, default.options = FALSE)

tmap_options_diff()

tmap_options_reset()

tmap_options_save(style)
}
\arguments{
\item{...}{List of tmap options to be set, or option names (characters) to be returned (see details)}

\item{crs}{Map crs (see \code{\link[=tm_shape]{tm_shape()}}). \code{NA} means the crs is specified in \code{\link[=tm_shape]{tm_shape()}}. The crs that is used by the transformation functions is defined in \code{\link[=tm_shape]{tm_shape()}}.}

\item{facet.max}{Maximum number of facets}

\item{facet.flip}{Should facets be flipped (in case of facet wrap)? This can also be set via \code{\link[=tm_facets_flip]{tm_facets_flip()}}}

\item{free.scales}{For backward compatibility: if this value is set, it will be used to impute the free arguments in the layer functions}

\item{raster.max_cells}{Maximum number of raster grid cells}

\item{show.messages}{Show messages?}

\item{show.warnings}{Show warnings?}

\item{output.format}{Output format}

\item{output.size}{Output size}

\item{output.dpi}{Output dpi}

\item{output.dpi.animation}{Output dpi for animations}

\item{value.const}{Default visual value constants e.g. the default fill color for \code{tm_shape(World) + tm_polygons()}. A list is required with per visual variable a value.}

\item{value.na}{Default visual values that are used to visualize NA data values. A list is required with per visual variable a value.}

\item{value.null}{Default visual values that are used to visualize null (out-of-scope) data values. A list is required with per visual variable a value.}

\item{value.blank}{Default visual values that correspond to blank. For color these are \code{"#00000000"} meaning transparent. A list is required with per visual variable a value.}

\item{values.var}{Default values when a data variable to mapped to a visual variable, e.g. a color palette. A list is required with per visual variable a value.}

\item{values.range}{Default range for values. See \code{values.range} of \code{\link[=tm_scale_categorical]{tm_scale_categorical()}}. A list is required with per visual variable a value.}

\item{value.neutral}{Default values for when a data variable to mapped to a visual variable, e.g. a color palette. A list is required with per visual variable a value.}

\item{values.scale}{Default scales (as in object sizes) for values. See \code{values.range} of \code{\link[=tm_scale_categorical]{tm_scale_categorical()}}. A list is required with per visual variable a value.}

\item{scales.var}{Default scale functions per visual variable and type of data variable. A list is required with per visual variable per data type.}

\item{scale.misc.args}{Default values of scale function-specific arguments. A list is required with per scale function and optional per visual variable.}

\item{continuous.nclass_per_legend_break}{The number of continuous legend breaks within one 'unit' (label).  The dafault value is 50.}

\item{continuous.nclasses}{the number of classes of a continuous scale. Should be odd.  The dafault value is 101.}

\item{label.format}{Format for the labels (was \code{legend.format} in tmap v3).}

\item{label.na}{Default label for missing values.}

\item{scale}{Overall scale of the map}

\item{asp}{Aspect ratio of each map. When \code{asp} is set to \code{NA} (default) the aspect ratio will be adjusted to the used shapes. When set to 0 the aspect ratio is
adjusted to the size of the device divided by the number of columns and rows.}

\item{bg.color}{Background color of the map.}

\item{outer.bg.color}{Background color of map outside the frame.}

\item{frame}{test123}

\item{frame.lwd}{test123}

\item{frame.r}{test123}

\item{frame.double_line}{test123}

\item{outer.margins}{test123}

\item{inner.margins}{test123}

\item{inner.margins.extra}{test123}

\item{meta.margins}{test123}

\item{meta.auto_margins}{test123}

\item{between_margin}{test123}

\item{panel.margin}{test123}

\item{component.offset}{test123}

\item{component.stack_margin}{test123}

\item{grid.mark.height}{test123}

\item{xylab.height}{test123}

\item{coords.height}{test123}

\item{xlab.show}{test123}

\item{xlab.text}{test123}

\item{xlab.size}{test123}

\item{xlab.color}{test123}

\item{xlab.rotation}{test123}

\item{xlab.space}{test123}

\item{xlab.fontface}{test123}

\item{xlab.fontfamily}{test123}

\item{xlab.side}{test123}

\item{ylab.show}{test123}

\item{ylab.text}{test123}

\item{ylab.size}{test123}

\item{ylab.color}{test123}

\item{ylab.rotation}{test123}

\item{ylab.space}{test123}

\item{ylab.fontface}{test123}

\item{ylab.fontfamily}{test123}

\item{ylab.side}{test123}

\item{panel.type}{test123}

\item{panel.wrap.pos}{test123}

\item{panel.xtab.pos}{test123}

\item{unit}{test123}

\item{color.sepia_intensity}{test123}

\item{color.saturation}{test123}

\item{color_vision_deficiency_sim}{test123}

\item{text.fontface}{test123}

\item{text.fontfamily}{test123}

\item{component.position}{test123}

\item{component.autoscale}{test123}

\item{legend.show}{test123}

\item{legend.design}{test123}

\item{legend.orientation}{test123}

\item{legend.position}{test123}

\item{legend.width}{test123}

\item{legend.height}{test123}

\item{legend.stack}{test123}

\item{legend.group.frame}{test123}

\item{legend.resize_as_group}{test123}

\item{legend.reverse}{test123}

\item{legend.na.show}{test123}

\item{legend.title.color}{test123}

\item{legend.title.size}{test123}

\item{legend.title.fontface}{test123}

\item{legend.title.fontfamily}{test123}

\item{legend.xlab.color}{test123}

\item{legend.xlab.size}{test123}

\item{legend.xlab.fontface}{test123}

\item{legend.xlab.fontfamily}{test123}

\item{legend.ylab.color}{test123}

\item{legend.ylab.size}{test123}

\item{legend.ylab.fontface}{test123}

\item{legend.ylab.fontfamily}{test123}

\item{legend.text.color}{test123}

\item{legend.text.size}{test123}

\item{legend.text.fontface}{test123}

\item{legend.text.fontfamily}{test123}

\item{legend.frame}{test123}

\item{legend.frame.lwd}{test123}

\item{legend.frame.r}{test123}

\item{legend.bg.color}{test123}

\item{legend.bg.alpha}{test123}

\item{legend.only}{test123}

\item{legend.settings.standard.portrait}{test123}

\item{legend.settings.standard.landscape}{test123}

\item{chart.show}{test123}

\item{chart.plot.axis.x}{test123}

\item{chart.plot.axis.y}{test123}

\item{chart.position}{test123}

\item{chart.width}{test123}

\item{chart.height}{test123}

\item{chart.stack}{test123}

\item{chart.group.frame}{test123}

\item{chart.resize_as_group}{test123}

\item{chart.reverse}{test123}

\item{chart.na.show}{test123}

\item{chart.title.color}{test123}

\item{chart.title.size}{test123}

\item{chart.title.fontface}{test123}

\item{chart.title.fontfamily}{test123}

\item{chart.xlab.color}{test123}

\item{chart.xlab.size}{test123}

\item{chart.xlab.fontface}{test123}

\item{chart.xlab.fontfamily}{test123}

\item{chart.ylab.color}{test123}

\item{chart.ylab.size}{test123}

\item{chart.ylab.fontface}{test123}

\item{chart.ylab.fontfamily}{test123}

\item{chart.text.color}{test123}

\item{chart.text.size}{test123}

\item{chart.text.fontface}{test123}

\item{chart.text.fontfamily}{test123}

\item{chart.frame}{test123}

\item{chart.frame.lwd}{test123}

\item{chart.frame.r}{test123}

\item{chart.bg.color}{test123}

\item{chart.bg.alpha}{test123}

\item{chart.object.color}{test123}

\item{title.show}{test123}

\item{title.size}{test123}

\item{title.color}{test123}

\item{title.fontface}{test123}

\item{title.fontfamily}{test123}

\item{title.bg.color}{test123}

\item{title.bg.alpha}{test123}

\item{title.padding}{test123}

\item{title.frame}{test123}

\item{title.frame.lwd}{test123}

\item{title.frame.r}{test123}

\item{title.stack}{test123}

\item{title.position}{test123}

\item{title.width}{test123}

\item{title.group.frame}{test123}

\item{title.resize_as_group}{test123}

\item{credits.show}{test123}

\item{credits.size}{test123}

\item{credits.color}{test123}

\item{credits.fontface}{test123}

\item{credits.fontfamily}{test123}

\item{credits.bg.color}{test123}

\item{credits.bg.alpha}{test123}

\item{credits.padding}{test123}

\item{credits.frame}{test123}

\item{credits.frame.lwd}{test123}

\item{credits.frame.r}{test123}

\item{credits.stack}{test123}

\item{credits.position}{test123}

\item{credits.width}{test123}

\item{credits.heigth}{test123}

\item{credits.group.frame}{test123}

\item{credits.resize_as_group}{test123}

\item{compass.north}{test123}

\item{compass.type}{test123}

\item{compass.text.size}{test123}

\item{compass.size}{test123}

\item{compass.show.labels}{test123}

\item{compass.cardinal.directions}{test123}

\item{compass.text.color}{test123}

\item{compass.color.dark}{test123}

\item{compass.color.light}{test123}

\item{compass.lwd}{test123}

\item{compass.bg.color}{test123}

\item{compass.bg.alpha}{test123}

\item{compass.margins}{test123}

\item{compass.show}{test123}

\item{compass.stack}{test123}

\item{compass.position}{test123}

\item{compass.frame}{test123}

\item{compass.frame.lwd}{test123}

\item{compass.frame.r}{test123}

\item{compass.group.frame}{test123}

\item{compass.resize_as_group}{test123}

\item{logo.height}{test123}

\item{logo.margins}{test123}

\item{logo.between_margin}{test123}

\item{logo.show}{test123}

\item{logo.stack}{test123}

\item{logo.position}{test123}

\item{logo.frame}{test123}

\item{logo.frame.lwd}{test123}

\item{logo.frame.r}{test123}

\item{logo.group.frame}{test123}

\item{logo.resize_as_group}{test123}

\item{scalebar.show}{test123}

\item{scalebar.breaks}{test123}

\item{scalebar.width}{test123}

\item{scalebar.text.size}{test123}

\item{scalebar.text.color}{test123}

\item{scalebar.color.dark}{test123}

\item{scalebar.color.light}{test123}

\item{scalebar.lwd}{test123}

\item{scalebar.bg.color}{test123}

\item{scalebar.bg.alpha}{test123}

\item{scalebar.size}{test123}

\item{scalebar.margins}{test123}

\item{scalebar.stack}{test123}

\item{scalebar.position}{test123}

\item{scalebar.frame}{test123}

\item{scalebar.frame.lwd}{test123}

\item{scalebar.frame.r}{test123}

\item{scalebar.group.frame}{test123}

\item{scalebar.resize_as_group}{test123}

\item{grid.show}{test123}

\item{grid.labels.pos}{test123}

\item{grid.x}{test123}

\item{grid.y}{test123}

\item{grid.n.x}{test123}

\item{grid.n.y}{test123}

\item{grid.crs}{test123}

\item{grid.col}{test123}

\item{grid.lwd}{test123}

\item{grid.alpha}{test123}

\item{grid.labels.show}{test123}

\item{grid.labels.size}{test123}

\item{grid.labels.col}{test123}

\item{grid.labels.rot}{test123}

\item{grid.labels.format}{test123}

\item{grid.labels.cardinal}{test123}

\item{grid.labels.margin.x}{test123}

\item{grid.labels.margin.y}{test123}

\item{grid.labels.space.x}{test123}

\item{grid.labels.space.y}{test123}

\item{grid.labels.inside_frame}{test123}

\item{grid.ticks}{test123}

\item{grid.lines}{test123}

\item{grid.ndiscr}{test123}

\item{mouse_coordinates.stack}{test123}

\item{mouse_coordinates.position}{test123}

\item{mouse_coordinates.show}{test123}

\item{minimap.server}{test123}

\item{minimap.toggle}{test123}

\item{minimap.stack}{test123}

\item{minimap.position}{test123}

\item{minimap.show}{test123}

\item{panel.show}{test123}

\item{panel.labels}{test123}

\item{panel.label.size}{test123}

\item{panel.label.color}{test123}

\item{panel.label.fontface}{test123}

\item{panel.label.fontfamily}{test123}

\item{panel.label.bg.color}{test123}

\item{panel.label.frame}{test123}

\item{panel.label.frame.lwd}{test123}

\item{panel.label.frame.r}{test123}

\item{panel.label.height}{test123}

\item{panel.label.rot}{test123}

\item{bbox}{test123}

\item{set.bounds}{test123}

\item{set.view}{test123}

\item{set.zoom.limits}{test123}

\item{qtm.scalebar}{test123}

\item{qtm.minimap}{test123}

\item{qtm.mouse_coordinates}{test123}

\item{earth.boundary}{test123}

\item{earth.boundary.color}{test123}

\item{earth.boundary.lwd}{test123}

\item{earth.datum}{test123}

\item{space.color}{test123}

\item{check_and_fix}{test123}

\item{basemap.show}{test123}

\item{basemap.server}{test123}

\item{basemap.alpha}{test123}

\item{basemap.zoom}{test123}

\item{tiles.show}{test123}

\item{tiles.server}{test123}

\item{tiles.alpha}{test123}

\item{tiles.zoom}{test123}

\item{attr.color}{test123}

\item{title, main.title}{deprecated See \code{\link[=tm_title]{tm_title()}}}

\item{view.legend.position}{deprecated. Use \code{legend.position} instead.}

\item{mode}{mode, e.g. \code{"plot"} or \code{"view"}}

\item{default.options}{return the default options or the current options?}

\item{style}{style, see \code{\link[=tmap_style]{tmap_style()}} for available styles}
}
\description{
tmap options
}
