% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapOnLoad.R, R/tmap_providers.R
\docType{data}
\name{.tmap_providers}
\alias{.tmap_providers}
\alias{tmap_provider_credits}
\alias{tmap_providers}
\title{Get basemap tiles providers}
\format{
An object of class \code{environment} of length 44.
}
\usage{
.tmap_providers

tmap_provider_credits(provider)

tmap_providers(mode, credits = FALSE, as.list = credits)
}
\arguments{
\item{provider}{provider name}

\item{mode}{mode. If not specified the default mode is used}

\item{credits}{If \code{TRUE} the credit (attribution) text is returned. If \code{FALSE} (default) the provider name.}

\item{as.list}{Should the output be returned as list where names are provider names? By default \code{TRUE} when \code{credits} is also \code{TRUE}.}
}
\value{
list or vector (see \code{as.list}) with providers (or credits). \code{\link[=tmap_provider_credits]{tmap_provider_credits()}} returns the credits text for the provided provider.
}
\description{
Get basemap tiles providers and the credits (attribution text). \code{\link[=tmap_providers]{tmap_providers()}} returns a list (or vector) with provider nams (or credits). \code{\link[=tmap_provider_credits]{tmap_provider_credits()}}
}
\keyword{datasets}
