% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_style.R
\name{tmap_style}
\alias{tmap_style}
\alias{tmap_style_load}
\alias{tmap_style_save}
\title{Set the default tmap style}
\usage{
tmap_style(style)

tmap_style_load(x)

tmap_style_save(style)
}
\arguments{
\item{style}{name of the style. When omitted, \code{tmap_style} returns the current style and also shows all available styles. When the style is specified, \code{tmap_style} sets the style accordingly. Note that in that case, all tmap options (see \code{\link{tmap_options}}) will be reset according to the style definition. See \code{\link{tm_layout}} for predefined styles, and \code{\link{tmap_style_catalogue}} for creating a catelogue.}

\item{x}{tmap options list (should be the same format as \code{tmap_options()})}
}
\value{
the style before changing
}
\description{
Set the default tmap style, which is contained in the global option \code{tmap.style} (see also \code{\link{tmap_options}}. The default style (i.e. when loading the package) is \code{"white"}.
}
\examples{
data(World)

current.style <- tmap_style("classic")
qtm(World, fill="life_exp", fill.title="Life expectancy")

tmap_style("cobalt")
qtm(World, fill="life_exp", fill.title="Life expectancy")

# restore current style
tmap_style(current.style)
}
\seealso{
\code{\link{tmap_options}} for tmap options, \code{\link{tmap_style_catalogue}} to create a style catelogue of all available styles, and \code{\link{tmap_options}} for tmap options.

\code{\link{tmap_options}} for tmap options
}
