% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_tip.R
\name{tmap_tip}
\alias{tmap_tip}
\title{Get a tip about tmap}
\usage{
tmap_tip(latest.version = FALSE)
}
\arguments{
\item{latest.version}{should only tips be generated from the latest version of tmap? False by default.}
}
\description{
Generates a tip with an example. The tip and example code are printed, and the example itself is executed.
}
\examples{
tmap_tip() 
}
