% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_tip.R
\name{tmap_tip}
\alias{tmap_tip}
\title{Get a tip about tmap}
\usage{
tmap_tip(from.version = NULL)
}
\arguments{
\item{from.version}{version number. Only tips regarding features from this version are shown.}
}
\description{
Generates a tip with an example. The tip and example code are printed, and the example itself is executed.
}
\examples{
tmap_tip()
tmap_tip(from.version = "3.0")
}
