% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_options.R
\name{tmapAddLayerOptions}
\alias{tmapAddLayerOptions}
\title{Internal tmap function to add a default value for the layer functions. Deprecated}
\usage{
tmapAddLayerOptions(option, id, value)
}
\arguments{
\item{option, }{one of: \code{"value.const"}, \code{"value.na"}, \code{"value.blank"}, \code{"values.var"}, \code{'values.range'}, \code{"value.neutral"}, \code{"scales.var"}}

\item{id}{name of the visual variable with layer, in the format \code{"x.y"},
where \code{x} is the visual variable and \code{y} is the layer.
It is also possible to set \code{x} only; then it applies to all layer functions.}

\item{value}{value}
}
\description{
Internal tmap function to add a default value for the layer functions. Deprecated. Use \code{\link[=tmapSubmitOptions]{tmapSubmitOptions()}} instead.
}
\keyword{internal}
