% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape.R
\name{tmapGetShapeMeta1}
\alias{tmapGetShapeMeta1}
\title{Internal method that extracts meta data from shape objects}
\usage{
tmapGetShapeMeta1(shp, o)
}
\arguments{
\item{shp}{the shape object}

\item{o}{the list of options}
}
\description{
Internal method that extracts meta data from shape objects
}
\keyword{internal}
