% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGridComp_leg_landscape.R,
%   R/tmapLeafletComp_leg_portrait.R
\name{tmapGridCompPrepare}
\alias{tmapGridCompPrepare}
\alias{tmapGridCompPrepare.default}
\alias{tmapLeafletCompPrepare}
\alias{tmapLeafletCompPrepare.default}
\title{Internal methods for map components in plot and view mode}
\usage{
tmapGridCompPrepare(comp, o)

\method{tmapGridCompPrepare}{default}(comp, o)

tmapLeafletCompPrepare(comp, o)

\method{tmapLeafletCompPrepare}{default}(comp, o)
}
\arguments{
\item{comp}{component}

\item{o}{options}
}
\description{
Internal methods for map components
}
\keyword{internal}
