% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapGridAux.R, R/tmapGridAux_grid.R,
%   R/tmapGridAux_tiles.R, R/tmapGridComp_leg_landscape.R,
%   R/tmapGridDataPlot.R, R/tmapGridDataPlot_lines.R,
%   R/tmapGridDataPlot_polygons.R, R/tmapGridDataPlot_raster.R,
%   R/tmapGridDataPlot_symbols.R, R/tmapGridDataPlot_text.R, R/tmapLeafletAux.R,
%   R/tmapLeafletComp_leg_portrait.R, R/tmapLeafletDataPlot_lines.R,
%   R/tmapLeafletDataPlot_polygons.R, R/tmapLeafletDataPlot_raster.R,
%   R/tmapLeafletDataPlot_symbols.R, R/tmapLeafletDataPlot_text.R,
%   R/tmapLeaflet_layers.R
\name{tmapGridAuxPrepare}
\alias{tmapGridAuxPrepare}
\alias{tmapGridAuxPrepare.default}
\alias{tmapGridAuxPlot}
\alias{tmapGridAuxPlot.default}
\alias{tmapGridAuxPrepare.tm_aux_graticules}
\alias{tmapGridAuxPrepare.tm_aux_grid}
\alias{tmapGridAuxPlot.tm_aux_basemap}
\alias{tmapGridAuxPlot.tm_aux_tiles}
\alias{tmapGridAuxPlot.tm_aux_graticules}
\alias{tmapGridAuxPlot.tm_aux_grid}
\alias{tmapGridAuxPrepare.tm_aux_basemap}
\alias{tmapGridAuxPrepare.tm_aux_tiles}
\alias{tmapGridCompPrepare}
\alias{tmapGridCompPrepare.default}
\alias{tmapGridDataPlot}
\alias{tmapGridDataPlot.default}
\alias{tmapGridDataPlot.tm_data_lines}
\alias{tmapGridDataPlot.tm_data_iso}
\alias{tmapGridDataPlot.tm_data_polygons}
\alias{tmapGridDataPlot.tm_data_fill}
\alias{tmapGridDataPlot.tm_data_borders}
\alias{tmapGridDataPlot.tm_data_raster}
\alias{tmapGridDataPlot.tm_data_symbols}
\alias{tmapGridDataPlot.tm_data_dots}
\alias{tmapGridDataPlot.tm_data_bubbles}
\alias{tmapGridDataPlot.tm_data_squares}
\alias{tmapGridDataPlot.tm_data_markers}
\alias{tmapGridDataPlot.tm_data_text}
\alias{tmapGridDataPlot.tm_data_labels}
\alias{tmapGridDataPlot.tm_data_labels_highlighted}
\alias{tmapLeafletAuxPrepare}
\alias{tmapLeafletAuxPlot}
\alias{tmapLeafletAuxPrepare.tm_aux_basemap}
\alias{tmapLeafletAuxPrepare.tm_aux_tiles}
\alias{tmapLeafletAuxPlot.tm_aux_basemap}
\alias{tmapLeafletAuxPlot.tm_aux_tiles}
\alias{tmapLeafletAuxPrepare.tm_aux_graticules}
\alias{tmapLeafletAuxPrepare.tm_aux_grid}
\alias{tmapLeafletAuxPlot.tm_aux_graticules}
\alias{tmapLeafletAuxPlot.tm_aux_grid}
\alias{tmapLeafletCompPrepare}
\alias{tmapLeafletCompPrepare.tm_chart_none}
\alias{tmapLeafletCompPrepare.default}
\alias{tmapLeafletDataPlot.tm_data_lines}
\alias{tmapLeafletDataPlot.tm_data_iso}
\alias{tmapLeafletDataPlot.tm_data_polygons}
\alias{tmapLeafletDataPlot.tm_data_fill}
\alias{tmapLeafletDataPlot.tm_data_borders}
\alias{tmapLeafletDataPlot.tm_data_raster}
\alias{tmapLeafletDataPlot.tm_data_symbols}
\alias{tmapLeafletDataPlot.tm_data_dots}
\alias{tmapLeafletDataPlot.tm_data_bubbles}
\alias{tmapLeafletDataPlot.tm_data_squares}
\alias{tmapLeafletDataPlot.tm_data_markers}
\alias{tmapLeafletDataPlot.tm_data_text}
\alias{tmapLeafletDataPlot.tm_data_labels}
\alias{tmapLeafletDataPlot.tm_data_labels_highlighted}
\alias{tmapLeafletDataPlot}
\alias{tmapLeafletDataPlot.default}
\title{Internal methods for map components in plot and view mode}
\usage{
tmapGridAuxPrepare(a, bs, id, o)

\method{tmapGridAuxPrepare}{default}(a, bs, id, o)

tmapGridAuxPlot(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPlot}{default}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPrepare}{tm_aux_graticules}(a, bs, id, o)

\method{tmapGridAuxPrepare}{tm_aux_grid}(a, bs, id, o)

\method{tmapGridAuxPlot}{tm_aux_basemap}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPlot}{tm_aux_tiles}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPlot}{tm_aux_graticules}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPlot}{tm_aux_grid}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapGridAuxPrepare}{tm_aux_basemap}(a, bs, id, o)

\method{tmapGridAuxPrepare}{tm_aux_tiles}(a, bs, id, o)

tmapGridCompPrepare(comp, o)

\method{tmapGridCompPrepare}{default}(comp, o)

tmapGridDataPlot(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{default}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_lines}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_iso}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_polygons}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_fill}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_borders}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_raster}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_symbols}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_dots}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_bubbles}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_squares}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_markers}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_text}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_labels}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapGridDataPlot}{tm_data_labels_highlighted}(
  a,
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapLeafletAuxPrepare(a, bs, id, o)

tmapLeafletAuxPlot(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapLeafletAuxPrepare}{tm_aux_basemap}(a, bs, id, o)

\method{tmapLeafletAuxPrepare}{tm_aux_tiles}(a, bs, id, o)

\method{tmapLeafletAuxPlot}{tm_aux_basemap}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapLeafletAuxPlot}{tm_aux_tiles}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapLeafletAuxPrepare}{tm_aux_graticules}(a, bs, id, o)

\method{tmapLeafletAuxPrepare}{tm_aux_grid}(a, bs, id, o)

\method{tmapLeafletAuxPlot}{tm_aux_graticules}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

\method{tmapLeafletAuxPlot}{tm_aux_grid}(
  a,
  bi,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapLeafletCompPrepare(comp, o)

\method{tmapLeafletCompPrepare}{tm_chart_none}(comp, o)

\method{tmapLeafletCompPrepare}{default}(comp, o)

\method{tmapLeafletDataPlot}{tm_data_lines}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_iso}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_polygons}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_fill}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_borders}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_raster}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_symbols}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_dots}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_bubbles}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_squares}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_markers}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_text}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_labels}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{tm_data_labels_highlighted}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapLeafletDataPlot(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

\method{tmapLeafletDataPlot}{default}(
  a,
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{o}{options}

\item{comp}{component}
}
\description{
Internal methods for map components
}
\keyword{internal}
