% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape.R
\name{tmapShape}
\alias{tmapShape}
\title{Internal method that processed shape objects}
\usage{
tmapShape(shp, is.main, crs, bbox, unit, filter, shp_name, smeta, o, tmf)
}
\arguments{
\item{shp}{shp}

\item{is.main}{is.main}

\item{crs}{crs}

\item{bbox}{Bounding box}

\item{unit}{unit}

\item{filter}{filter}

\item{shp_name}{shp_name}

\item{smeta}{smeta}

\item{o}{o}

\item{tmf}{tmf}
}
\description{
Internal method that processed shape objects
}
\keyword{internal}
