% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapShape.R
\name{tmapSubsetShp}
\alias{tmapSubsetShp}
\title{Internal method that subsets data from shape objects}
\usage{
tmapSubsetShp(shp, vars)
}
\arguments{
\item{shp}{shape}

\item{vars}{a vector of variable names}
}
\description{
Internal method that subsets data from shape objects
}
\keyword{internal}
