% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v4-not-implemented.R
\name{v4-not-yet}
\alias{v4-not-yet}
\alias{tm_squares}
\alias{tm_iso}
\alias{tm_markers}
\alias{tm_logo}
\alias{tm_minimap}
\title{These functions will be implemented soon}
\usage{
tm_squares()

tm_iso()

tm_markers()

tm_logo()

tm_minimap()
}
\description{
These functions will be implemented soon
}
