% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v4-not-implemented.R
\name{v4-not-yet}
\alias{v4-not-yet}
\alias{tm_iso}
\alias{tm_markers}
\alias{tm_logo}
\alias{tm_minimap}
\alias{tmap_grob}
\alias{tmapOutput}
\alias{tmapProxy}
\alias{renderTmap}
\alias{tm_remove_layer}
\alias{tm_rgba}
\alias{tmap_tip}
\alias{tmap_style_catalog}
\alias{tmap_style_catalogue}
\title{Features not yet implemented in tmap4}
\usage{
tm_iso(...)

tm_markers(...)

tm_logo(...)

tm_minimap(...)

tmap_grob(...)

tmapOutput(...)

tmapProxy(...)

renderTmap(...)

tm_remove_layer(...)

tm_rgba(...)

tmap_tip(...)

tmap_style_catalog(...)

tmap_style_catalogue(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Features not yet implemented in tmap4
}
