% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shape.R
\name{write_shape}
\alias{write_shape}
\title{Write shape file}
\usage{
write_shape(shp, file)
}
\arguments{
\item{shp}{a shape object}

\item{file}{file name (including directory)}
}
\description{
Write a shape object to an ESRI shape file.
}
\details{
This function is a convenient wrapper of rgdal's \code{\link[rgdal:writeOGR]{writeOGR}}.
}

